feat: sematic v1.25.0
