+++ packages/opentelemetry-semantic-conventions/src/trace/SemanticAttributes.ts
@@ -26,14 +26,6 @@ import { createConstMap } from '../internal/utils';
 
 // Temporary local constants to assign to the individual exports and the namespaced version
 // Required to avoid the namespace exports using the unminifiable export names for some package types
-const TMP_FAAS_INVOKED_NAME = 'faas.invoked_name';
-const TMP_FAAS_INVOKED_PROVIDER = 'faas.invoked_provider';
-const TMP_FAAS_INVOKED_REGION = 'faas.invoked_region';
-const TMP_FAAS_TRIGGER = 'faas.trigger';
-const TMP_PEER_SERVICE = 'peer.service';
-const TMP_ENDUSER_ID = 'enduser.id';
-const TMP_ENDUSER_ROLE = 'enduser.role';
-const TMP_ENDUSER_SCOPE = 'enduser.scope';
 const TMP_EVENT_NAME = 'event.name';
 const TMP_LOG_RECORD_UID = 'log.record.uid';
 const TMP_LOG_IOSTREAM = 'log.iostream';
@@ -45,9 +37,9 @@ const TMP_IOS_STATE = 'ios.state';
 const TMP_ANDROID_STATE = 'android.state';
 const TMP_POOL_NAME = 'pool.name';
 const TMP_STATE = 'state';
+const TMP_ASPNETCORE_RATE_LIMITING_RESULT = 'aspnetcore.rate_limiting.result';
 const TMP_ASPNETCORE_DIAGNOSTICS_HANDLER_TYPE = 'aspnetcore.diagnostics.handler.type';
 const TMP_ASPNETCORE_RATE_LIMITING_POLICY = 'aspnetcore.rate_limiting.policy';
-const TMP_ASPNETCORE_RATE_LIMITING_RESULT = 'aspnetcore.rate_limiting.result';
 const TMP_ASPNETCORE_REQUEST_IS_UNHANDLED = 'aspnetcore.request.is_unhandled';
 const TMP_ASPNETCORE_ROUTING_IS_FALLBACK = 'aspnetcore.routing.is_fallback';
 const TMP_SIGNALR_CONNECTION_STATUS = 'signalr.connection.status';
@@ -55,6 +47,7 @@ const TMP_SIGNALR_TRANSPORT = 'signalr.transport';
 const TMP_JVM_BUFFER_POOL_NAME = 'jvm.buffer.pool.name';
 const TMP_JVM_MEMORY_POOL_NAME = 'jvm.memory.pool.name';
 const TMP_JVM_MEMORY_TYPE = 'jvm.memory.type';
+const TMP_PROCESS_CPU_STATE = 'process.cpu.state';
 const TMP_SYSTEM_DEVICE = 'system.device';
 const TMP_SYSTEM_CPU_LOGICAL_NUMBER = 'system.cpu.logical_number';
 const TMP_SYSTEM_CPU_STATE = 'system.cpu.state';
@@ -67,15 +60,64 @@ const TMP_SYSTEM_FILESYSTEM_MOUNTPOINT = 'system.filesystem.mountpoint';
 const TMP_SYSTEM_FILESYSTEM_STATE = 'system.filesystem.state';
 const TMP_SYSTEM_FILESYSTEM_TYPE = 'system.filesystem.type';
 const TMP_SYSTEM_NETWORK_STATE = 'system.network.state';
-const TMP_SYSTEM_PROCESSES_STATUS = 'system.processes.status';
+const TMP_SYSTEM_PROCESS_STATUS = 'system.process.status';
+const TMP_ANDROID_OS_API_LEVEL = 'android.os.api_level';
+const TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS = 'aws.dynamodb.attribute_definitions';
+const TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET = 'aws.dynamodb.attributes_to_get';
+const TMP_AWS_DYNAMODB_CONSISTENT_READ = 'aws.dynamodb.consistent_read';
+const TMP_AWS_DYNAMODB_CONSUMED_CAPACITY = 'aws.dynamodb.consumed_capacity';
+const TMP_AWS_DYNAMODB_COUNT = 'aws.dynamodb.count';
+const TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE = 'aws.dynamodb.exclusive_start_table';
+const TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES = 'aws.dynamodb.global_secondary_index_updates';
+const TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES = 'aws.dynamodb.global_secondary_indexes';
+const TMP_AWS_DYNAMODB_INDEX_NAME = 'aws.dynamodb.index_name';
+const TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS = 'aws.dynamodb.item_collection_metrics';
+const TMP_AWS_DYNAMODB_LIMIT = 'aws.dynamodb.limit';
+const TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES = 'aws.dynamodb.local_secondary_indexes';
+const TMP_AWS_DYNAMODB_PROJECTION = 'aws.dynamodb.projection';
+const TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY = 'aws.dynamodb.provisioned_read_capacity';
+const TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY = 'aws.dynamodb.provisioned_write_capacity';
+const TMP_AWS_DYNAMODB_SCAN_FORWARD = 'aws.dynamodb.scan_forward';
+const TMP_AWS_DYNAMODB_SCANNED_COUNT = 'aws.dynamodb.scanned_count';
+const TMP_AWS_DYNAMODB_SEGMENT = 'aws.dynamodb.segment';
+const TMP_AWS_DYNAMODB_SELECT = 'aws.dynamodb.select';
+const TMP_AWS_DYNAMODB_TABLE_COUNT = 'aws.dynamodb.table_count';
+const TMP_AWS_DYNAMODB_TABLE_NAMES = 'aws.dynamodb.table_names';
+const TMP_AWS_DYNAMODB_TOTAL_SEGMENTS = 'aws.dynamodb.total_segments';
+const TMP_BROWSER_BRANDS = 'browser.brands';
+const TMP_BROWSER_LANGUAGE = 'browser.language';
+const TMP_BROWSER_MOBILE = 'browser.mobile';
+const TMP_BROWSER_PLATFORM = 'browser.platform';
 const TMP_CLIENT_ADDRESS = 'client.address';
 const TMP_CLIENT_PORT = 'client.port';
+const TMP_CLOUD_ACCOUNT_ID = 'cloud.account.id';
+const TMP_CLOUD_AVAILABILITY_ZONE = 'cloud.availability_zone';
+const TMP_CLOUD_PLATFORM = 'cloud.platform';
+const TMP_CLOUD_PROVIDER = 'cloud.provider';
+const TMP_CLOUD_REGION = 'cloud.region';
+const TMP_CLOUD_RESOURCE_ID = 'cloud.resource_id';
+const TMP_CLOUDEVENTS_EVENT_ID = 'cloudevents.event_id';
+const TMP_CLOUDEVENTS_EVENT_SOURCE = 'cloudevents.event_source';
+const TMP_CLOUDEVENTS_EVENT_SPEC_VERSION = 'cloudevents.event_spec_version';
+const TMP_CLOUDEVENTS_EVENT_SUBJECT = 'cloudevents.event_subject';
+const TMP_CLOUDEVENTS_EVENT_TYPE = 'cloudevents.event_type';
 const TMP_CODE_COLUMN = 'code.column';
 const TMP_CODE_FILEPATH = 'code.filepath';
 const TMP_CODE_FUNCTION = 'code.function';
 const TMP_CODE_LINENO = 'code.lineno';
 const TMP_CODE_NAMESPACE = 'code.namespace';
 const TMP_CODE_STACKTRACE = 'code.stacktrace';
+const TMP_CONTAINER_COMMAND = 'container.command';
+const TMP_CONTAINER_COMMAND_ARGS = 'container.command_args';
+const TMP_CONTAINER_COMMAND_LINE = 'container.command_line';
+const TMP_CONTAINER_CPU_STATE = 'container.cpu.state';
+const TMP_CONTAINER_ID = 'container.id';
+const TMP_CONTAINER_IMAGE_ID = 'container.image.id';
+const TMP_CONTAINER_IMAGE_NAME = 'container.image.name';
+const TMP_CONTAINER_IMAGE_REPO_DIGESTS = 'container.image.repo_digests';
+const TMP_CONTAINER_IMAGE_TAGS = 'container.image.tags';
+const TMP_CONTAINER_NAME = 'container.name';
+const TMP_CONTAINER_RUNTIME = 'container.runtime';
 const TMP_DB_CASSANDRA_CONSISTENCY_LEVEL = 'db.cassandra.consistency_level';
 const TMP_DB_CASSANDRA_COORDINATOR_DC = 'db.cassandra.coordinator.dc';
 const TMP_DB_CASSANDRA_COORDINATOR_ID = 'db.cassandra.coordinator.id';
@@ -83,7 +125,6 @@ const TMP_DB_CASSANDRA_IDEMPOTENCE = 'db.cassandra.idempotence';
 const TMP_DB_CASSANDRA_PAGE_SIZE = 'db.cassandra.page_size';
 const TMP_DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT = 'db.cassandra.speculative_execution_count';
 const TMP_DB_CASSANDRA_TABLE = 'db.cassandra.table';
-const TMP_DB_CONNECTION_STRING = 'db.connection_string';
 const TMP_DB_COSMOSDB_CLIENT_ID = 'db.cosmosdb.client_id';
 const TMP_DB_COSMOSDB_CONNECTION_MODE = 'db.cosmosdb.connection_mode';
 const TMP_DB_COSMOSDB_CONTAINER = 'db.cosmosdb.container';
@@ -93,9 +134,7 @@ const TMP_DB_COSMOSDB_REQUEST_CONTENT_LENGTH = 'db.cosmosdb.request_content_leng
 const TMP_DB_COSMOSDB_STATUS_CODE = 'db.cosmosdb.status_code';
 const TMP_DB_COSMOSDB_SUB_STATUS_CODE = 'db.cosmosdb.sub_status_code';
 const TMP_DB_ELASTICSEARCH_CLUSTER_NAME = 'db.elasticsearch.cluster.name';
-const TMP_DB_ELASTICSEARCH_NODE_NAME = 'db.elasticsearch.node.name';
 const TMP_DB_INSTANCE_ID = 'db.instance.id';
-const TMP_DB_JDBC_DRIVER_CLASSNAME = 'db.jdbc.driver_classname';
 const TMP_DB_MONGODB_COLLECTION = 'db.mongodb.collection';
 const TMP_DB_MSSQL_INSTANCE_NAME = 'db.mssql.instance_name';
 const TMP_DB_NAME = 'db.name';
@@ -105,6 +144,10 @@ const TMP_DB_SQL_TABLE = 'db.sql.table';
 const TMP_DB_STATEMENT = 'db.statement';
 const TMP_DB_SYSTEM = 'db.system';
 const TMP_DB_USER = 'db.user';
+const TMP_DEPLOYMENT_ENVIRONMENT = 'deployment.environment';
+const TMP_DB_CONNECTION_STRING = 'db.connection_string';
+const TMP_DB_ELASTICSEARCH_NODE_NAME = 'db.elasticsearch.node.name';
+const TMP_DB_JDBC_DRIVER_CLASSNAME = 'db.jdbc.driver_classname';
 const TMP_HTTP_FLAVOR = 'http.flavor';
 const TMP_HTTP_METHOD = 'http.method';
 const TMP_HTTP_REQUEST_CONTENT_LENGTH = 'http.request_content_length';
@@ -114,6 +157,7 @@ const TMP_HTTP_STATUS_CODE = 'http.status_code';
 const TMP_HTTP_TARGET = 'http.target';
 const TMP_HTTP_URL = 'http.url';
 const TMP_HTTP_USER_AGENT = 'http.user_agent';
+const TMP_MESSAGING_KAFKA_DESTINATION_PARTITION = 'messaging.kafka.destination.partition';
 const TMP_NET_HOST_NAME = 'net.host.name';
 const TMP_NET_HOST_PORT = 'net.host.port';
 const TMP_NET_PEER_NAME = 'net.peer.name';
@@ -127,32 +171,110 @@ const TMP_NET_SOCK_PEER_ADDR = 'net.sock.peer.addr';
 const TMP_NET_SOCK_PEER_NAME = 'net.sock.peer.name';
 const TMP_NET_SOCK_PEER_PORT = 'net.sock.peer.port';
 const TMP_NET_TRANSPORT = 'net.transport';
+const TMP_SYSTEM_PROCESSES_STATUS = 'system.processes.status';
 const TMP_DESTINATION_ADDRESS = 'destination.address';
 const TMP_DESTINATION_PORT = 'destination.port';
+const TMP_DEVICE_ID = 'device.id';
+const TMP_DEVICE_MANUFACTURER = 'device.manufacturer';
+const TMP_DEVICE_MODEL_IDENTIFIER = 'device.model.identifier';
+const TMP_DEVICE_MODEL_NAME = 'device.model.name';
 const TMP_DISK_IO_DIRECTION = 'disk.io.direction';
+const TMP_DNS_QUESTION_NAME = 'dns.question.name';
+const TMP_ENDUSER_ID = 'enduser.id';
+const TMP_ENDUSER_ROLE = 'enduser.role';
+const TMP_ENDUSER_SCOPE = 'enduser.scope';
 const TMP_ERROR_TYPE = 'error.type';
 const TMP_EXCEPTION_ESCAPED = 'exception.escaped';
 const TMP_EXCEPTION_MESSAGE = 'exception.message';
 const TMP_EXCEPTION_STACKTRACE = 'exception.stacktrace';
 const TMP_EXCEPTION_TYPE = 'exception.type';
+const TMP_FAAS_COLDSTART = 'faas.coldstart';
+const TMP_FAAS_CRON = 'faas.cron';
+const TMP_FAAS_DOCUMENT_COLLECTION = 'faas.document.collection';
+const TMP_FAAS_DOCUMENT_NAME = 'faas.document.name';
+const TMP_FAAS_DOCUMENT_OPERATION = 'faas.document.operation';
+const TMP_FAAS_DOCUMENT_TIME = 'faas.document.time';
+const TMP_FAAS_INSTANCE = 'faas.instance';
+const TMP_FAAS_INVOCATION_ID = 'faas.invocation_id';
+const TMP_FAAS_INVOKED_NAME = 'faas.invoked_name';
+const TMP_FAAS_INVOKED_PROVIDER = 'faas.invoked_provider';
+const TMP_FAAS_INVOKED_REGION = 'faas.invoked_region';
+const TMP_FAAS_MAX_MEMORY = 'faas.max_memory';
+const TMP_FAAS_NAME = 'faas.name';
+const TMP_FAAS_TIME = 'faas.time';
+const TMP_FAAS_TRIGGER = 'faas.trigger';
+const TMP_FAAS_VERSION = 'faas.version';
+const TMP_FEATURE_FLAG_KEY = 'feature_flag.key';
+const TMP_FEATURE_FLAG_PROVIDER_NAME = 'feature_flag.provider_name';
+const TMP_FEATURE_FLAG_VARIANT = 'feature_flag.variant';
+const TMP_FILE_DIRECTORY = 'file.directory';
+const TMP_FILE_EXTENSION = 'file.extension';
+const TMP_FILE_NAME = 'file.name';
+const TMP_FILE_PATH = 'file.path';
+const TMP_FILE_SIZE = 'file.size';
+const TMP_GCP_CLOUD_RUN_JOB_EXECUTION = 'gcp.cloud_run.job.execution';
+const TMP_GCP_CLOUD_RUN_JOB_TASK_INDEX = 'gcp.cloud_run.job.task_index';
+const TMP_GCP_GCE_INSTANCE_HOSTNAME = 'gcp.gce.instance.hostname';
+const TMP_GCP_GCE_INSTANCE_NAME = 'gcp.gce.instance.name';
+const TMP_HOST_ARCH = 'host.arch';
+const TMP_HOST_CPU_CACHE_L2_SIZE = 'host.cpu.cache.l2.size';
+const TMP_HOST_CPU_FAMILY = 'host.cpu.family';
+const TMP_HOST_CPU_MODEL_ID = 'host.cpu.model.id';
+const TMP_HOST_CPU_MODEL_NAME = 'host.cpu.model.name';
+const TMP_HOST_CPU_STEPPING = 'host.cpu.stepping';
+const TMP_HOST_CPU_VENDOR_ID = 'host.cpu.vendor.id';
+const TMP_HOST_ID = 'host.id';
+const TMP_HOST_IMAGE_ID = 'host.image.id';
+const TMP_HOST_IMAGE_NAME = 'host.image.name';
+const TMP_HOST_IMAGE_VERSION = 'host.image.version';
+const TMP_HOST_IP = 'host.ip';
+const TMP_HOST_MAC = 'host.mac';
+const TMP_HOST_NAME = 'host.name';
+const TMP_HOST_TYPE = 'host.type';
+const TMP_HTTP_CONNECTION_STATE = 'http.connection.state';
 const TMP_HTTP_REQUEST_BODY_SIZE = 'http.request.body.size';
 const TMP_HTTP_REQUEST_METHOD = 'http.request.method';
 const TMP_HTTP_REQUEST_METHOD_ORIGINAL = 'http.request.method_original';
 const TMP_HTTP_REQUEST_RESEND_COUNT = 'http.request.resend_count';
+const TMP_HTTP_REQUEST_SIZE = 'http.request.size';
 const TMP_HTTP_RESPONSE_BODY_SIZE = 'http.response.body.size';
+const TMP_HTTP_RESPONSE_SIZE = 'http.response.size';
 const TMP_HTTP_RESPONSE_STATUS_CODE = 'http.response.status_code';
 const TMP_HTTP_ROUTE = 'http.route';
+const TMP_K8S_CLUSTER_NAME = 'k8s.cluster.name';
+const TMP_K8S_CLUSTER_UID = 'k8s.cluster.uid';
+const TMP_K8S_CONTAINER_NAME = 'k8s.container.name';
+const TMP_K8S_CONTAINER_RESTART_COUNT = 'k8s.container.restart_count';
+const TMP_K8S_CRONJOB_NAME = 'k8s.cronjob.name';
+const TMP_K8S_CRONJOB_UID = 'k8s.cronjob.uid';
+const TMP_K8S_DAEMONSET_NAME = 'k8s.daemonset.name';
+const TMP_K8S_DAEMONSET_UID = 'k8s.daemonset.uid';
+const TMP_K8S_DEPLOYMENT_NAME = 'k8s.deployment.name';
+const TMP_K8S_DEPLOYMENT_UID = 'k8s.deployment.uid';
+const TMP_K8S_JOB_NAME = 'k8s.job.name';
+const TMP_K8S_JOB_UID = 'k8s.job.uid';
+const TMP_K8S_NAMESPACE_NAME = 'k8s.namespace.name';
+const TMP_K8S_NODE_NAME = 'k8s.node.name';
+const TMP_K8S_NODE_UID = 'k8s.node.uid';
+const TMP_K8S_POD_NAME = 'k8s.pod.name';
+const TMP_K8S_POD_UID = 'k8s.pod.uid';
+const TMP_K8S_REPLICASET_NAME = 'k8s.replicaset.name';
+const TMP_K8S_REPLICASET_UID = 'k8s.replicaset.uid';
+const TMP_K8S_STATEFULSET_NAME = 'k8s.statefulset.name';
+const TMP_K8S_STATEFULSET_UID = 'k8s.statefulset.uid';
 const TMP_MESSAGING_BATCH_MESSAGE_COUNT = 'messaging.batch.message_count';
 const TMP_MESSAGING_CLIENT_ID = 'messaging.client_id';
 const TMP_MESSAGING_DESTINATION_ANONYMOUS = 'messaging.destination.anonymous';
 const TMP_MESSAGING_DESTINATION_NAME = 'messaging.destination.name';
+const TMP_MESSAGING_DESTINATION_PARTITION_ID = 'messaging.destination.partition.id';
 const TMP_MESSAGING_DESTINATION_TEMPLATE = 'messaging.destination.template';
 const TMP_MESSAGING_DESTINATION_TEMPORARY = 'messaging.destination.temporary';
 const TMP_MESSAGING_DESTINATION_PUBLISH_ANONYMOUS = 'messaging.destination_publish.anonymous';
 const TMP_MESSAGING_DESTINATION_PUBLISH_NAME = 'messaging.destination_publish.name';
+const TMP_MESSAGING_EVENTHUBS_CONSUMER_GROUP = 'messaging.eventhubs.consumer.group';
+const TMP_MESSAGING_EVENTHUBS_MESSAGE_ENQUEUED_TIME = 'messaging.eventhubs.message.enqueued_time';
 const TMP_MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY = 'messaging.gcp_pubsub.message.ordering_key';
 const TMP_MESSAGING_KAFKA_CONSUMER_GROUP = 'messaging.kafka.consumer.group';
-const TMP_MESSAGING_KAFKA_DESTINATION_PARTITION = 'messaging.kafka.destination.partition';
 const TMP_MESSAGING_KAFKA_MESSAGE_KEY = 'messaging.kafka.message.key';
 const TMP_MESSAGING_KAFKA_MESSAGE_OFFSET = 'messaging.kafka.message.offset';
 const TMP_MESSAGING_KAFKA_MESSAGE_TOMBSTONE = 'messaging.kafka.message.tombstone';
@@ -162,6 +284,7 @@ const TMP_MESSAGING_MESSAGE_ENVELOPE_SIZE = 'messaging.message.envelope.size';
 const TMP_MESSAGING_MESSAGE_ID = 'messaging.message.id';
 const TMP_MESSAGING_OPERATION = 'messaging.operation';
 const TMP_MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY = 'messaging.rabbitmq.destination.routing_key';
+const TMP_MESSAGING_RABBITMQ_MESSAGE_DELIVERY_TAG = 'messaging.rabbitmq.message.delivery_tag';
 const TMP_MESSAGING_ROCKETMQ_CLIENT_GROUP = 'messaging.rocketmq.client_group';
 const TMP_MESSAGING_ROCKETMQ_CONSUMPTION_MODEL = 'messaging.rocketmq.consumption_model';
 const TMP_MESSAGING_ROCKETMQ_MESSAGE_DELAY_TIME_LEVEL = 'messaging.rocketmq.message.delay_time_level';
@@ -171,6 +294,10 @@ const TMP_MESSAGING_ROCKETMQ_MESSAGE_KEYS = 'messaging.rocketmq.message.keys';
 const TMP_MESSAGING_ROCKETMQ_MESSAGE_TAG = 'messaging.rocketmq.message.tag';
 const TMP_MESSAGING_ROCKETMQ_MESSAGE_TYPE = 'messaging.rocketmq.message.type';
 const TMP_MESSAGING_ROCKETMQ_NAMESPACE = 'messaging.rocketmq.namespace';
+const TMP_MESSAGING_SERVICEBUS_DESTINATION_SUBSCRIPTION_NAME = 'messaging.servicebus.destination.subscription_name';
+const TMP_MESSAGING_SERVICEBUS_DISPOSITION_STATUS = 'messaging.servicebus.disposition_status';
+const TMP_MESSAGING_SERVICEBUS_MESSAGE_DELIVERY_COUNT = 'messaging.servicebus.message.delivery_count';
+const TMP_MESSAGING_SERVICEBUS_MESSAGE_ENQUEUED_TIME = 'messaging.servicebus.message.enqueued_time';
 const TMP_MESSAGING_SYSTEM = 'messaging.system';
 const TMP_NETWORK_CARRIER_ICC = 'network.carrier.icc';
 const TMP_NETWORK_CARRIER_MCC = 'network.carrier.mcc';
@@ -187,6 +314,24 @@ const TMP_NETWORK_PROTOCOL_NAME = 'network.protocol.name';
 const TMP_NETWORK_PROTOCOL_VERSION = 'network.protocol.version';
 const TMP_NETWORK_TRANSPORT = 'network.transport';
 const TMP_NETWORK_TYPE = 'network.type';
+const TMP_OCI_MANIFEST_DIGEST = 'oci.manifest.digest';
+const TMP_OS_BUILD_ID = 'os.build_id';
+const TMP_OS_DESCRIPTION = 'os.description';
+const TMP_OS_NAME = 'os.name';
+const TMP_OS_TYPE = 'os.type';
+const TMP_OS_VERSION = 'os.version';
+const TMP_PEER_SERVICE = 'peer.service';
+const TMP_PROCESS_COMMAND = 'process.command';
+const TMP_PROCESS_COMMAND_ARGS = 'process.command_args';
+const TMP_PROCESS_COMMAND_LINE = 'process.command_line';
+const TMP_PROCESS_EXECUTABLE_NAME = 'process.executable.name';
+const TMP_PROCESS_EXECUTABLE_PATH = 'process.executable.path';
+const TMP_PROCESS_OWNER = 'process.owner';
+const TMP_PROCESS_PARENT_PID = 'process.parent_pid';
+const TMP_PROCESS_PID = 'process.pid';
+const TMP_PROCESS_RUNTIME_DESCRIPTION = 'process.runtime.description';
+const TMP_PROCESS_RUNTIME_NAME = 'process.runtime.name';
+const TMP_PROCESS_RUNTIME_VERSION = 'process.runtime.version';
 const TMP_RPC_CONNECT_RPC_ERROR_CODE = 'rpc.connect_rpc.error_code';
 const TMP_RPC_GRPC_STATUS_CODE = 'rpc.grpc.status_code';
 const TMP_RPC_JSONRPC_ERROR_CODE = 'rpc.jsonrpc.error_code';
@@ -198,8 +343,19 @@ const TMP_RPC_SERVICE = 'rpc.service';
 const TMP_RPC_SYSTEM = 'rpc.system';
 const TMP_SERVER_ADDRESS = 'server.address';
 const TMP_SERVER_PORT = 'server.port';
+const TMP_SERVICE_INSTANCE_ID = 'service.instance.id';
+const TMP_SERVICE_NAME = 'service.name';
+const TMP_SERVICE_NAMESPACE = 'service.namespace';
+const TMP_SERVICE_VERSION = 'service.version';
+const TMP_SESSION_ID = 'session.id';
+const TMP_SESSION_PREVIOUS_ID = 'session.previous_id';
 const TMP_SOURCE_ADDRESS = 'source.address';
 const TMP_SOURCE_PORT = 'source.port';
+const TMP_TELEMETRY_SDK_LANGUAGE = 'telemetry.sdk.language';
+const TMP_TELEMETRY_SDK_NAME = 'telemetry.sdk.name';
+const TMP_TELEMETRY_SDK_VERSION = 'telemetry.sdk.version';
+const TMP_TELEMETRY_DISTRO_NAME = 'telemetry.distro.name';
+const TMP_TELEMETRY_DISTRO_VERSION = 'telemetry.distro.version';
 const TMP_THREAD_ID = 'thread.id';
 const TMP_THREAD_NAME = 'thread.name';
 const TMP_TLS_CIPHER = 'tls.cipher';
@@ -231,57 +387,26 @@ const TMP_TLS_SERVER_JA3S = 'tls.server.ja3s';
 const TMP_TLS_SERVER_NOT_AFTER = 'tls.server.not_after';
 const TMP_TLS_SERVER_NOT_BEFORE = 'tls.server.not_before';
 const TMP_TLS_SERVER_SUBJECT = 'tls.server.subject';
+const TMP_URL_DOMAIN = 'url.domain';
+const TMP_URL_EXTENSION = 'url.extension';
 const TMP_URL_FRAGMENT = 'url.fragment';
 const TMP_URL_FULL = 'url.full';
+const TMP_URL_ORIGINAL = 'url.original';
 const TMP_URL_PATH = 'url.path';
+const TMP_URL_PORT = 'url.port';
 const TMP_URL_QUERY = 'url.query';
+const TMP_URL_REGISTERED_DOMAIN = 'url.registered_domain';
 const TMP_URL_SCHEME = 'url.scheme';
+const TMP_URL_SUBDOMAIN = 'url.subdomain';
+const TMP_URL_TOP_LEVEL_DOMAIN = 'url.top_level_domain';
+const TMP_USER_AGENT_NAME = 'user_agent.name';
 const TMP_USER_AGENT_ORIGINAL = 'user_agent.original';
-const TMP_SESSION_ID = 'session.id';
-const TMP_SESSION_PREVIOUS_ID = 'session.previous_id';
+const TMP_USER_AGENT_VERSION = 'user_agent.version';
 const TMP_AWS_LAMBDA_INVOKED_ARN = 'aws.lambda.invoked_arn';
-const TMP_CLOUDEVENTS_EVENT_ID = 'cloudevents.event_id';
-const TMP_CLOUDEVENTS_EVENT_SOURCE = 'cloudevents.event_source';
-const TMP_CLOUDEVENTS_EVENT_SPEC_VERSION = 'cloudevents.event_spec_version';
-const TMP_CLOUDEVENTS_EVENT_SUBJECT = 'cloudevents.event_subject';
-const TMP_CLOUDEVENTS_EVENT_TYPE = 'cloudevents.event_type';
 const TMP_OPENTRACING_REF_TYPE = 'opentracing.ref_type';
 const TMP_OTEL_STATUS_CODE = 'otel.status_code';
 const TMP_OTEL_STATUS_DESCRIPTION = 'otel.status_description';
-const TMP_FAAS_INVOCATION_ID = 'faas.invocation_id';
-const TMP_FAAS_DOCUMENT_COLLECTION = 'faas.document.collection';
-const TMP_FAAS_DOCUMENT_NAME = 'faas.document.name';
-const TMP_FAAS_DOCUMENT_OPERATION = 'faas.document.operation';
-const TMP_FAAS_DOCUMENT_TIME = 'faas.document.time';
-const TMP_FAAS_CRON = 'faas.cron';
-const TMP_FAAS_TIME = 'faas.time';
-const TMP_FAAS_COLDSTART = 'faas.coldstart';
-const TMP_FEATURE_FLAG_KEY = 'feature_flag.key';
-const TMP_FEATURE_FLAG_PROVIDER_NAME = 'feature_flag.provider_name';
-const TMP_FEATURE_FLAG_VARIANT = 'feature_flag.variant';
 const TMP_AWS_REQUEST_ID = 'aws.request_id';
-const TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET = 'aws.dynamodb.attributes_to_get';
-const TMP_AWS_DYNAMODB_CONSISTENT_READ = 'aws.dynamodb.consistent_read';
-const TMP_AWS_DYNAMODB_CONSUMED_CAPACITY = 'aws.dynamodb.consumed_capacity';
-const TMP_AWS_DYNAMODB_INDEX_NAME = 'aws.dynamodb.index_name';
-const TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS = 'aws.dynamodb.item_collection_metrics';
-const TMP_AWS_DYNAMODB_LIMIT = 'aws.dynamodb.limit';
-const TMP_AWS_DYNAMODB_PROJECTION = 'aws.dynamodb.projection';
-const TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY = 'aws.dynamodb.provisioned_read_capacity';
-const TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY = 'aws.dynamodb.provisioned_write_capacity';
-const TMP_AWS_DYNAMODB_SELECT = 'aws.dynamodb.select';
-const TMP_AWS_DYNAMODB_TABLE_NAMES = 'aws.dynamodb.table_names';
-const TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES = 'aws.dynamodb.global_secondary_indexes';
-const TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES = 'aws.dynamodb.local_secondary_indexes';
-const TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE = 'aws.dynamodb.exclusive_start_table';
-const TMP_AWS_DYNAMODB_TABLE_COUNT = 'aws.dynamodb.table_count';
-const TMP_AWS_DYNAMODB_SCAN_FORWARD = 'aws.dynamodb.scan_forward';
-const TMP_AWS_DYNAMODB_COUNT = 'aws.dynamodb.count';
-const TMP_AWS_DYNAMODB_SCANNED_COUNT = 'aws.dynamodb.scanned_count';
-const TMP_AWS_DYNAMODB_SEGMENT = 'aws.dynamodb.segment';
-const TMP_AWS_DYNAMODB_TOTAL_SEGMENTS = 'aws.dynamodb.total_segments';
-const TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS = 'aws.dynamodb.attribute_definitions';
-const TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES = 'aws.dynamodb.global_secondary_index_updates';
 const TMP_AWS_S3_BUCKET = 'aws.s3.bucket';
 const TMP_AWS_S3_COPY_SOURCE = 'aws.s3.copy_source';
 const TMP_AWS_S3_DELETE = 'aws.s3.delete';
@@ -297,55 +422,9 @@ const TMP_MESSAGE_TYPE = 'message.type';
 const TMP_MESSAGE_UNCOMPRESSED_SIZE = 'message.uncompressed_size';
 
 /**
-* The name of the invoked function.
-*
-* Note: SHOULD be equal to the `faas.name` resource attribute of the invoked function.
-*/
-export const SEMATTRS_FAAS_INVOKED_NAME = TMP_FAAS_INVOKED_NAME;
-
-/**
-* The cloud provider of the invoked function.
-*
-* Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
-*/
-export const SEMATTRS_FAAS_INVOKED_PROVIDER = TMP_FAAS_INVOKED_PROVIDER;
-
-/**
-* The cloud region of the invoked function.
-*
-* Note: SHOULD be equal to the `cloud.region` resource attribute of the invoked function.
-*/
-export const SEMATTRS_FAAS_INVOKED_REGION = TMP_FAAS_INVOKED_REGION;
-
-/**
-* Type of the trigger which caused this function invocation.
-*/
-export const SEMATTRS_FAAS_TRIGGER = TMP_FAAS_TRIGGER;
-
-/**
-* The [`service.name`](/docs/resource/README.md#service) of the remote service. SHOULD be equal to the actual `service.name` resource attribute of the remote service if any.
-*/
-export const SEMATTRS_PEER_SERVICE = TMP_PEER_SERVICE;
-
-/**
-* Username or client_id extracted from the access token or [Authorization](https://tools.ietf.org/html/rfc7235#section-4.2) header in the inbound request from outside the system.
-*/
-export const SEMATTRS_ENDUSER_ID = TMP_ENDUSER_ID;
-
-/**
-* Actual/assumed role the client is making the request under extracted from token or application security context.
-*/
-export const SEMATTRS_ENDUSER_ROLE = TMP_ENDUSER_ROLE;
-
-/**
-* Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an [OAuth 2.0 Access Token](https://tools.ietf.org/html/rfc6749#section-3.3) or an attribute value in a [SAML 2.0 Assertion](http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html).
-*/
-export const SEMATTRS_ENDUSER_SCOPE = TMP_ENDUSER_SCOPE;
-
-/**
 * Identifies the class / type of event.
 *
-* Note: Event names are subject to the same rules as [attribute names](https://github.com/open-telemetry/opentelemetry-specification/tree/v1.26.0/specification/common/attribute-naming.md). Notably, event names are namespaced to avoid collisions and provide a clean separation of semantics for events in separate domains like browser, mobile, and kubernetes.
+* Note: Event names are subject to the same rules as [attribute names](https://github.com/open-telemetry/opentelemetry-specification/tree/v1.31.0/specification/common/attribute-naming.md). Notably, event names are namespaced to avoid collisions and provide a clean separation of semantics for events in separate domains like browser, mobile, and kubernetes.
 */
 export const SEMATTRS_EVENT_NAME = TMP_EVENT_NAME;
 
@@ -397,7 +476,7 @@ export const SEMATTRS_IOS_STATE = TMP_IOS_STATE;
 export const SEMATTRS_ANDROID_STATE = TMP_ANDROID_STATE;
 
 /**
-* The name of the connection pool; unique within the instrumented application. In case the connection pool implementation doesn&#39;t provide a name, then the [db.connection_string](/docs/database/database-spans.md#connection-level-attributes) should be used.
+* The name of the connection pool; unique within the instrumented application. In case the connection pool implementation doesn&#39;t provide a name, instrumentation should use a combination of `server.address` and `server.port` attributes formatted as `server.address:server.port`.
 */
 export const SEMATTRS_POOL_NAME = TMP_POOL_NAME;
 
@@ -405,6 +484,11 @@ export const SEMATTRS_POOL_NAME = TMP_POOL_NAME;
 * The state of a connection in the pool.
 */
 export const SEMATTRS_STATE = TMP_STATE;
+
+/**
+* Rate-limiting result, shows whether the lease was acquired or contains a rejection reason.
+*/
+export const SEMATTRS_ASPNETCORE_RATE_LIMITING_RESULT = TMP_ASPNETCORE_RATE_LIMITING_RESULT;
 
 /**
 * Full type name of the [`IExceptionHandler`](https://learn.microsoft.com/dotnet/api/microsoft.aspnetcore.diagnostics.iexceptionhandler) implementation that handled the exception.
@@ -415,11 +499,6 @@ export const SEMATTRS_ASPNETCORE_DIAGNOSTICS_HANDLER_TYPE = TMP_ASPNETCORE_DIAGN
 * Rate limiting policy name.
 */
 export const SEMATTRS_ASPNETCORE_RATE_LIMITING_POLICY = TMP_ASPNETCORE_RATE_LIMITING_POLICY;
-
-/**
-* Rate-limiting result, shows whether the lease was acquired or contains a rejection reason.
-*/
-export const SEMATTRS_ASPNETCORE_RATE_LIMITING_RESULT = TMP_ASPNETCORE_RATE_LIMITING_RESULT;
 
 /**
 * Flag indicating if request was handled by the application pipeline.
@@ -461,6 +540,11 @@ export const SEMATTRS_JVM_MEMORY_POOL_NAME = TMP_JVM_MEMORY_POOL_NAME;
 export const SEMATTRS_JVM_MEMORY_TYPE = TMP_JVM_MEMORY_TYPE;
 
 /**
+* The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+*/
+export const SEMATTRS_PROCESS_CPU_STATE = TMP_PROCESS_CPU_STATE;
+
+/**
 * The device identifier.
 */
 export const SEMATTRS_SYSTEM_DEVICE = TMP_SYSTEM_DEVICE;
@@ -471,7 +555,7 @@ export const SEMATTRS_SYSTEM_DEVICE = TMP_SYSTEM_DEVICE;
 export const SEMATTRS_SYSTEM_CPU_LOGICAL_NUMBER = TMP_SYSTEM_CPU_LOGICAL_NUMBER;
 
 /**
-* The state of the CPU.
+* The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
 */
 export const SEMATTRS_SYSTEM_CPU_STATE = TMP_SYSTEM_CPU_STATE;
 
@@ -523,7 +607,151 @@ export const SEMATTRS_SYSTEM_NETWORK_STATE = TMP_SYSTEM_NETWORK_STATE;
 /**
 * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
 */
-export const SEMATTRS_SYSTEM_PROCESSES_STATUS = TMP_SYSTEM_PROCESSES_STATUS;
+export const SEMATTRS_SYSTEM_PROCESS_STATUS = TMP_SYSTEM_PROCESS_STATUS;
+
+/**
+* Uniquely identifies the framework API revision offered by a version (`os.version`) of the android operating system. More information can be found [here](https://developer.android.com/guide/topics/manifest/uses-sdk-element#ApiLevels).
+*/
+export const SEMATTRS_ANDROID_OS_API_LEVEL = TMP_ANDROID_OS_API_LEVEL;
+
+/**
+* The JSON-serialized value of each item in the `AttributeDefinitions` request field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS = TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS;
+
+/**
+* The value of the `AttributesToGet` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_ATTRIBUTES_TO_GET = TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET;
+
+/**
+* The value of the `ConsistentRead` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_CONSISTENT_READ = TMP_AWS_DYNAMODB_CONSISTENT_READ;
+
+/**
+* The JSON-serialized value of each item in the `ConsumedCapacity` response field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_CONSUMED_CAPACITY = TMP_AWS_DYNAMODB_CONSUMED_CAPACITY;
+
+/**
+* The value of the `Count` response parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_COUNT = TMP_AWS_DYNAMODB_COUNT;
+
+/**
+* The value of the `ExclusiveStartTableName` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_EXCLUSIVE_START_TABLE = TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE;
+
+/**
+* The JSON-serialized value of each item in the `GlobalSecondaryIndexUpdates` request field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES = TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES;
+
+/**
+* The JSON-serialized value of each item of the `GlobalSecondaryIndexes` request field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES = TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES;
+
+/**
+* The value of the `IndexName` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_INDEX_NAME = TMP_AWS_DYNAMODB_INDEX_NAME;
+
+/**
+* The JSON-serialized value of the `ItemCollectionMetrics` response field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_ITEM_COLLECTION_METRICS = TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS;
+
+/**
+* The value of the `Limit` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_LIMIT = TMP_AWS_DYNAMODB_LIMIT;
+
+/**
+* The JSON-serialized value of each item of the `LocalSecondaryIndexes` request field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES = TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES;
+
+/**
+* The value of the `ProjectionExpression` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_PROJECTION = TMP_AWS_DYNAMODB_PROJECTION;
+
+/**
+* The value of the `ProvisionedThroughput.ReadCapacityUnits` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY = TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY;
+
+/**
+* The value of the `ProvisionedThroughput.WriteCapacityUnits` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY = TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY;
+
+/**
+* The value of the `ScanIndexForward` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_SCAN_FORWARD = TMP_AWS_DYNAMODB_SCAN_FORWARD;
+
+/**
+* The value of the `ScannedCount` response parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_SCANNED_COUNT = TMP_AWS_DYNAMODB_SCANNED_COUNT;
+
+/**
+* The value of the `Segment` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_SEGMENT = TMP_AWS_DYNAMODB_SEGMENT;
+
+/**
+* The value of the `Select` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_SELECT = TMP_AWS_DYNAMODB_SELECT;
+
+/**
+* The number of items in the `TableNames` response parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_TABLE_COUNT = TMP_AWS_DYNAMODB_TABLE_COUNT;
+
+/**
+* The keys in the `RequestItems` object field.
+*/
+export const SEMATTRS_AWS_DYNAMODB_TABLE_NAMES = TMP_AWS_DYNAMODB_TABLE_NAMES;
+
+/**
+* The value of the `TotalSegments` request parameter.
+*/
+export const SEMATTRS_AWS_DYNAMODB_TOTAL_SEGMENTS = TMP_AWS_DYNAMODB_TOTAL_SEGMENTS;
+
+/**
+* Array of brand name and version separated by a space.
+*
+* Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.brands`).
+*/
+export const SEMATTRS_BROWSER_BRANDS = TMP_BROWSER_BRANDS;
+
+/**
+* Preferred language of the user using the browser.
+*
+* Note: This value is intended to be taken from the Navigator API `navigator.language`.
+*/
+export const SEMATTRS_BROWSER_LANGUAGE = TMP_BROWSER_LANGUAGE;
+
+/**
+* A boolean that is true if the browser is running on a mobile device.
+*
+* Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.mobile`). If unavailable, this attribute SHOULD be left unset.
+*/
+export const SEMATTRS_BROWSER_MOBILE = TMP_BROWSER_MOBILE;
+
+/**
+* The platform on which the browser is running.
+*
+* Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.platform`). If unavailable, the legacy `navigator.platform` API SHOULD NOT be used instead and this attribute SHOULD be left unset in order for the values to be consistent.
+The list of possible values is defined in the [W3C User-Agent Client Hints specification](https://wicg.github.io/ua-client-hints/#sec-ch-ua-platform). Note that some (but not all) of these values can overlap with values in the [`os.type` and `os.name` attributes](./os.md). However, for consistency, the values in the `browser.platform` attribute should capture the exact value that the user agent provides.
+*/
+export const SEMATTRS_BROWSER_PLATFORM = TMP_BROWSER_PLATFORM;
 
 /**
 * Client address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
@@ -538,6 +766,85 @@ export const SEMATTRS_CLIENT_ADDRESS = TMP_CLIENT_ADDRESS;
 * Note: When observed from the server side, and when communicating through an intermediary, `client.port` SHOULD represent the client port behind any intermediaries,  for example proxies, if it&#39;s available.
 */
 export const SEMATTRS_CLIENT_PORT = TMP_CLIENT_PORT;
+
+/**
+* The cloud account ID the resource is assigned to.
+*/
+export const SEMATTRS_CLOUD_ACCOUNT_ID = TMP_CLOUD_ACCOUNT_ID;
+
+/**
+* Cloud regions often have multiple, isolated locations known as zones to increase availability. Availability zone represents the zone where the resource is running.
+*
+* Note: Availability zones are called &#34;zones&#34; on Alibaba Cloud and Google Cloud.
+*/
+export const SEMATTRS_CLOUD_AVAILABILITY_ZONE = TMP_CLOUD_AVAILABILITY_ZONE;
+
+/**
+* The cloud platform in use.
+*
+* Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+*/
+export const SEMATTRS_CLOUD_PLATFORM = TMP_CLOUD_PLATFORM;
+
+/**
+* Name of the cloud provider.
+*/
+export const SEMATTRS_CLOUD_PROVIDER = TMP_CLOUD_PROVIDER;
+
+/**
+* The geographical region the resource is running.
+*
+* Note: Refer to your provider&#39;s docs to see the available regions, for example [Alibaba Cloud regions](https://www.alibabacloud.com/help/doc-detail/40654.htm), [AWS regions](https://aws.amazon.com/about-aws/global-infrastructure/regions_az/), [Azure regions](https://azure.microsoft.com/global-infrastructure/geographies/), [Google Cloud regions](https://cloud.google.com/about/locations), or [Tencent Cloud regions](https://www.tencentcloud.com/document/product/213/6091).
+*/
+export const SEMATTRS_CLOUD_REGION = TMP_CLOUD_REGION;
+
+/**
+* Cloud provider-specific native identifier of the monitored cloud resource (e.g. an [ARN](https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html) on AWS, a [fully qualified resource ID](https://learn.microsoft.com/rest/api/resources/resources/get-by-id) on Azure, a [full resource name](https://cloud.google.com/apis/design/resource_names#full_resource_name) on GCP).
+*
+* Note: On some cloud providers, it may not be possible to determine the full ID at startup,
+so it may be necessary to set `cloud.resource_id` as a span attribute instead.
+
+The exact value to use for `cloud.resource_id` depends on the cloud provider.
+The following well-known definitions MUST be used if you set this attribute and they apply:
+
+* **AWS Lambda:** The function [ARN](https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html).
+  Take care not to use the &#34;invoked ARN&#34; directly but replace any
+  [alias suffix](https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html)
+  with the resolved function version, as the same runtime instance may be invokable with
+  multiple different aliases.
+* **GCP:** The [URI of the resource](https://cloud.google.com/iam/docs/full-resource-names)
+* **Azure:** The [Fully Qualified Resource ID](https://docs.microsoft.com/rest/api/resources/resources/get-by-id) of the invoked function,
+  *not* the function app, having the form
+  `/subscriptions/&lt;SUBSCIPTION_GUID&gt;/resourceGroups/&lt;RG&gt;/providers/Microsoft.Web/sites/&lt;FUNCAPP&gt;/functions/&lt;FUNC&gt;`.
+  This means that a span attribute MUST be used, as an Azure function app can host multiple functions that would usually share
+  a TracerProvider.
+*/
+export const SEMATTRS_CLOUD_RESOURCE_ID = TMP_CLOUD_RESOURCE_ID;
+
+/**
+* The [event_id](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id) uniquely identifies the event.
+*/
+export const SEMATTRS_CLOUDEVENTS_EVENT_ID = TMP_CLOUDEVENTS_EVENT_ID;
+
+/**
+* The [source](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1) identifies the context in which an event happened.
+*/
+export const SEMATTRS_CLOUDEVENTS_EVENT_SOURCE = TMP_CLOUDEVENTS_EVENT_SOURCE;
+
+/**
+* The [version of the CloudEvents specification](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion) which the event uses.
+*/
+export const SEMATTRS_CLOUDEVENTS_EVENT_SPEC_VERSION = TMP_CLOUDEVENTS_EVENT_SPEC_VERSION;
+
+/**
+* The [subject](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject) of the event in the context of the event producer (identified by source).
+*/
+export const SEMATTRS_CLOUDEVENTS_EVENT_SUBJECT = TMP_CLOUDEVENTS_EVENT_SUBJECT;
+
+/**
+* The [event_type](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type) contains a value describing the type of event related to the originating occurrence.
+*/
+export const SEMATTRS_CLOUDEVENTS_EVENT_TYPE = TMP_CLOUDEVENTS_EVENT_TYPE;
 
 /**
 * The column number in `code.filepath` best representing the operation. It SHOULD point within the code unit named in `code.function`.
@@ -568,6 +875,69 @@ export const SEMATTRS_CODE_NAMESPACE = TMP_CODE_NAMESPACE;
 * A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.
 */
 export const SEMATTRS_CODE_STACKTRACE = TMP_CODE_STACKTRACE;
+
+/**
+* The command used to run the container (i.e. the command name).
+*
+* Note: If using embedded credentials or sensitive data, it is recommended to remove them to prevent potential leakage.
+*/
+export const SEMATTRS_CONTAINER_COMMAND = TMP_CONTAINER_COMMAND;
+
+/**
+* All the command arguments (including the command/executable itself) run by the container. [2].
+*/
+export const SEMATTRS_CONTAINER_COMMAND_ARGS = TMP_CONTAINER_COMMAND_ARGS;
+
+/**
+* The full command run by the container as a single string representing the full command. [2].
+*/
+export const SEMATTRS_CONTAINER_COMMAND_LINE = TMP_CONTAINER_COMMAND_LINE;
+
+/**
+* The CPU state for this data point.
+*/
+export const SEMATTRS_CONTAINER_CPU_STATE = TMP_CONTAINER_CPU_STATE;
+
+/**
+* Container ID. Usually a UUID, as for example used to [identify Docker containers](https://docs.docker.com/engine/reference/run/#container-identification). The UUID might be abbreviated.
+*/
+export const SEMATTRS_CONTAINER_ID = TMP_CONTAINER_ID;
+
+/**
+* Runtime specific image identifier. Usually a hash algorithm followed by a UUID.
+*
+* Note: Docker defines a sha256 of the image id; `container.image.id` corresponds to the `Image` field from the Docker container inspect [API](https://docs.docker.com/engine/api/v1.43/#tag/Container/operation/ContainerInspect) endpoint.
+K8s defines a link to the container registry repository with digest `&#34;imageID&#34;: &#34;registry.azurecr.io /namespace/service/dockerfile@sha256:bdeabd40c3a8a492eaf9e8e44d0ebbb84bac7ee25ac0cf8a7159d25f62555625&#34;`.
+The ID is assinged by the container runtime and can vary in different environments. Consider using `oci.manifest.digest` if it is important to identify the same image in different environments/runtimes.
+*/
+export const SEMATTRS_CONTAINER_IMAGE_ID = TMP_CONTAINER_IMAGE_ID;
+
+/**
+* Name of the image the container was built on.
+*/
+export const SEMATTRS_CONTAINER_IMAGE_NAME = TMP_CONTAINER_IMAGE_NAME;
+
+/**
+* Repo digests of the container image as provided by the container runtime.
+*
+* Note: [Docker](https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect) and [CRI](https://github.com/kubernetes/cri-api/blob/c75ef5b473bbe2d0a4fc92f82235efd665ea8e9f/pkg/apis/runtime/v1/api.proto#L1237-L1238) report those under the `RepoDigests` field.
+*/
+export const SEMATTRS_CONTAINER_IMAGE_REPO_DIGESTS = TMP_CONTAINER_IMAGE_REPO_DIGESTS;
+
+/**
+* Container image tags. An example can be found in [Docker Image Inspect](https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect). Should be only the `&lt;tag&gt;` section of the full name for example from `registry.example.com/my-org/my-image:&lt;tag&gt;`.
+*/
+export const SEMATTRS_CONTAINER_IMAGE_TAGS = TMP_CONTAINER_IMAGE_TAGS;
+
+/**
+* Container name used by container runtime.
+*/
+export const SEMATTRS_CONTAINER_NAME = TMP_CONTAINER_NAME;
+
+/**
+* The container runtime managing this container.
+*/
+export const SEMATTRS_CONTAINER_RUNTIME = TMP_CONTAINER_RUNTIME;
 
 /**
 * The consistency level of the query. Based on consistency values from [CQL](https://docs.datastax.com/en/cassandra-oss/3.0/cassandra/dml/dmlConfigConsistency.html).
@@ -605,11 +975,6 @@ export const SEMATTRS_DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT = TMP_DB_CASSANDR
 * Note: This mirrors the db.sql.table attribute but references cassandra rather than sql. It is not recommended to attempt any client-side parsing of `db.statement` just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.
 */
 export const SEMATTRS_DB_CASSANDRA_TABLE = TMP_DB_CASSANDRA_TABLE;
-
-/**
-* The connection string used to connect to the database. It is recommended to remove embedded credentials.
-*/
-export const SEMATTRS_DB_CONNECTION_STRING = TMP_DB_CONNECTION_STRING;
 
 /**
 * Unique Cosmos client instance id.
@@ -657,19 +1022,9 @@ export const SEMATTRS_DB_COSMOSDB_SUB_STATUS_CODE = TMP_DB_COSMOSDB_SUB_STATUS_C
 export const SEMATTRS_DB_ELASTICSEARCH_CLUSTER_NAME = TMP_DB_ELASTICSEARCH_CLUSTER_NAME;
 
 /**
-* Represents the human-readable identifier of the node/instance to which a request was routed.
-*/
-export const SEMATTRS_DB_ELASTICSEARCH_NODE_NAME = TMP_DB_ELASTICSEARCH_NODE_NAME;
-
-/**
 * An identifier (address, unique name, or any other identifier) of the database instance that is executing queries or mutations on the current connection. This is useful in cases where the database is running in a clustered environment and the instrumentation is able to record the node executing the query. The client may obtain this value in databases like MySQL using queries like `select @@hostname`.
 */
 export const SEMATTRS_DB_INSTANCE_ID = TMP_DB_INSTANCE_ID;
-
-/**
-* The fully-qualified class name of the [Java Database Connectivity (JDBC)](https://docs.oracle.com/javase/8/docs/technotes/guides/jdbc/) driver used to connect.
-*/
-export const SEMATTRS_DB_JDBC_DRIVER_CLASSNAME = TMP_DB_JDBC_DRIVER_CLASSNAME;
 
 /**
 * The MongoDB collection being accessed within the database stated in `db.name`.
@@ -723,6 +1078,40 @@ export const SEMATTRS_DB_SYSTEM = TMP_DB_SYSTEM;
 * Username for accessing the database.
 */
 export const SEMATTRS_DB_USER = TMP_DB_USER;
+
+/**
+* Name of the [deployment environment](https://wikipedia.org/wiki/Deployment_environment) (aka deployment tier).
+*
+* Note: `deployment.environment` does not affect the uniqueness constraints defined through
+the `service.namespace`, `service.name` and `service.instance.id` resource attributes.
+This implies that resources carrying the following attribute combinations MUST be
+considered to be identifying the same service:
+
+* `service.name=frontend`, `deployment.environment=production`
+* `service.name=frontend`, `deployment.environment=staging`.
+*/
+export const SEMATTRS_DEPLOYMENT_ENVIRONMENT = TMP_DEPLOYMENT_ENVIRONMENT;
+
+/**
+* Deprecated, use `server.address`, `server.port` attributes instead.
+*
+* @deprecated "Replaced by `server.address` and `server.port`.".
+*/
+export const SEMATTRS_DB_CONNECTION_STRING = TMP_DB_CONNECTION_STRING;
+
+/**
+* Deprecated, use `db.instance.id` instead.
+*
+* @deprecated Replaced by `db.instance.id`.
+*/
+export const SEMATTRS_DB_ELASTICSEARCH_NODE_NAME = TMP_DB_ELASTICSEARCH_NODE_NAME;
+
+/**
+* Removed, no replacement at this time.
+*
+* @deprecated Removed as not used.
+*/
+export const SEMATTRS_DB_JDBC_DRIVER_CLASSNAME = TMP_DB_JDBC_DRIVER_CLASSNAME;
 
 /**
 * Deprecated, use `network.protocol.name` instead.
@@ -786,6 +1175,13 @@ export const SEMATTRS_HTTP_URL = TMP_HTTP_URL;
 * @deprecated Replaced by `user_agent.original`.
 */
 export const SEMATTRS_HTTP_USER_AGENT = TMP_HTTP_USER_AGENT;
+
+/**
+* &#34;Deprecated, use `messaging.destination.partition.id` instead.&#34;.
+*
+* @deprecated Replaced by `messaging.destination.partition.id`.
+*/
+export const SEMATTRS_MESSAGING_KAFKA_DESTINATION_PARTITION = TMP_MESSAGING_KAFKA_DESTINATION_PARTITION;
 
 /**
 * Deprecated, use `server.address`.
@@ -879,6 +1275,13 @@ export const SEMATTRS_NET_SOCK_PEER_PORT = TMP_NET_SOCK_PEER_PORT;
 export const SEMATTRS_NET_TRANSPORT = TMP_NET_TRANSPORT;
 
 /**
+* Deprecated, use `system.process.status` instead.
+*
+* @deprecated Replaced by `system.process.status`.
+*/
+export const SEMATTRS_SYSTEM_PROCESSES_STATUS = TMP_SYSTEM_PROCESSES_STATUS;
+
+/**
 * Destination address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
 *
 * Note: When observed from the source side, and when communicating through an intermediary, `destination.address` SHOULD represent the destination address behind any intermediaries, for example proxies, if it&#39;s available.
@@ -891,9 +1294,59 @@ export const SEMATTRS_DESTINATION_ADDRESS = TMP_DESTINATION_ADDRESS;
 export const SEMATTRS_DESTINATION_PORT = TMP_DESTINATION_PORT;
 
 /**
+* A unique identifier representing the device.
+*
+* Note: The device identifier MUST only be defined using the values outlined below. This value is not an advertising identifier and MUST NOT be used as such. On iOS (Swift or Objective-C), this value MUST be equal to the [vendor identifier](https://developer.apple.com/documentation/uikit/uidevice/1620059-identifierforvendor). On Android (Java or Kotlin), this value MUST be equal to the Firebase Installation ID or a globally unique UUID which is persisted across sessions in your application. More information can be found [here](https://developer.android.com/training/articles/user-data-ids) on best practices and exact implementation details. Caution should be taken when storing personal data or anything which can identify a user. GDPR and data protection laws may apply, ensure you do your own due diligence.
+*/
+export const SEMATTRS_DEVICE_ID = TMP_DEVICE_ID;
+
+/**
+* The name of the device manufacturer.
+*
+* Note: The Android OS provides this field via [Build](https://developer.android.com/reference/android/os/Build#MANUFACTURER). iOS apps SHOULD hardcode the value `Apple`.
+*/
+export const SEMATTRS_DEVICE_MANUFACTURER = TMP_DEVICE_MANUFACTURER;
+
+/**
+* The model identifier for the device.
+*
+* Note: It&#39;s recommended this value represents a machine-readable version of the model identifier rather than the market or consumer-friendly name of the device.
+*/
+export const SEMATTRS_DEVICE_MODEL_IDENTIFIER = TMP_DEVICE_MODEL_IDENTIFIER;
+
+/**
+* The marketing name for the device model.
+*
+* Note: It&#39;s recommended this value represents a human-readable version of the device model rather than a machine-readable alternative.
+*/
+export const SEMATTRS_DEVICE_MODEL_NAME = TMP_DEVICE_MODEL_NAME;
+
+/**
 * The disk IO operation direction.
 */
 export const SEMATTRS_DISK_IO_DIRECTION = TMP_DISK_IO_DIRECTION;
+
+/**
+* The name being queried.
+*
+* Note: If the name field contains non-printable characters (below 32 or above 126), those characters should be represented as escaped base 10 integers (\DDD). Back slashes and quotes should be escaped. Tabs, carriage returns, and line feeds should be converted to \t, \r, and \n respectively.
+*/
+export const SEMATTRS_DNS_QUESTION_NAME = TMP_DNS_QUESTION_NAME;
+
+/**
+* Username or client_id extracted from the access token or [Authorization](https://tools.ietf.org/html/rfc7235#section-4.2) header in the inbound request from outside the system.
+*/
+export const SEMATTRS_ENDUSER_ID = TMP_ENDUSER_ID;
+
+/**
+* Actual/assumed role the client is making the request under extracted from token or application security context.
+*/
+export const SEMATTRS_ENDUSER_ROLE = TMP_ENDUSER_ROLE;
+
+/**
+* Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an [OAuth 2.0 Access Token](https://tools.ietf.org/html/rfc6749#section-3.3) or an attribute value in a [SAML 2.0 Assertion](http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html).
+*/
+export const SEMATTRS_ENDUSER_SCOPE = TMP_ENDUSER_SCOPE;
 
 /**
 * Describes a class of error the operation ended with.
@@ -954,6 +1407,280 @@ export const SEMATTRS_EXCEPTION_STACKTRACE = TMP_EXCEPTION_STACKTRACE;
 export const SEMATTRS_EXCEPTION_TYPE = TMP_EXCEPTION_TYPE;
 
 /**
+* A boolean that is true if the serverless function is executed for the first time (aka cold-start).
+*/
+export const SEMATTRS_FAAS_COLDSTART = TMP_FAAS_COLDSTART;
+
+/**
+* A string containing the schedule period as [Cron Expression](https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm).
+*/
+export const SEMATTRS_FAAS_CRON = TMP_FAAS_CRON;
+
+/**
+* The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
+*/
+export const SEMATTRS_FAAS_DOCUMENT_COLLECTION = TMP_FAAS_DOCUMENT_COLLECTION;
+
+/**
+* The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.
+*/
+export const SEMATTRS_FAAS_DOCUMENT_NAME = TMP_FAAS_DOCUMENT_NAME;
+
+/**
+* Describes the type of the operation that was performed on the data.
+*/
+export const SEMATTRS_FAAS_DOCUMENT_OPERATION = TMP_FAAS_DOCUMENT_OPERATION;
+
+/**
+* A string containing the time when the data was accessed in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
+*/
+export const SEMATTRS_FAAS_DOCUMENT_TIME = TMP_FAAS_DOCUMENT_TIME;
+
+/**
+* The execution environment ID as a string, that will be potentially reused for other invocations to the same function/function version.
+*
+* Note: * **AWS Lambda:** Use the (full) log stream name.
+*/
+export const SEMATTRS_FAAS_INSTANCE = TMP_FAAS_INSTANCE;
+
+/**
+* The invocation ID of the current function invocation.
+*/
+export const SEMATTRS_FAAS_INVOCATION_ID = TMP_FAAS_INVOCATION_ID;
+
+/**
+* The name of the invoked function.
+*
+* Note: SHOULD be equal to the `faas.name` resource attribute of the invoked function.
+*/
+export const SEMATTRS_FAAS_INVOKED_NAME = TMP_FAAS_INVOKED_NAME;
+
+/**
+* The cloud provider of the invoked function.
+*
+* Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+*/
+export const SEMATTRS_FAAS_INVOKED_PROVIDER = TMP_FAAS_INVOKED_PROVIDER;
+
+/**
+* The cloud region of the invoked function.
+*
+* Note: SHOULD be equal to the `cloud.region` resource attribute of the invoked function.
+*/
+export const SEMATTRS_FAAS_INVOKED_REGION = TMP_FAAS_INVOKED_REGION;
+
+/**
+* The amount of memory available to the serverless function converted to Bytes.
+*
+* Note: It&#39;s recommended to set this attribute since e.g. too little memory can easily stop a Java AWS Lambda function from working correctly. On AWS Lambda, the environment variable `AWS_LAMBDA_FUNCTION_MEMORY_SIZE` provides this information (which must be multiplied by 1,048,576).
+*/
+export const SEMATTRS_FAAS_MAX_MEMORY = TMP_FAAS_MAX_MEMORY;
+
+/**
+* The name of the single function that this runtime instance executes.
+*
+* Note: This is the name of the function as configured/deployed on the FaaS
+platform and is usually different from the name of the callback
+function (which may be stored in the
+[`code.namespace`/`code.function`](/docs/general/attributes.md#source-code-attributes)
+span attributes).
+
+For some cloud providers, the above definition is ambiguous. The following
+definition of function name MUST be used for this attribute
+(and consequently the span name) for the listed cloud providers/products:
+
+* **Azure:**  The full name `&lt;FUNCAPP&gt;/&lt;FUNC&gt;`, i.e., function app name
+  followed by a forward slash followed by the function name (this form
+  can also be seen in the resource JSON for the function).
+  This means that a span attribute MUST be used, as an Azure function
+  app can host multiple functions that would usually share
+  a TracerProvider (see also the `cloud.resource_id` attribute).
+*/
+export const SEMATTRS_FAAS_NAME = TMP_FAAS_NAME;
+
+/**
+* A string containing the function invocation time in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
+*/
+export const SEMATTRS_FAAS_TIME = TMP_FAAS_TIME;
+
+/**
+* Type of the trigger which caused this function invocation.
+*/
+export const SEMATTRS_FAAS_TRIGGER = TMP_FAAS_TRIGGER;
+
+/**
+* The immutable version of the function being executed.
+*
+* Note: Depending on the cloud provider and platform, use:
+
+* **AWS Lambda:** The [function version](https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html)
+  (an integer represented as a decimal string).
+* **Google Cloud Run (Services):** The [revision](https://cloud.google.com/run/docs/managing/revisions)
+  (i.e., the function name plus the revision suffix).
+* **Google Cloud Functions:** The value of the
+  [`K_REVISION` environment variable](https://cloud.google.com/functions/docs/env-var#runtime_environment_variables_set_automatically).
+* **Azure Functions:** Not applicable. Do not set this attribute.
+*/
+export const SEMATTRS_FAAS_VERSION = TMP_FAAS_VERSION;
+
+/**
+* The unique identifier of the feature flag.
+*/
+export const SEMATTRS_FEATURE_FLAG_KEY = TMP_FEATURE_FLAG_KEY;
+
+/**
+* The name of the service provider that performs the flag evaluation.
+*/
+export const SEMATTRS_FEATURE_FLAG_PROVIDER_NAME = TMP_FEATURE_FLAG_PROVIDER_NAME;
+
+/**
+* SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.
+*
+* Note: A semantic identifier, commonly referred to as a variant, provides a means
+for referring to a value without including the value itself. This can
+provide additional context for understanding the meaning behind a value.
+For example, the variant `red` maybe be used for the value `#c05543`.
+
+A stringified version of the value can be used in situations where a
+semantic identifier is unavailable. String representation of the value
+should be determined by the implementer.
+*/
+export const SEMATTRS_FEATURE_FLAG_VARIANT = TMP_FEATURE_FLAG_VARIANT;
+
+/**
+* Directory where the file is located. It should include the drive letter, when appropriate.
+*/
+export const SEMATTRS_FILE_DIRECTORY = TMP_FILE_DIRECTORY;
+
+/**
+* File extension, excluding the leading dot.
+*
+* Note: When the file name has multiple extensions (example.tar.gz), only the last one should be captured (&#34;gz&#34;, not &#34;tar.gz&#34;).
+*/
+export const SEMATTRS_FILE_EXTENSION = TMP_FILE_EXTENSION;
+
+/**
+* Name of the file including the extension, without the directory.
+*/
+export const SEMATTRS_FILE_NAME = TMP_FILE_NAME;
+
+/**
+* Full path to the file, including the file name. It should include the drive letter, when appropriate.
+*/
+export const SEMATTRS_FILE_PATH = TMP_FILE_PATH;
+
+/**
+* File size in bytes.
+*/
+export const SEMATTRS_FILE_SIZE = TMP_FILE_SIZE;
+
+/**
+* The name of the Cloud Run [execution](https://cloud.google.com/run/docs/managing/job-executions) being run for the Job, as set by the [`CLOUD_RUN_EXECUTION`](https://cloud.google.com/run/docs/container-contract#jobs-env-vars) environment variable.
+*/
+export const SEMATTRS_GCP_CLOUD_RUN_JOB_EXECUTION = TMP_GCP_CLOUD_RUN_JOB_EXECUTION;
+
+/**
+* The index for a task within an execution as provided by the [`CLOUD_RUN_TASK_INDEX`](https://cloud.google.com/run/docs/container-contract#jobs-env-vars) environment variable.
+*/
+export const SEMATTRS_GCP_CLOUD_RUN_JOB_TASK_INDEX = TMP_GCP_CLOUD_RUN_JOB_TASK_INDEX;
+
+/**
+* The hostname of a GCE instance. This is the full value of the default or [custom hostname](https://cloud.google.com/compute/docs/instances/custom-hostname-vm).
+*/
+export const SEMATTRS_GCP_GCE_INSTANCE_HOSTNAME = TMP_GCP_GCE_INSTANCE_HOSTNAME;
+
+/**
+* The instance name of a GCE instance. This is the value provided by `host.name`, the visible name of the instance in the Cloud Console UI, and the prefix for the default hostname of the instance as defined by the [default internal DNS name](https://cloud.google.com/compute/docs/internal-dns#instance-fully-qualified-domain-names).
+*/
+export const SEMATTRS_GCP_GCE_INSTANCE_NAME = TMP_GCP_GCE_INSTANCE_NAME;
+
+/**
+* The CPU architecture the host system is running on.
+*/
+export const SEMATTRS_HOST_ARCH = TMP_HOST_ARCH;
+
+/**
+* The amount of level 2 memory cache available to the processor (in Bytes).
+*/
+export const SEMATTRS_HOST_CPU_CACHE_L2_SIZE = TMP_HOST_CPU_CACHE_L2_SIZE;
+
+/**
+* Family or generation of the CPU.
+*/
+export const SEMATTRS_HOST_CPU_FAMILY = TMP_HOST_CPU_FAMILY;
+
+/**
+* Model identifier. It provides more granular information about the CPU, distinguishing it from other CPUs within the same family.
+*/
+export const SEMATTRS_HOST_CPU_MODEL_ID = TMP_HOST_CPU_MODEL_ID;
+
+/**
+* Model designation of the processor.
+*/
+export const SEMATTRS_HOST_CPU_MODEL_NAME = TMP_HOST_CPU_MODEL_NAME;
+
+/**
+* Stepping or core revisions.
+*/
+export const SEMATTRS_HOST_CPU_STEPPING = TMP_HOST_CPU_STEPPING;
+
+/**
+* Processor manufacturer identifier. A maximum 12-character string.
+*
+* Note: [CPUID](https://wiki.osdev.org/CPUID) command returns the vendor ID string in EBX, EDX and ECX registers. Writing these to memory in this order results in a 12-character string.
+*/
+export const SEMATTRS_HOST_CPU_VENDOR_ID = TMP_HOST_CPU_VENDOR_ID;
+
+/**
+* Unique host ID. For Cloud, this must be the instance_id assigned by the cloud provider. For non-containerized systems, this should be the `machine-id`. See the table below for the sources to use to determine the `machine-id` based on operating system.
+*/
+export const SEMATTRS_HOST_ID = TMP_HOST_ID;
+
+/**
+* VM image ID or host OS image ID. For Cloud, this value is from the provider.
+*/
+export const SEMATTRS_HOST_IMAGE_ID = TMP_HOST_IMAGE_ID;
+
+/**
+* Name of the VM image or OS install the host was instantiated from.
+*/
+export const SEMATTRS_HOST_IMAGE_NAME = TMP_HOST_IMAGE_NAME;
+
+/**
+* The version string of the VM image or host OS as defined in [Version Attributes](/docs/resource/README.md#version-attributes).
+*/
+export const SEMATTRS_HOST_IMAGE_VERSION = TMP_HOST_IMAGE_VERSION;
+
+/**
+* Available IP addresses of the host, excluding loopback interfaces.
+*
+* Note: IPv4 Addresses MUST be specified in dotted-quad notation. IPv6 addresses MUST be specified in the [RFC 5952](https://www.rfc-editor.org/rfc/rfc5952.html) format.
+*/
+export const SEMATTRS_HOST_IP = TMP_HOST_IP;
+
+/**
+* Available MAC addresses of the host, excluding loopback interfaces.
+*
+* Note: MAC Addresses MUST be represented in [IEEE RA hexadecimal form](https://standards.ieee.org/wp-content/uploads/import/documents/tutorials/eui.pdf): as hyphen-separated octets in uppercase hexadecimal form from most to least significant.
+*/
+export const SEMATTRS_HOST_MAC = TMP_HOST_MAC;
+
+/**
+* Name of the host. On Unix systems, it may contain what the hostname command returns, or the fully qualified hostname, or another name specified by the user.
+*/
+export const SEMATTRS_HOST_NAME = TMP_HOST_NAME;
+
+/**
+* Type of host. For Cloud, this must be the machine type.
+*/
+export const SEMATTRS_HOST_TYPE = TMP_HOST_TYPE;
+
+/**
+* State of the HTTP connection in the HTTP connection pool.
+*/
+export const SEMATTRS_HTTP_CONNECTION_STATE = TMP_HTTP_CONNECTION_STATE;
+
+/**
 * The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the [Content-Length](https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length) header. For requests using transport encoding, this should be the compressed size.
 */
 export const SEMATTRS_HTTP_REQUEST_BODY_SIZE = TMP_HTTP_REQUEST_BODY_SIZE;
@@ -991,9 +1718,19 @@ export const SEMATTRS_HTTP_REQUEST_METHOD_ORIGINAL = TMP_HTTP_REQUEST_METHOD_ORI
 export const SEMATTRS_HTTP_REQUEST_RESEND_COUNT = TMP_HTTP_REQUEST_RESEND_COUNT;
 
 /**
+* The total size of the request in bytes. This should be the total number of bytes sent over the wire, including the request line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and request body if any.
+*/
+export const SEMATTRS_HTTP_REQUEST_SIZE = TMP_HTTP_REQUEST_SIZE;
+
+/**
 * The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the [Content-Length](https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length) header. For requests using transport encoding, this should be the compressed size.
 */
 export const SEMATTRS_HTTP_RESPONSE_BODY_SIZE = TMP_HTTP_RESPONSE_BODY_SIZE;
+
+/**
+* The total size of the response in bytes. This should be the total number of bytes sent over the wire, including the status line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and response body and trailers if any.
+*/
+export const SEMATTRS_HTTP_RESPONSE_SIZE = TMP_HTTP_RESPONSE_SIZE;
 
 /**
 * [HTTP response status code](https://tools.ietf.org/html/rfc7231#section-6).
@@ -1007,6 +1744,134 @@ export const SEMATTRS_HTTP_RESPONSE_STATUS_CODE = TMP_HTTP_RESPONSE_STATUS_CODE;
 SHOULD include the [application root](/docs/http/http-spans.md#http-server-definitions) if there is one.
 */
 export const SEMATTRS_HTTP_ROUTE = TMP_HTTP_ROUTE;
+
+/**
+* The name of the cluster.
+*/
+export const SEMATTRS_K8S_CLUSTER_NAME = TMP_K8S_CLUSTER_NAME;
+
+/**
+* A pseudo-ID for the cluster, set to the UID of the `kube-system` namespace.
+*
+* Note: K8s doesn&#39;t have support for obtaining a cluster ID. If this is ever
+added, we will recommend collecting the `k8s.cluster.uid` through the
+official APIs. In the meantime, we are able to use the `uid` of the
+`kube-system` namespace as a proxy for cluster ID. Read on for the
+rationale.
+
+Every object created in a K8s cluster is assigned a distinct UID. The
+`kube-system` namespace is used by Kubernetes itself and will exist
+for the lifetime of the cluster. Using the `uid` of the `kube-system`
+namespace is a reasonable proxy for the K8s ClusterID as it will only
+change if the cluster is rebuilt. Furthermore, Kubernetes UIDs are
+UUIDs as standardized by
+[ISO/IEC 9834-8 and ITU-T X.667](https://www.itu.int/ITU-T/studygroups/com17/oid.html).
+Which states:
+
+&gt; If generated according to one of the mechanisms defined in Rec.
+  ITU-T X.667 | ISO/IEC 9834-8, a UUID is either guaranteed to be
+  different from all other UUIDs generated before 3603 A.D., or is
+  extremely likely to be different (depending on the mechanism chosen).
+
+Therefore, UIDs between clusters should be extremely unlikely to
+conflict.
+*/
+export const SEMATTRS_K8S_CLUSTER_UID = TMP_K8S_CLUSTER_UID;
+
+/**
+* The name of the Container from Pod specification, must be unique within a Pod. Container runtime usually uses different globally unique name (`container.name`).
+*/
+export const SEMATTRS_K8S_CONTAINER_NAME = TMP_K8S_CONTAINER_NAME;
+
+/**
+* Number of times the container was restarted. This attribute can be used to identify a particular container (running or stopped) within a container spec.
+*/
+export const SEMATTRS_K8S_CONTAINER_RESTART_COUNT = TMP_K8S_CONTAINER_RESTART_COUNT;
+
+/**
+* The name of the CronJob.
+*/
+export const SEMATTRS_K8S_CRONJOB_NAME = TMP_K8S_CRONJOB_NAME;
+
+/**
+* The UID of the CronJob.
+*/
+export const SEMATTRS_K8S_CRONJOB_UID = TMP_K8S_CRONJOB_UID;
+
+/**
+* The name of the DaemonSet.
+*/
+export const SEMATTRS_K8S_DAEMONSET_NAME = TMP_K8S_DAEMONSET_NAME;
+
+/**
+* The UID of the DaemonSet.
+*/
+export const SEMATTRS_K8S_DAEMONSET_UID = TMP_K8S_DAEMONSET_UID;
+
+/**
+* The name of the Deployment.
+*/
+export const SEMATTRS_K8S_DEPLOYMENT_NAME = TMP_K8S_DEPLOYMENT_NAME;
+
+/**
+* The UID of the Deployment.
+*/
+export const SEMATTRS_K8S_DEPLOYMENT_UID = TMP_K8S_DEPLOYMENT_UID;
+
+/**
+* The name of the Job.
+*/
+export const SEMATTRS_K8S_JOB_NAME = TMP_K8S_JOB_NAME;
+
+/**
+* The UID of the Job.
+*/
+export const SEMATTRS_K8S_JOB_UID = TMP_K8S_JOB_UID;
+
+/**
+* The name of the namespace that the pod is running in.
+*/
+export const SEMATTRS_K8S_NAMESPACE_NAME = TMP_K8S_NAMESPACE_NAME;
+
+/**
+* The name of the Node.
+*/
+export const SEMATTRS_K8S_NODE_NAME = TMP_K8S_NODE_NAME;
+
+/**
+* The UID of the Node.
+*/
+export const SEMATTRS_K8S_NODE_UID = TMP_K8S_NODE_UID;
+
+/**
+* The name of the Pod.
+*/
+export const SEMATTRS_K8S_POD_NAME = TMP_K8S_POD_NAME;
+
+/**
+* The UID of the Pod.
+*/
+export const SEMATTRS_K8S_POD_UID = TMP_K8S_POD_UID;
+
+/**
+* The name of the ReplicaSet.
+*/
+export const SEMATTRS_K8S_REPLICASET_NAME = TMP_K8S_REPLICASET_NAME;
+
+/**
+* The UID of the ReplicaSet.
+*/
+export const SEMATTRS_K8S_REPLICASET_UID = TMP_K8S_REPLICASET_UID;
+
+/**
+* The name of the StatefulSet.
+*/
+export const SEMATTRS_K8S_STATEFULSET_NAME = TMP_K8S_STATEFULSET_NAME;
+
+/**
+* The UID of the StatefulSet.
+*/
+export const SEMATTRS_K8S_STATEFULSET_UID = TMP_K8S_STATEFULSET_UID;
 
 /**
 * The number of messages sent, received, or processed in the scope of the batching operation.
@@ -1034,6 +1899,11 @@ the broker doesn&#39;t have such notion, the destination name SHOULD uniquely id
 export const SEMATTRS_MESSAGING_DESTINATION_NAME = TMP_MESSAGING_DESTINATION_NAME;
 
 /**
+* The identifier of the partition messages are sent to or received from, unique within the `messaging.destination.name`.
+*/
+export const SEMATTRS_MESSAGING_DESTINATION_PARTITION_ID = TMP_MESSAGING_DESTINATION_PARTITION_ID;
+
+/**
 * Low cardinality representation of the messaging destination name.
 *
 * Note: Destination names could be constructed from templates. An example would be a destination name involving a user name or product id. Although the destination name in this case is of high cardinality, the underlying template is of low cardinality and can be effectively used for grouping and aggregation.
@@ -1059,6 +1929,16 @@ the broker doesn&#39;t have such notion, the original destination name SHOULD un
 export const SEMATTRS_MESSAGING_DESTINATION_PUBLISH_NAME = TMP_MESSAGING_DESTINATION_PUBLISH_NAME;
 
 /**
+* The name of the consumer group the event consumer is associated with.
+*/
+export const SEMATTRS_MESSAGING_EVENTHUBS_CONSUMER_GROUP = TMP_MESSAGING_EVENTHUBS_CONSUMER_GROUP;
+
+/**
+* The UTC epoch seconds at which the message has been accepted and stored in the entity.
+*/
+export const SEMATTRS_MESSAGING_EVENTHUBS_MESSAGE_ENQUEUED_TIME = TMP_MESSAGING_EVENTHUBS_MESSAGE_ENQUEUED_TIME;
+
+/**
 * The ordering key for a given message. If the attribute is not present, the message does not have an ordering key.
 */
 export const SEMATTRS_MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY = TMP_MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY;
@@ -1067,11 +1947,6 @@ export const SEMATTRS_MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY = TMP_MESSAGING_
 * Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.
 */
 export const SEMATTRS_MESSAGING_KAFKA_CONSUMER_GROUP = TMP_MESSAGING_KAFKA_CONSUMER_GROUP;
-
-/**
-* Partition the message is sent to.
-*/
-export const SEMATTRS_MESSAGING_KAFKA_DESTINATION_PARTITION = TMP_MESSAGING_KAFKA_DESTINATION_PARTITION;
 
 /**
 * Message keys in Kafka are used for grouping alike messages to ensure they&#39;re processed on the same partition. They differ from `messaging.message.id` in that they&#39;re not unique. If the key is `null`, the attribute MUST NOT be set.
@@ -1129,6 +2004,11 @@ export const SEMATTRS_MESSAGING_OPERATION = TMP_MESSAGING_OPERATION;
 export const SEMATTRS_MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY = TMP_MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY;
 
 /**
+* RabbitMQ message delivery tag.
+*/
+export const SEMATTRS_MESSAGING_RABBITMQ_MESSAGE_DELIVERY_TAG = TMP_MESSAGING_RABBITMQ_MESSAGE_DELIVERY_TAG;
+
+/**
 * Name of the RocketMQ producer/consumer group that is handling the message. The client type is identified by the SpanKind.
 */
 export const SEMATTRS_MESSAGING_ROCKETMQ_CLIENT_GROUP = TMP_MESSAGING_ROCKETMQ_CLIENT_GROUP;
@@ -1172,6 +2052,26 @@ export const SEMATTRS_MESSAGING_ROCKETMQ_MESSAGE_TYPE = TMP_MESSAGING_ROCKETMQ_M
 * Namespace of RocketMQ resources, resources in different namespaces are individual.
 */
 export const SEMATTRS_MESSAGING_ROCKETMQ_NAMESPACE = TMP_MESSAGING_ROCKETMQ_NAMESPACE;
+
+/**
+* The name of the subscription in the topic messages are received from.
+*/
+export const SEMATTRS_MESSAGING_SERVICEBUS_DESTINATION_SUBSCRIPTION_NAME = TMP_MESSAGING_SERVICEBUS_DESTINATION_SUBSCRIPTION_NAME;
+
+/**
+* Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+*/
+export const SEMATTRS_MESSAGING_SERVICEBUS_DISPOSITION_STATUS = TMP_MESSAGING_SERVICEBUS_DISPOSITION_STATUS;
+
+/**
+* Number of deliveries that have been attempted for this message.
+*/
+export const SEMATTRS_MESSAGING_SERVICEBUS_MESSAGE_DELIVERY_COUNT = TMP_MESSAGING_SERVICEBUS_MESSAGE_DELIVERY_COUNT;
+
+/**
+* The UTC epoch seconds at which the message has been accepted and stored in the entity.
+*/
+export const SEMATTRS_MESSAGING_SERVICEBUS_MESSAGE_ENQUEUED_TIME = TMP_MESSAGING_SERVICEBUS_MESSAGE_ENQUEUED_TIME;
 
 /**
 * An identifier for the messaging system being used. See below for a list of well-known identifiers.
@@ -1241,9 +2141,9 @@ export const SEMATTRS_NETWORK_PEER_PORT = TMP_NETWORK_PEER_PORT;
 export const SEMATTRS_NETWORK_PROTOCOL_NAME = TMP_NETWORK_PROTOCOL_NAME;
 
 /**
-* Version of the protocol specified in `network.protocol.name`.
+* The actual version of the protocol used for network communication.
 *
-* Note: `network.protocol.version` refers to the version of the protocol used and might be different from the protocol client&#39;s version. If the HTTP client has a version of `0.27.2`, but sends HTTP version `1.1`, this attribute should be set to `1.1`.
+* Note: If protocol version is subject to negotiation (for example using [ALPN](https://www.rfc-editor.org/rfc/rfc7301.html)), this attribute SHOULD be set to the negotiated version. If the actual protocol version is not known, this attribute SHOULD NOT be set.
 */
 export const SEMATTRS_NETWORK_PROTOCOL_VERSION = TMP_NETWORK_PROTOCOL_VERSION;
 
@@ -1264,6 +2164,99 @@ export const SEMATTRS_NETWORK_TRANSPORT = TMP_NETWORK_TRANSPORT;
 * Note: The value SHOULD be normalized to lowercase.
 */
 export const SEMATTRS_NETWORK_TYPE = TMP_NETWORK_TYPE;
+
+/**
+* The digest of the OCI image manifest. For container images specifically is the digest by which the container image is known.
+*
+* Note: Follows [OCI Image Manifest Specification](https://github.com/opencontainers/image-spec/blob/main/manifest.md), and specifically the [Digest property](https://github.com/opencontainers/image-spec/blob/main/descriptor.md#digests).
+An example can be found in [Example Image Manifest](https://docs.docker.com/registry/spec/manifest-v2-2/#example-image-manifest).
+*/
+export const SEMATTRS_OCI_MANIFEST_DIGEST = TMP_OCI_MANIFEST_DIGEST;
+
+/**
+* Unique identifier for a particular build or compilation of the operating system.
+*/
+export const SEMATTRS_OS_BUILD_ID = TMP_OS_BUILD_ID;
+
+/**
+* Human readable (not intended to be parsed) OS version information, like e.g. reported by `ver` or `lsb_release -a` commands.
+*/
+export const SEMATTRS_OS_DESCRIPTION = TMP_OS_DESCRIPTION;
+
+/**
+* Human readable operating system name.
+*/
+export const SEMATTRS_OS_NAME = TMP_OS_NAME;
+
+/**
+* The operating system type.
+*/
+export const SEMATTRS_OS_TYPE = TMP_OS_TYPE;
+
+/**
+* The version string of the operating system as defined in [Version Attributes](/docs/resource/README.md#version-attributes).
+*/
+export const SEMATTRS_OS_VERSION = TMP_OS_VERSION;
+
+/**
+* The [`service.name`](/docs/resource/README.md#service) of the remote service. SHOULD be equal to the actual `service.name` resource attribute of the remote service if any.
+*/
+export const SEMATTRS_PEER_SERVICE = TMP_PEER_SERVICE;
+
+/**
+* The command used to launch the process (i.e. the command name). On Linux based systems, can be set to the zeroth string in `proc/[pid]/cmdline`. On Windows, can be set to the first parameter extracted from `GetCommandLineW`.
+*/
+export const SEMATTRS_PROCESS_COMMAND = TMP_PROCESS_COMMAND;
+
+/**
+* All the command arguments (including the command/executable itself) as received by the process. On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according to the list of null-delimited strings extracted from `proc/[pid]/cmdline`. For libc-based executables, this would be the full argv vector passed to `main`.
+*/
+export const SEMATTRS_PROCESS_COMMAND_ARGS = TMP_PROCESS_COMMAND_ARGS;
+
+/**
+* The full command used to launch the process as a single string representing the full command. On Windows, can be set to the result of `GetCommandLineW`. Do not set this if you have to assemble it just for monitoring; use `process.command_args` instead.
+*/
+export const SEMATTRS_PROCESS_COMMAND_LINE = TMP_PROCESS_COMMAND_LINE;
+
+/**
+* The name of the process executable. On Linux based systems, can be set to the `Name` in `proc/[pid]/status`. On Windows, can be set to the base name of `GetProcessImageFileNameW`.
+*/
+export const SEMATTRS_PROCESS_EXECUTABLE_NAME = TMP_PROCESS_EXECUTABLE_NAME;
+
+/**
+* The full path to the process executable. On Linux based systems, can be set to the target of `proc/[pid]/exe`. On Windows, can be set to the result of `GetProcessImageFileNameW`.
+*/
+export const SEMATTRS_PROCESS_EXECUTABLE_PATH = TMP_PROCESS_EXECUTABLE_PATH;
+
+/**
+* The username of the user that owns the process.
+*/
+export const SEMATTRS_PROCESS_OWNER = TMP_PROCESS_OWNER;
+
+/**
+* Parent Process identifier (PPID).
+*/
+export const SEMATTRS_PROCESS_PARENT_PID = TMP_PROCESS_PARENT_PID;
+
+/**
+* Process identifier (PID).
+*/
+export const SEMATTRS_PROCESS_PID = TMP_PROCESS_PID;
+
+/**
+* An additional description about the runtime of the process, for example a specific vendor customization of the runtime environment.
+*/
+export const SEMATTRS_PROCESS_RUNTIME_DESCRIPTION = TMP_PROCESS_RUNTIME_DESCRIPTION;
+
+/**
+* The name of the runtime of this process. For compiled native binaries, this SHOULD be the name of the compiler.
+*/
+export const SEMATTRS_PROCESS_RUNTIME_NAME = TMP_PROCESS_RUNTIME_NAME;
+
+/**
+* The version of the runtime of this process, as returned by the runtime without modification.
+*/
+export const SEMATTRS_PROCESS_RUNTIME_VERSION = TMP_PROCESS_RUNTIME_VERSION;
 
 /**
 * The [error codes](https://connect.build/docs/protocol/#error-codes) of the Connect request. Error codes are always string values.
@@ -1329,6 +2322,67 @@ export const SEMATTRS_SERVER_ADDRESS = TMP_SERVER_ADDRESS;
 export const SEMATTRS_SERVER_PORT = TMP_SERVER_PORT;
 
 /**
+* The string ID of the service instance.
+*
+* Note: MUST be unique for each instance of the same `service.namespace,service.name` pair (in other words
+`service.namespace,service.name,service.instance.id` triplet MUST be globally unique). The ID helps to
+distinguish instances of the same service that exist at the same time (e.g. instances of a horizontally scaled
+service).
+
+Implementations, such as SDKs, are recommended to generate a random Version 1 or Version 4 [RFC
+4122](https://www.ietf.org/rfc/rfc4122.txt) UUID, but are free to use an inherent unique ID as the source of
+this value if stability is desirable. In that case, the ID SHOULD be used as source of a UUID Version 5 and
+SHOULD use the following UUID as the namespace: `4d63009a-8d0f-11ee-aad7-4c796ed8e320`.
+
+UUIDs are typically recommended, as only an opaque value for the purposes of identifying a service instance is
+needed. Similar to what can be seen in the man page for the
+[`/etc/machine-id`](https://www.freedesktop.org/software/systemd/man/machine-id.html) file, the underlying
+data, such as pod name and namespace should be treated as confidential, being the user&#39;s choice to expose it
+or not via another resource attribute.
+
+For applications running behind an application server (like unicorn), we do not recommend using one identifier
+for all processes participating in the application. Instead, it&#39;s recommended each division (e.g. a worker
+thread in unicorn) to have its own instance.id.
+
+It&#39;s not recommended for a Collector to set `service.instance.id` if it can&#39;t unambiguously determine the
+service instance that is generating that telemetry. For instance, creating an UUID based on `pod.name` will
+likely be wrong, as the Collector might not know from which container within that pod the telemetry originated.
+However, Collectors can set the `service.instance.id` if they can unambiguously determine the service instance
+for that telemetry. This is typically the case for scraping receivers, as they know the target address and
+port.
+*/
+export const SEMATTRS_SERVICE_INSTANCE_ID = TMP_SERVICE_INSTANCE_ID;
+
+/**
+* Logical name of the service.
+*
+* Note: MUST be the same for all instances of horizontally scaled services. If the value was not specified, SDKs MUST fallback to `unknown_service:` concatenated with [`process.executable.name`](process.md#process), e.g. `unknown_service:bash`. If `process.executable.name` is not available, the value MUST be set to `unknown_service`.
+*/
+export const SEMATTRS_SERVICE_NAME = TMP_SERVICE_NAME;
+
+/**
+* A namespace for `service.name`.
+*
+* Note: A string value having a meaning that helps to distinguish a group of services, for example the team name that owns a group of services. `service.name` is expected to be unique within the same namespace. If `service.namespace` is not specified in the Resource then `service.name` is expected to be unique for all services that have no explicit namespace defined (so the empty/unspecified namespace is simply one more valid namespace). Zero-length namespace string is assumed equal to unspecified namespace.
+*/
+export const SEMATTRS_SERVICE_NAMESPACE = TMP_SERVICE_NAMESPACE;
+
+/**
+* The version string of the service API or implementation. The format is not defined by these conventions.
+*/
+export const SEMATTRS_SERVICE_VERSION = TMP_SERVICE_VERSION;
+
+/**
+* A unique id to identify a session.
+*/
+export const SEMATTRS_SESSION_ID = TMP_SESSION_ID;
+
+/**
+* The previous `session.id` for this user, when known.
+*/
+export const SEMATTRS_SESSION_PREVIOUS_ID = TMP_SESSION_PREVIOUS_ID;
+
+/**
 * Source address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
 *
 * Note: When observed from the destination side, and when communicating through an intermediary, `source.address` SHOULD represent the source address behind any intermediaries, for example proxies, if it&#39;s available.
@@ -1339,6 +2393,41 @@ export const SEMATTRS_SOURCE_ADDRESS = TMP_SOURCE_ADDRESS;
 * Source port number.
 */
 export const SEMATTRS_SOURCE_PORT = TMP_SOURCE_PORT;
+
+/**
+* The language of the telemetry SDK.
+*/
+export const SEMATTRS_TELEMETRY_SDK_LANGUAGE = TMP_TELEMETRY_SDK_LANGUAGE;
+
+/**
+* The name of the telemetry SDK as defined above.
+*
+* Note: The OpenTelemetry SDK MUST set the `telemetry.sdk.name` attribute to `opentelemetry`.
+If another SDK, like a fork or a vendor-provided implementation, is used, this SDK MUST set the
+`telemetry.sdk.name` attribute to the fully-qualified class or module name of this SDK&#39;s main entry point
+or another suitable identifier depending on the language.
+The identifier `opentelemetry` is reserved and MUST NOT be used in this case.
+All custom identifiers SHOULD be stable across different versions of an implementation.
+*/
+export const SEMATTRS_TELEMETRY_SDK_NAME = TMP_TELEMETRY_SDK_NAME;
+
+/**
+* The version string of the telemetry SDK.
+*/
+export const SEMATTRS_TELEMETRY_SDK_VERSION = TMP_TELEMETRY_SDK_VERSION;
+
+/**
+* The name of the auto instrumentation agent or distribution, if used.
+*
+* Note: Official auto instrumentation agents and distributions SHOULD set the `telemetry.distro.name` attribute to
+a string starting with `opentelemetry-`, e.g. `opentelemetry-java-instrumentation`.
+*/
+export const SEMATTRS_TELEMETRY_DISTRO_NAME = TMP_TELEMETRY_DISTRO_NAME;
+
+/**
+* The version string of the auto instrumentation agent or distribution, if used.
+*/
+export const SEMATTRS_TELEMETRY_DISTRO_VERSION = TMP_TELEMETRY_DISTRO_VERSION;
 
 /**
 * Current &#34;managed&#34; thread ID (as opposed to OS thread ID).
@@ -1498,6 +2587,20 @@ export const SEMATTRS_TLS_SERVER_NOT_BEFORE = TMP_TLS_SERVER_NOT_BEFORE;
 export const SEMATTRS_TLS_SERVER_SUBJECT = TMP_TLS_SERVER_SUBJECT;
 
 /**
+* Domain extracted from the `url.full`, such as &#34;opentelemetry.io&#34;.
+*
+* Note: In some cases a URL may refer to an IP and/or port directly, without a domain name. In this case, the IP address would go to the domain field. If the URL contains a [literal IPv6 address](https://www.rfc-editor.org/rfc/rfc2732#section-2) enclosed by `[` and `]`, the `[` and `]` characters should also be captured in the domain field.
+*/
+export const SEMATTRS_URL_DOMAIN = TMP_URL_DOMAIN;
+
+/**
+* The file extension extracted from the `url.full`, excluding the leading dot.
+*
+* Note: The file extension is only set if it exists, as not every url has a file extension. When the file name has multiple extensions `example.tar.gz`, only the last one should be captured `gz`, not `tar.gz`.
+*/
+export const SEMATTRS_URL_EXTENSION = TMP_URL_EXTENSION;
+
+/**
 * The [URI fragment](https://www.rfc-editor.org/rfc/rfc3986#section-3.5) component.
 */
 export const SEMATTRS_URL_FRAGMENT = TMP_URL_FRAGMENT;
@@ -1507,21 +2610,43 @@ export const SEMATTRS_URL_FRAGMENT = TMP_URL_FRAGMENT;
 *
 * Note: For network calls, URL usually has `scheme://host[:port][path][?query][#fragment]` format, where the fragment is not transmitted over HTTP, but if it is known, it SHOULD be included nevertheless.
 `url.full` MUST NOT contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case username and password SHOULD be redacted and attribute&#39;s value SHOULD be `https://REDACTED:REDACTED@www.example.com/`.
-`url.full` SHOULD capture the absolute URL when it is available (or can be reconstructed) and SHOULD NOT be validated or modified except for sanitizing purposes.
+`url.full` SHOULD capture the absolute URL when it is available (or can be reconstructed). Sensitive content provided in `url.full` SHOULD be scrubbed when instrumentations can identify it.
 */
 export const SEMATTRS_URL_FULL = TMP_URL_FULL;
 
 /**
+* Unmodified original URL as seen in the event source.
+*
+* Note: In network monitoring, the observed URL may be a full URL, whereas in access logs, the URL is often just represented as a path. This field is meant to represent the URL as it was observed, complete or not.
+`url.original` might contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case password and username SHOULD NOT be redacted and attribute&#39;s value SHOULD remain the same.
+*/
+export const SEMATTRS_URL_ORIGINAL = TMP_URL_ORIGINAL;
+
+/**
 * The [URI path](https://www.rfc-editor.org/rfc/rfc3986#section-3.3) component.
+*
+* Note: Sensitive content provided in `url.path` SHOULD be scrubbed when instrumentations can identify it.
 */
 export const SEMATTRS_URL_PATH = TMP_URL_PATH;
 
 /**
+* Port extracted from the `url.full`.
+*/
+export const SEMATTRS_URL_PORT = TMP_URL_PORT;
+
+/**
 * The [URI query](https://www.rfc-editor.org/rfc/rfc3986#section-3.4) component.
 *
-* Note: Sensitive content provided in query string SHOULD be scrubbed when instrumentations can identify it.
+* Note: Sensitive content provided in `url.query` SHOULD be scrubbed when instrumentations can identify it.
 */
 export const SEMATTRS_URL_QUERY = TMP_URL_QUERY;
+
+/**
+* The highest registered url domain, stripped of the subdomain.
+*
+* Note: This value can be determined precisely with the [public suffix list](http://publicsuffix.org). For example, the registered domain for `foo.example.com` is `example.com`. Trying to approximate this by simply taking the last two labels will not work well for TLDs such as `co.uk`.
+*/
+export const SEMATTRS_URL_REGISTERED_DOMAIN = TMP_URL_REGISTERED_DOMAIN;
 
 /**
 * The [URI scheme](https://www.rfc-editor.org/rfc/rfc3986#section-3.1) component identifying the used protocol.
@@ -1529,19 +2654,37 @@ export const SEMATTRS_URL_QUERY = TMP_URL_QUERY;
 export const SEMATTRS_URL_SCHEME = TMP_URL_SCHEME;
 
 /**
+* The subdomain portion of a fully qualified domain name includes all of the names except the host name under the registered_domain. In a partially qualified domain, or if the qualification level of the full name cannot be determined, subdomain contains all of the names below the registered domain.
+*
+* Note: The subdomain portion of `www.east.mydomain.co.uk` is `east`. If the domain has multiple levels of subdomain, such as `sub2.sub1.example.com`, the subdomain field should contain `sub2.sub1`, with no trailing period.
+*/
+export const SEMATTRS_URL_SUBDOMAIN = TMP_URL_SUBDOMAIN;
+
+/**
+* The effective top level domain (eTLD), also known as the domain suffix, is the last part of the domain name. For example, the top level domain for example.com is `com`.
+*
+* Note: This value can be determined precisely with the [public suffix list](http://publicsuffix.org).
+*/
+export const SEMATTRS_URL_TOP_LEVEL_DOMAIN = TMP_URL_TOP_LEVEL_DOMAIN;
+
+/**
+* Name of the user-agent extracted from original. Usually refers to the browser&#39;s name.
+*
+* Note: [Example](https://www.whatsmyua.info) of extracting browser&#39;s name from original string. In the case of using a user-agent for non-browser products, such as microservices with multiple names/versions inside the `user_agent.original`, the most significant name SHOULD be selected. In such a scenario it should align with `user_agent.version`.
+*/
+export const SEMATTRS_USER_AGENT_NAME = TMP_USER_AGENT_NAME;
+
+/**
 * Value of the [HTTP User-Agent](https://www.rfc-editor.org/rfc/rfc9110.html#field.user-agent) header sent by the client.
 */
 export const SEMATTRS_USER_AGENT_ORIGINAL = TMP_USER_AGENT_ORIGINAL;
 
 /**
-* A unique id to identify a session.
+* Version of the user-agent extracted from original. Usually refers to the browser&#39;s version.
+*
+* Note: [Example](https://www.whatsmyua.info) of extracting browser&#39;s version from original string. In the case of using a user-agent for non-browser products, such as microservices with multiple names/versions inside the `user_agent.original`, the most significant version SHOULD be selected. In such a scenario it should align with `user_agent.name`.
 */
-export const SEMATTRS_SESSION_ID = TMP_SESSION_ID;
-
-/**
-* The previous `session.id` for this user, when known.
-*/
-export const SEMATTRS_SESSION_PREVIOUS_ID = TMP_SESSION_PREVIOUS_ID;
+export const SEMATTRS_USER_AGENT_VERSION = TMP_USER_AGENT_VERSION;
 
 /**
 * The full invoked ARN as provided on the `Context` passed to the function (`Lambda-Runtime-Invoked-Function-Arn` header on the `/runtime/invocation/next` applicable).
@@ -1549,31 +2692,6 @@ export const SEMATTRS_SESSION_PREVIOUS_ID = TMP_SESSION_PREVIOUS_ID;
 * Note: This may be different from `cloud.resource_id` if an alias is involved.
 */
 export const SEMATTRS_AWS_LAMBDA_INVOKED_ARN = TMP_AWS_LAMBDA_INVOKED_ARN;
-
-/**
-* The [event_id](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id) uniquely identifies the event.
-*/
-export const SEMATTRS_CLOUDEVENTS_EVENT_ID = TMP_CLOUDEVENTS_EVENT_ID;
-
-/**
-* The [source](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1) identifies the context in which an event happened.
-*/
-export const SEMATTRS_CLOUDEVENTS_EVENT_SOURCE = TMP_CLOUDEVENTS_EVENT_SOURCE;
-
-/**
-* The [version of the CloudEvents specification](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion) which the event uses.
-*/
-export const SEMATTRS_CLOUDEVENTS_EVENT_SPEC_VERSION = TMP_CLOUDEVENTS_EVENT_SPEC_VERSION;
-
-/**
-* The [subject](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject) of the event in the context of the event producer (identified by source).
-*/
-export const SEMATTRS_CLOUDEVENTS_EVENT_SUBJECT = TMP_CLOUDEVENTS_EVENT_SUBJECT;
-
-/**
-* The [event_type](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type) contains a value describing the type of event related to the originating occurrence.
-*/
-export const SEMATTRS_CLOUDEVENTS_EVENT_TYPE = TMP_CLOUDEVENTS_EVENT_TYPE;
 
 /**
 * Parent-child Reference type.
@@ -1593,183 +2711,9 @@ export const SEMATTRS_OTEL_STATUS_CODE = TMP_OTEL_STATUS_CODE;
 export const SEMATTRS_OTEL_STATUS_DESCRIPTION = TMP_OTEL_STATUS_DESCRIPTION;
 
 /**
-* The invocation ID of the current function invocation.
-*/
-export const SEMATTRS_FAAS_INVOCATION_ID = TMP_FAAS_INVOCATION_ID;
-
-/**
-* The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
-*/
-export const SEMATTRS_FAAS_DOCUMENT_COLLECTION = TMP_FAAS_DOCUMENT_COLLECTION;
-
-/**
-* The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.
-*/
-export const SEMATTRS_FAAS_DOCUMENT_NAME = TMP_FAAS_DOCUMENT_NAME;
-
-/**
-* Describes the type of the operation that was performed on the data.
-*/
-export const SEMATTRS_FAAS_DOCUMENT_OPERATION = TMP_FAAS_DOCUMENT_OPERATION;
-
-/**
-* A string containing the time when the data was accessed in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
-*/
-export const SEMATTRS_FAAS_DOCUMENT_TIME = TMP_FAAS_DOCUMENT_TIME;
-
-/**
-* A string containing the schedule period as [Cron Expression](https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm).
-*/
-export const SEMATTRS_FAAS_CRON = TMP_FAAS_CRON;
-
-/**
-* A string containing the function invocation time in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
-*/
-export const SEMATTRS_FAAS_TIME = TMP_FAAS_TIME;
-
-/**
-* A boolean that is true if the serverless function is executed for the first time (aka cold-start).
-*/
-export const SEMATTRS_FAAS_COLDSTART = TMP_FAAS_COLDSTART;
-
-/**
-* The unique identifier of the feature flag.
-*/
-export const SEMATTRS_FEATURE_FLAG_KEY = TMP_FEATURE_FLAG_KEY;
-
-/**
-* The name of the service provider that performs the flag evaluation.
-*/
-export const SEMATTRS_FEATURE_FLAG_PROVIDER_NAME = TMP_FEATURE_FLAG_PROVIDER_NAME;
-
-/**
-* SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.
-*
-* Note: A semantic identifier, commonly referred to as a variant, provides a means
-for referring to a value without including the value itself. This can
-provide additional context for understanding the meaning behind a value.
-For example, the variant `red` maybe be used for the value `#c05543`.
-
-A stringified version of the value can be used in situations where a
-semantic identifier is unavailable. String representation of the value
-should be determined by the implementer.
-*/
-export const SEMATTRS_FEATURE_FLAG_VARIANT = TMP_FEATURE_FLAG_VARIANT;
-
-/**
 * The AWS request ID as returned in the response headers `x-amz-request-id` or `x-amz-requestid`.
 */
 export const SEMATTRS_AWS_REQUEST_ID = TMP_AWS_REQUEST_ID;
-
-/**
-* The value of the `AttributesToGet` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_ATTRIBUTES_TO_GET = TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET;
-
-/**
-* The value of the `ConsistentRead` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_CONSISTENT_READ = TMP_AWS_DYNAMODB_CONSISTENT_READ;
-
-/**
-* The JSON-serialized value of each item in the `ConsumedCapacity` response field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_CONSUMED_CAPACITY = TMP_AWS_DYNAMODB_CONSUMED_CAPACITY;
-
-/**
-* The value of the `IndexName` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_INDEX_NAME = TMP_AWS_DYNAMODB_INDEX_NAME;
-
-/**
-* The JSON-serialized value of the `ItemCollectionMetrics` response field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_ITEM_COLLECTION_METRICS = TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS;
-
-/**
-* The value of the `Limit` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_LIMIT = TMP_AWS_DYNAMODB_LIMIT;
-
-/**
-* The value of the `ProjectionExpression` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_PROJECTION = TMP_AWS_DYNAMODB_PROJECTION;
-
-/**
-* The value of the `ProvisionedThroughput.ReadCapacityUnits` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY = TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY;
-
-/**
-* The value of the `ProvisionedThroughput.WriteCapacityUnits` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY = TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY;
-
-/**
-* The value of the `Select` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_SELECT = TMP_AWS_DYNAMODB_SELECT;
-
-/**
-* The keys in the `RequestItems` object field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_TABLE_NAMES = TMP_AWS_DYNAMODB_TABLE_NAMES;
-
-/**
-* The JSON-serialized value of each item of the `GlobalSecondaryIndexes` request field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES = TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES;
-
-/**
-* The JSON-serialized value of each item of the `LocalSecondaryIndexes` request field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES = TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES;
-
-/**
-* The value of the `ExclusiveStartTableName` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_EXCLUSIVE_START_TABLE = TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE;
-
-/**
-* The the number of items in the `TableNames` response parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_TABLE_COUNT = TMP_AWS_DYNAMODB_TABLE_COUNT;
-
-/**
-* The value of the `ScanIndexForward` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_SCAN_FORWARD = TMP_AWS_DYNAMODB_SCAN_FORWARD;
-
-/**
-* The value of the `Count` response parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_COUNT = TMP_AWS_DYNAMODB_COUNT;
-
-/**
-* The value of the `ScannedCount` response parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_SCANNED_COUNT = TMP_AWS_DYNAMODB_SCANNED_COUNT;
-
-/**
-* The value of the `Segment` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_SEGMENT = TMP_AWS_DYNAMODB_SEGMENT;
-
-/**
-* The value of the `TotalSegments` request parameter.
-*/
-export const SEMATTRS_AWS_DYNAMODB_TOTAL_SEGMENTS = TMP_AWS_DYNAMODB_TOTAL_SEGMENTS;
-
-/**
-* The JSON-serialized value of each item in the `AttributeDefinitions` request field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS = TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS;
-
-/**
-* The JSON-serialized value of each item in the the `GlobalSecondaryIndexUpdates` request field.
-*/
-export const SEMATTRS_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES = TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES;
 
 /**
 * The S3 bucket name the request refers to. Corresponds to the `--bucket` parameter of the [S3 API](https://docs.aws.amazon.com/cli/latest/reference/s3api/index.html) operations.
@@ -1896,55 +2840,9 @@ export const SEMATTRS_MESSAGE_UNCOMPRESSED_SIZE = TMP_MESSAGE_UNCOMPRESSED_SIZE;
 export type SemanticAttributes = {
 
   /**
-  * The name of the invoked function.
-  *
-  * Note: SHOULD be equal to the `faas.name` resource attribute of the invoked function.
-  */
-  FAAS_INVOKED_NAME: 'faas.invoked_name',
-
-  /**
-  * The cloud provider of the invoked function.
-  *
-  * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
-  */
-  FAAS_INVOKED_PROVIDER: 'faas.invoked_provider',
-
-  /**
-  * The cloud region of the invoked function.
-  *
-  * Note: SHOULD be equal to the `cloud.region` resource attribute of the invoked function.
-  */
-  FAAS_INVOKED_REGION: 'faas.invoked_region',
-
-  /**
-  * Type of the trigger which caused this function invocation.
-  */
-  FAAS_TRIGGER: 'faas.trigger',
-
-  /**
-  * The [`service.name`](/docs/resource/README.md#service) of the remote service. SHOULD be equal to the actual `service.name` resource attribute of the remote service if any.
-  */
-  PEER_SERVICE: 'peer.service',
-
-  /**
-  * Username or client_id extracted from the access token or [Authorization](https://tools.ietf.org/html/rfc7235#section-4.2) header in the inbound request from outside the system.
-  */
-  ENDUSER_ID: 'enduser.id',
-
-  /**
-  * Actual/assumed role the client is making the request under extracted from token or application security context.
-  */
-  ENDUSER_ROLE: 'enduser.role',
-
-  /**
-  * Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an [OAuth 2.0 Access Token](https://tools.ietf.org/html/rfc6749#section-3.3) or an attribute value in a [SAML 2.0 Assertion](http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html).
-  */
-  ENDUSER_SCOPE: 'enduser.scope',
-
-  /**
   * Identifies the class / type of event.
   *
-  * Note: Event names are subject to the same rules as [attribute names](https://github.com/open-telemetry/opentelemetry-specification/tree/v1.26.0/specification/common/attribute-naming.md). Notably, event names are namespaced to avoid collisions and provide a clean separation of semantics for events in separate domains like browser, mobile, and kubernetes.
+  * Note: Event names are subject to the same rules as [attribute names](https://github.com/open-telemetry/opentelemetry-specification/tree/v1.31.0/specification/common/attribute-naming.md). Notably, event names are namespaced to avoid collisions and provide a clean separation of semantics for events in separate domains like browser, mobile, and kubernetes.
   */
   EVENT_NAME: 'event.name',
 
@@ -1996,7 +2894,7 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   ANDROID_STATE: 'android.state',
 
   /**
-  * The name of the connection pool; unique within the instrumented application. In case the connection pool implementation doesn&#39;t provide a name, then the [db.connection_string](/docs/database/database-spans.md#connection-level-attributes) should be used.
+  * The name of the connection pool; unique within the instrumented application. In case the connection pool implementation doesn&#39;t provide a name, instrumentation should use a combination of `server.address` and `server.port` attributes formatted as `server.address:server.port`.
   */
   POOL_NAME: 'pool.name',
 
@@ -2004,6 +2902,11 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * The state of a connection in the pool.
   */
   STATE: 'state',
+
+  /**
+  * Rate-limiting result, shows whether the lease was acquired or contains a rejection reason.
+  */
+  ASPNETCORE_RATE_LIMITING_RESULT: 'aspnetcore.rate_limiting.result',
 
   /**
   * Full type name of the [`IExceptionHandler`](https://learn.microsoft.com/dotnet/api/microsoft.aspnetcore.diagnostics.iexceptionhandler) implementation that handled the exception.
@@ -2014,11 +2917,6 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * Rate limiting policy name.
   */
   ASPNETCORE_RATE_LIMITING_POLICY: 'aspnetcore.rate_limiting.policy',
-
-  /**
-  * Rate-limiting result, shows whether the lease was acquired or contains a rejection reason.
-  */
-  ASPNETCORE_RATE_LIMITING_RESULT: 'aspnetcore.rate_limiting.result',
 
   /**
   * Flag indicating if request was handled by the application pipeline.
@@ -2060,6 +2958,11 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   JVM_MEMORY_TYPE: 'jvm.memory.type',
 
   /**
+  * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+  */
+  PROCESS_CPU_STATE: 'process.cpu.state',
+
+  /**
   * The device identifier.
   */
   SYSTEM_DEVICE: 'system.device',
@@ -2070,7 +2973,7 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   SYSTEM_CPU_LOGICAL_NUMBER: 'system.cpu.logical_number',
 
   /**
-  * The state of the CPU.
+  * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
   */
   SYSTEM_CPU_STATE: 'system.cpu.state',
 
@@ -2122,7 +3025,151 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   /**
   * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
   */
-  SYSTEM_PROCESSES_STATUS: 'system.processes.status',
+  SYSTEM_PROCESS_STATUS: 'system.process.status',
+
+  /**
+  * Uniquely identifies the framework API revision offered by a version (`os.version`) of the android operating system. More information can be found [here](https://developer.android.com/guide/topics/manifest/uses-sdk-element#ApiLevels).
+  */
+  ANDROID_OS_API_LEVEL: 'android.os.api_level',
+
+  /**
+  * The JSON-serialized value of each item in the `AttributeDefinitions` request field.
+  */
+  AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS: 'aws.dynamodb.attribute_definitions',
+
+  /**
+  * The value of the `AttributesToGet` request parameter.
+  */
+  AWS_DYNAMODB_ATTRIBUTES_TO_GET: 'aws.dynamodb.attributes_to_get',
+
+  /**
+  * The value of the `ConsistentRead` request parameter.
+  */
+  AWS_DYNAMODB_CONSISTENT_READ: 'aws.dynamodb.consistent_read',
+
+  /**
+  * The JSON-serialized value of each item in the `ConsumedCapacity` response field.
+  */
+  AWS_DYNAMODB_CONSUMED_CAPACITY: 'aws.dynamodb.consumed_capacity',
+
+  /**
+  * The value of the `Count` response parameter.
+  */
+  AWS_DYNAMODB_COUNT: 'aws.dynamodb.count',
+
+  /**
+  * The value of the `ExclusiveStartTableName` request parameter.
+  */
+  AWS_DYNAMODB_EXCLUSIVE_START_TABLE: 'aws.dynamodb.exclusive_start_table',
+
+  /**
+  * The JSON-serialized value of each item in the `GlobalSecondaryIndexUpdates` request field.
+  */
+  AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES: 'aws.dynamodb.global_secondary_index_updates',
+
+  /**
+  * The JSON-serialized value of each item of the `GlobalSecondaryIndexes` request field.
+  */
+  AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES: 'aws.dynamodb.global_secondary_indexes',
+
+  /**
+  * The value of the `IndexName` request parameter.
+  */
+  AWS_DYNAMODB_INDEX_NAME: 'aws.dynamodb.index_name',
+
+  /**
+  * The JSON-serialized value of the `ItemCollectionMetrics` response field.
+  */
+  AWS_DYNAMODB_ITEM_COLLECTION_METRICS: 'aws.dynamodb.item_collection_metrics',
+
+  /**
+  * The value of the `Limit` request parameter.
+  */
+  AWS_DYNAMODB_LIMIT: 'aws.dynamodb.limit',
+
+  /**
+  * The JSON-serialized value of each item of the `LocalSecondaryIndexes` request field.
+  */
+  AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES: 'aws.dynamodb.local_secondary_indexes',
+
+  /**
+  * The value of the `ProjectionExpression` request parameter.
+  */
+  AWS_DYNAMODB_PROJECTION: 'aws.dynamodb.projection',
+
+  /**
+  * The value of the `ProvisionedThroughput.ReadCapacityUnits` request parameter.
+  */
+  AWS_DYNAMODB_PROVISIONED_READ_CAPACITY: 'aws.dynamodb.provisioned_read_capacity',
+
+  /**
+  * The value of the `ProvisionedThroughput.WriteCapacityUnits` request parameter.
+  */
+  AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY: 'aws.dynamodb.provisioned_write_capacity',
+
+  /**
+  * The value of the `ScanIndexForward` request parameter.
+  */
+  AWS_DYNAMODB_SCAN_FORWARD: 'aws.dynamodb.scan_forward',
+
+  /**
+  * The value of the `ScannedCount` response parameter.
+  */
+  AWS_DYNAMODB_SCANNED_COUNT: 'aws.dynamodb.scanned_count',
+
+  /**
+  * The value of the `Segment` request parameter.
+  */
+  AWS_DYNAMODB_SEGMENT: 'aws.dynamodb.segment',
+
+  /**
+  * The value of the `Select` request parameter.
+  */
+  AWS_DYNAMODB_SELECT: 'aws.dynamodb.select',
+
+  /**
+  * The number of items in the `TableNames` response parameter.
+  */
+  AWS_DYNAMODB_TABLE_COUNT: 'aws.dynamodb.table_count',
+
+  /**
+  * The keys in the `RequestItems` object field.
+  */
+  AWS_DYNAMODB_TABLE_NAMES: 'aws.dynamodb.table_names',
+
+  /**
+  * The value of the `TotalSegments` request parameter.
+  */
+  AWS_DYNAMODB_TOTAL_SEGMENTS: 'aws.dynamodb.total_segments',
+
+  /**
+  * Array of brand name and version separated by a space.
+  *
+  * Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.brands`).
+  */
+  BROWSER_BRANDS: 'browser.brands',
+
+  /**
+  * Preferred language of the user using the browser.
+  *
+  * Note: This value is intended to be taken from the Navigator API `navigator.language`.
+  */
+  BROWSER_LANGUAGE: 'browser.language',
+
+  /**
+  * A boolean that is true if the browser is running on a mobile device.
+  *
+  * Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.mobile`). If unavailable, this attribute SHOULD be left unset.
+  */
+  BROWSER_MOBILE: 'browser.mobile',
+
+  /**
+  * The platform on which the browser is running.
+  *
+  * Note: This value is intended to be taken from the [UA client hints API](https://wicg.github.io/ua-client-hints/#interface) (`navigator.userAgentData.platform`). If unavailable, the legacy `navigator.platform` API SHOULD NOT be used instead and this attribute SHOULD be left unset in order for the values to be consistent.
+The list of possible values is defined in the [W3C User-Agent Client Hints specification](https://wicg.github.io/ua-client-hints/#sec-ch-ua-platform). Note that some (but not all) of these values can overlap with values in the [`os.type` and `os.name` attributes](./os.md). However, for consistency, the values in the `browser.platform` attribute should capture the exact value that the user agent provides.
+  */
+  BROWSER_PLATFORM: 'browser.platform',
 
   /**
   * Client address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
@@ -2137,6 +3184,85 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * Note: When observed from the server side, and when communicating through an intermediary, `client.port` SHOULD represent the client port behind any intermediaries,  for example proxies, if it&#39;s available.
   */
   CLIENT_PORT: 'client.port',
+
+  /**
+  * The cloud account ID the resource is assigned to.
+  */
+  CLOUD_ACCOUNT_ID: 'cloud.account.id',
+
+  /**
+  * Cloud regions often have multiple, isolated locations known as zones to increase availability. Availability zone represents the zone where the resource is running.
+  *
+  * Note: Availability zones are called &#34;zones&#34; on Alibaba Cloud and Google Cloud.
+  */
+  CLOUD_AVAILABILITY_ZONE: 'cloud.availability_zone',
+
+  /**
+  * The cloud platform in use.
+  *
+  * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+  */
+  CLOUD_PLATFORM: 'cloud.platform',
+
+  /**
+  * Name of the cloud provider.
+  */
+  CLOUD_PROVIDER: 'cloud.provider',
+
+  /**
+  * The geographical region the resource is running.
+  *
+  * Note: Refer to your provider&#39;s docs to see the available regions, for example [Alibaba Cloud regions](https://www.alibabacloud.com/help/doc-detail/40654.htm), [AWS regions](https://aws.amazon.com/about-aws/global-infrastructure/regions_az/), [Azure regions](https://azure.microsoft.com/global-infrastructure/geographies/), [Google Cloud regions](https://cloud.google.com/about/locations), or [Tencent Cloud regions](https://www.tencentcloud.com/document/product/213/6091).
+  */
+  CLOUD_REGION: 'cloud.region',
+
+  /**
+  * Cloud provider-specific native identifier of the monitored cloud resource (e.g. an [ARN](https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html) on AWS, a [fully qualified resource ID](https://learn.microsoft.com/rest/api/resources/resources/get-by-id) on Azure, a [full resource name](https://cloud.google.com/apis/design/resource_names#full_resource_name) on GCP).
+  *
+  * Note: On some cloud providers, it may not be possible to determine the full ID at startup,
+so it may be necessary to set `cloud.resource_id` as a span attribute instead.
+
+The exact value to use for `cloud.resource_id` depends on the cloud provider.
+The following well-known definitions MUST be used if you set this attribute and they apply:
+
+* **AWS Lambda:** The function [ARN](https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html).
+  Take care not to use the &#34;invoked ARN&#34; directly but replace any
+  [alias suffix](https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html)
+  with the resolved function version, as the same runtime instance may be invokable with
+  multiple different aliases.
+* **GCP:** The [URI of the resource](https://cloud.google.com/iam/docs/full-resource-names)
+* **Azure:** The [Fully Qualified Resource ID](https://docs.microsoft.com/rest/api/resources/resources/get-by-id) of the invoked function,
+  *not* the function app, having the form
+  `/subscriptions/&lt;SUBSCIPTION_GUID&gt;/resourceGroups/&lt;RG&gt;/providers/Microsoft.Web/sites/&lt;FUNCAPP&gt;/functions/&lt;FUNC&gt;`.
+  This means that a span attribute MUST be used, as an Azure function app can host multiple functions that would usually share
+  a TracerProvider.
+  */
+  CLOUD_RESOURCE_ID: 'cloud.resource_id',
+
+  /**
+  * The [event_id](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id) uniquely identifies the event.
+  */
+  CLOUDEVENTS_EVENT_ID: 'cloudevents.event_id',
+
+  /**
+  * The [source](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1) identifies the context in which an event happened.
+  */
+  CLOUDEVENTS_EVENT_SOURCE: 'cloudevents.event_source',
+
+  /**
+  * The [version of the CloudEvents specification](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion) which the event uses.
+  */
+  CLOUDEVENTS_EVENT_SPEC_VERSION: 'cloudevents.event_spec_version',
+
+  /**
+  * The [subject](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject) of the event in the context of the event producer (identified by source).
+  */
+  CLOUDEVENTS_EVENT_SUBJECT: 'cloudevents.event_subject',
+
+  /**
+  * The [event_type](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type) contains a value describing the type of event related to the originating occurrence.
+  */
+  CLOUDEVENTS_EVENT_TYPE: 'cloudevents.event_type',
 
   /**
   * The column number in `code.filepath` best representing the operation. It SHOULD point within the code unit named in `code.function`.
@@ -2167,6 +3293,69 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.
   */
   CODE_STACKTRACE: 'code.stacktrace',
+
+  /**
+  * The command used to run the container (i.e. the command name).
+  *
+  * Note: If using embedded credentials or sensitive data, it is recommended to remove them to prevent potential leakage.
+  */
+  CONTAINER_COMMAND: 'container.command',
+
+  /**
+  * All the command arguments (including the command/executable itself) run by the container. [2].
+  */
+  CONTAINER_COMMAND_ARGS: 'container.command_args',
+
+  /**
+  * The full command run by the container as a single string representing the full command. [2].
+  */
+  CONTAINER_COMMAND_LINE: 'container.command_line',
+
+  /**
+  * The CPU state for this data point.
+  */
+  CONTAINER_CPU_STATE: 'container.cpu.state',
+
+  /**
+  * Container ID. Usually a UUID, as for example used to [identify Docker containers](https://docs.docker.com/engine/reference/run/#container-identification). The UUID might be abbreviated.
+  */
+  CONTAINER_ID: 'container.id',
+
+  /**
+  * Runtime specific image identifier. Usually a hash algorithm followed by a UUID.
+  *
+  * Note: Docker defines a sha256 of the image id; `container.image.id` corresponds to the `Image` field from the Docker container inspect [API](https://docs.docker.com/engine/api/v1.43/#tag/Container/operation/ContainerInspect) endpoint.
+K8s defines a link to the container registry repository with digest `&#34;imageID&#34;: &#34;registry.azurecr.io /namespace/service/dockerfile@sha256:bdeabd40c3a8a492eaf9e8e44d0ebbb84bac7ee25ac0cf8a7159d25f62555625&#34;`.
+The ID is assinged by the container runtime and can vary in different environments. Consider using `oci.manifest.digest` if it is important to identify the same image in different environments/runtimes.
+  */
+  CONTAINER_IMAGE_ID: 'container.image.id',
+
+  /**
+  * Name of the image the container was built on.
+  */
+  CONTAINER_IMAGE_NAME: 'container.image.name',
+
+  /**
+  * Repo digests of the container image as provided by the container runtime.
+  *
+  * Note: [Docker](https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect) and [CRI](https://github.com/kubernetes/cri-api/blob/c75ef5b473bbe2d0a4fc92f82235efd665ea8e9f/pkg/apis/runtime/v1/api.proto#L1237-L1238) report those under the `RepoDigests` field.
+  */
+  CONTAINER_IMAGE_REPO_DIGESTS: 'container.image.repo_digests',
+
+  /**
+  * Container image tags. An example can be found in [Docker Image Inspect](https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect). Should be only the `&lt;tag&gt;` section of the full name for example from `registry.example.com/my-org/my-image:&lt;tag&gt;`.
+  */
+  CONTAINER_IMAGE_TAGS: 'container.image.tags',
+
+  /**
+  * Container name used by container runtime.
+  */
+  CONTAINER_NAME: 'container.name',
+
+  /**
+  * The container runtime managing this container.
+  */
+  CONTAINER_RUNTIME: 'container.runtime',
 
   /**
   * The consistency level of the query. Based on consistency values from [CQL](https://docs.datastax.com/en/cassandra-oss/3.0/cassandra/dml/dmlConfigConsistency.html).
@@ -2204,11 +3393,6 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * Note: This mirrors the db.sql.table attribute but references cassandra rather than sql. It is not recommended to attempt any client-side parsing of `db.statement` just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.
   */
   DB_CASSANDRA_TABLE: 'db.cassandra.table',
-
-  /**
-  * The connection string used to connect to the database. It is recommended to remove embedded credentials.
-  */
-  DB_CONNECTION_STRING: 'db.connection_string',
 
   /**
   * Unique Cosmos client instance id.
@@ -2256,19 +3440,9 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   DB_ELASTICSEARCH_CLUSTER_NAME: 'db.elasticsearch.cluster.name',
 
   /**
-  * Represents the human-readable identifier of the node/instance to which a request was routed.
-  */
-  DB_ELASTICSEARCH_NODE_NAME: 'db.elasticsearch.node.name',
-
-  /**
   * An identifier (address, unique name, or any other identifier) of the database instance that is executing queries or mutations on the current connection. This is useful in cases where the database is running in a clustered environment and the instrumentation is able to record the node executing the query. The client may obtain this value in databases like MySQL using queries like `select @@hostname`.
   */
   DB_INSTANCE_ID: 'db.instance.id',
-
-  /**
-  * The fully-qualified class name of the [Java Database Connectivity (JDBC)](https://docs.oracle.com/javase/8/docs/technotes/guides/jdbc/) driver used to connect.
-  */
-  DB_JDBC_DRIVER_CLASSNAME: 'db.jdbc.driver_classname',
 
   /**
   * The MongoDB collection being accessed within the database stated in `db.name`.
@@ -2322,6 +3496,40 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * Username for accessing the database.
   */
   DB_USER: 'db.user',
+
+  /**
+  * Name of the [deployment environment](https://wikipedia.org/wiki/Deployment_environment) (aka deployment tier).
+  *
+  * Note: `deployment.environment` does not affect the uniqueness constraints defined through
+the `service.namespace`, `service.name` and `service.instance.id` resource attributes.
+This implies that resources carrying the following attribute combinations MUST be
+considered to be identifying the same service:
+
+* `service.name=frontend`, `deployment.environment=production`
+* `service.name=frontend`, `deployment.environment=staging`.
+  */
+  DEPLOYMENT_ENVIRONMENT: 'deployment.environment',
+
+  /**
+  * Deprecated, use `server.address`, `server.port` attributes instead.
+  *
+  * @deprecated "Replaced by `server.address` and `server.port`.".
+  */
+  DB_CONNECTION_STRING: 'db.connection_string',
+
+  /**
+  * Deprecated, use `db.instance.id` instead.
+  *
+  * @deprecated Replaced by `db.instance.id`.
+  */
+  DB_ELASTICSEARCH_NODE_NAME: 'db.elasticsearch.node.name',
+
+  /**
+  * Removed, no replacement at this time.
+  *
+  * @deprecated Removed as not used.
+  */
+  DB_JDBC_DRIVER_CLASSNAME: 'db.jdbc.driver_classname',
 
   /**
   * Deprecated, use `network.protocol.name` instead.
@@ -2385,6 +3593,13 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   * @deprecated Replaced by `user_agent.original`.
   */
   HTTP_USER_AGENT: 'http.user_agent',
+
+  /**
+  * &#34;Deprecated, use `messaging.destination.partition.id` instead.&#34;.
+  *
+  * @deprecated Replaced by `messaging.destination.partition.id`.
+  */
+  MESSAGING_KAFKA_DESTINATION_PARTITION: 'messaging.kafka.destination.partition',
 
   /**
   * Deprecated, use `server.address`.
@@ -2478,6 +3693,13 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   NET_TRANSPORT: 'net.transport',
 
   /**
+  * Deprecated, use `system.process.status` instead.
+  *
+  * @deprecated Replaced by `system.process.status`.
+  */
+  SYSTEM_PROCESSES_STATUS: 'system.processes.status',
+
+  /**
   * Destination address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
   *
   * Note: When observed from the source side, and when communicating through an intermediary, `destination.address` SHOULD represent the destination address behind any intermediaries, for example proxies, if it&#39;s available.
@@ -2490,9 +3712,59 @@ The id MAY be an [Universally Unique Lexicographically Sortable Identifier (ULID
   DESTINATION_PORT: 'destination.port',
 
   /**
+  * A unique identifier representing the device.
+  *
+  * Note: The device identifier MUST only be defined using the values outlined below. This value is not an advertising identifier and MUST NOT be used as such. On iOS (Swift or Objective-C), this value MUST be equal to the [vendor identifier](https://developer.apple.com/documentation/uikit/uidevice/1620059-identifierforvendor). On Android (Java or Kotlin), this value MUST be equal to the Firebase Installation ID or a globally unique UUID which is persisted across sessions in your application. More information can be found [here](https://developer.android.com/training/articles/user-data-ids) on best practices and exact implementation details. Caution should be taken when storing personal data or anything which can identify a user. GDPR and data protection laws may apply, ensure you do your own due diligence.
+  */
+  DEVICE_ID: 'device.id',
+
+  /**
+  * The name of the device manufacturer.
+  *
+  * Note: The Android OS provides this field via [Build](https://developer.android.com/reference/android/os/Build#MANUFACTURER). iOS apps SHOULD hardcode the value `Apple`.
+  */
+  DEVICE_MANUFACTURER: 'device.manufacturer',
+
+  /**
+  * The model identifier for the device.
+  *
+  * Note: It&#39;s recommended this value represents a machine-readable version of the model identifier rather than the market or consumer-friendly name of the device.
+  */
+  DEVICE_MODEL_IDENTIFIER: 'device.model.identifier',
+
+  /**
+  * The marketing name for the device model.
+  *
+  * Note: It&#39;s recommended this value represents a human-readable version of the device model rather than a machine-readable alternative.
+  */
+  DEVICE_MODEL_NAME: 'device.model.name',
+
+  /**
   * The disk IO operation direction.
   */
   DISK_IO_DIRECTION: 'disk.io.direction',
+
+  /**
+  * The name being queried.
+  *
+  * Note: If the name field contains non-printable characters (below 32 or above 126), those characters should be represented as escaped base 10 integers (\DDD). Back slashes and quotes should be escaped. Tabs, carriage returns, and line feeds should be converted to \t, \r, and \n respectively.
+  */
+  DNS_QUESTION_NAME: 'dns.question.name',
+
+  /**
+  * Username or client_id extracted from the access token or [Authorization](https://tools.ietf.org/html/rfc7235#section-4.2) header in the inbound request from outside the system.
+  */
+  ENDUSER_ID: 'enduser.id',
+
+  /**
+  * Actual/assumed role the client is making the request under extracted from token or application security context.
+  */
+  ENDUSER_ROLE: 'enduser.role',
+
+  /**
+  * Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an [OAuth 2.0 Access Token](https://tools.ietf.org/html/rfc6749#section-3.3) or an attribute value in a [SAML 2.0 Assertion](http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html).
+  */
+  ENDUSER_SCOPE: 'enduser.scope',
 
   /**
   * Describes a class of error the operation ended with.
@@ -2553,6 +3825,280 @@ clear whether the exception will escape.
   EXCEPTION_TYPE: 'exception.type',
 
   /**
+  * A boolean that is true if the serverless function is executed for the first time (aka cold-start).
+  */
+  FAAS_COLDSTART: 'faas.coldstart',
+
+  /**
+  * A string containing the schedule period as [Cron Expression](https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm).
+  */
+  FAAS_CRON: 'faas.cron',
+
+  /**
+  * The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
+  */
+  FAAS_DOCUMENT_COLLECTION: 'faas.document.collection',
+
+  /**
+  * The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.
+  */
+  FAAS_DOCUMENT_NAME: 'faas.document.name',
+
+  /**
+  * Describes the type of the operation that was performed on the data.
+  */
+  FAAS_DOCUMENT_OPERATION: 'faas.document.operation',
+
+  /**
+  * A string containing the time when the data was accessed in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
+  */
+  FAAS_DOCUMENT_TIME: 'faas.document.time',
+
+  /**
+  * The execution environment ID as a string, that will be potentially reused for other invocations to the same function/function version.
+  *
+  * Note: * **AWS Lambda:** Use the (full) log stream name.
+  */
+  FAAS_INSTANCE: 'faas.instance',
+
+  /**
+  * The invocation ID of the current function invocation.
+  */
+  FAAS_INVOCATION_ID: 'faas.invocation_id',
+
+  /**
+  * The name of the invoked function.
+  *
+  * Note: SHOULD be equal to the `faas.name` resource attribute of the invoked function.
+  */
+  FAAS_INVOKED_NAME: 'faas.invoked_name',
+
+  /**
+  * The cloud provider of the invoked function.
+  *
+  * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+  */
+  FAAS_INVOKED_PROVIDER: 'faas.invoked_provider',
+
+  /**
+  * The cloud region of the invoked function.
+  *
+  * Note: SHOULD be equal to the `cloud.region` resource attribute of the invoked function.
+  */
+  FAAS_INVOKED_REGION: 'faas.invoked_region',
+
+  /**
+  * The amount of memory available to the serverless function converted to Bytes.
+  *
+  * Note: It&#39;s recommended to set this attribute since e.g. too little memory can easily stop a Java AWS Lambda function from working correctly. On AWS Lambda, the environment variable `AWS_LAMBDA_FUNCTION_MEMORY_SIZE` provides this information (which must be multiplied by 1,048,576).
+  */
+  FAAS_MAX_MEMORY: 'faas.max_memory',
+
+  /**
+  * The name of the single function that this runtime instance executes.
+  *
+  * Note: This is the name of the function as configured/deployed on the FaaS
+platform and is usually different from the name of the callback
+function (which may be stored in the
+[`code.namespace`/`code.function`](/docs/general/attributes.md#source-code-attributes)
+span attributes).
+
+For some cloud providers, the above definition is ambiguous. The following
+definition of function name MUST be used for this attribute
+(and consequently the span name) for the listed cloud providers/products:
+
+* **Azure:**  The full name `&lt;FUNCAPP&gt;/&lt;FUNC&gt;`, i.e., function app name
+  followed by a forward slash followed by the function name (this form
+  can also be seen in the resource JSON for the function).
+  This means that a span attribute MUST be used, as an Azure function
+  app can host multiple functions that would usually share
+  a TracerProvider (see also the `cloud.resource_id` attribute).
+  */
+  FAAS_NAME: 'faas.name',
+
+  /**
+  * A string containing the function invocation time in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
+  */
+  FAAS_TIME: 'faas.time',
+
+  /**
+  * Type of the trigger which caused this function invocation.
+  */
+  FAAS_TRIGGER: 'faas.trigger',
+
+  /**
+  * The immutable version of the function being executed.
+  *
+  * Note: Depending on the cloud provider and platform, use:
+
+* **AWS Lambda:** The [function version](https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html)
+  (an integer represented as a decimal string).
+* **Google Cloud Run (Services):** The [revision](https://cloud.google.com/run/docs/managing/revisions)
+  (i.e., the function name plus the revision suffix).
+* **Google Cloud Functions:** The value of the
+  [`K_REVISION` environment variable](https://cloud.google.com/functions/docs/env-var#runtime_environment_variables_set_automatically).
+* **Azure Functions:** Not applicable. Do not set this attribute.
+  */
+  FAAS_VERSION: 'faas.version',
+
+  /**
+  * The unique identifier of the feature flag.
+  */
+  FEATURE_FLAG_KEY: 'feature_flag.key',
+
+  /**
+  * The name of the service provider that performs the flag evaluation.
+  */
+  FEATURE_FLAG_PROVIDER_NAME: 'feature_flag.provider_name',
+
+  /**
+  * SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.
+  *
+  * Note: A semantic identifier, commonly referred to as a variant, provides a means
+for referring to a value without including the value itself. This can
+provide additional context for understanding the meaning behind a value.
+For example, the variant `red` maybe be used for the value `#c05543`.
+
+A stringified version of the value can be used in situations where a
+semantic identifier is unavailable. String representation of the value
+should be determined by the implementer.
+  */
+  FEATURE_FLAG_VARIANT: 'feature_flag.variant',
+
+  /**
+  * Directory where the file is located. It should include the drive letter, when appropriate.
+  */
+  FILE_DIRECTORY: 'file.directory',
+
+  /**
+  * File extension, excluding the leading dot.
+  *
+  * Note: When the file name has multiple extensions (example.tar.gz), only the last one should be captured (&#34;gz&#34;, not &#34;tar.gz&#34;).
+  */
+  FILE_EXTENSION: 'file.extension',
+
+  /**
+  * Name of the file including the extension, without the directory.
+  */
+  FILE_NAME: 'file.name',
+
+  /**
+  * Full path to the file, including the file name. It should include the drive letter, when appropriate.
+  */
+  FILE_PATH: 'file.path',
+
+  /**
+  * File size in bytes.
+  */
+  FILE_SIZE: 'file.size',
+
+  /**
+  * The name of the Cloud Run [execution](https://cloud.google.com/run/docs/managing/job-executions) being run for the Job, as set by the [`CLOUD_RUN_EXECUTION`](https://cloud.google.com/run/docs/container-contract#jobs-env-vars) environment variable.
+  */
+  GCP_CLOUD_RUN_JOB_EXECUTION: 'gcp.cloud_run.job.execution',
+
+  /**
+  * The index for a task within an execution as provided by the [`CLOUD_RUN_TASK_INDEX`](https://cloud.google.com/run/docs/container-contract#jobs-env-vars) environment variable.
+  */
+  GCP_CLOUD_RUN_JOB_TASK_INDEX: 'gcp.cloud_run.job.task_index',
+
+  /**
+  * The hostname of a GCE instance. This is the full value of the default or [custom hostname](https://cloud.google.com/compute/docs/instances/custom-hostname-vm).
+  */
+  GCP_GCE_INSTANCE_HOSTNAME: 'gcp.gce.instance.hostname',
+
+  /**
+  * The instance name of a GCE instance. This is the value provided by `host.name`, the visible name of the instance in the Cloud Console UI, and the prefix for the default hostname of the instance as defined by the [default internal DNS name](https://cloud.google.com/compute/docs/internal-dns#instance-fully-qualified-domain-names).
+  */
+  GCP_GCE_INSTANCE_NAME: 'gcp.gce.instance.name',
+
+  /**
+  * The CPU architecture the host system is running on.
+  */
+  HOST_ARCH: 'host.arch',
+
+  /**
+  * The amount of level 2 memory cache available to the processor (in Bytes).
+  */
+  HOST_CPU_CACHE_L2_SIZE: 'host.cpu.cache.l2.size',
+
+  /**
+  * Family or generation of the CPU.
+  */
+  HOST_CPU_FAMILY: 'host.cpu.family',
+
+  /**
+  * Model identifier. It provides more granular information about the CPU, distinguishing it from other CPUs within the same family.
+  */
+  HOST_CPU_MODEL_ID: 'host.cpu.model.id',
+
+  /**
+  * Model designation of the processor.
+  */
+  HOST_CPU_MODEL_NAME: 'host.cpu.model.name',
+
+  /**
+  * Stepping or core revisions.
+  */
+  HOST_CPU_STEPPING: 'host.cpu.stepping',
+
+  /**
+  * Processor manufacturer identifier. A maximum 12-character string.
+  *
+  * Note: [CPUID](https://wiki.osdev.org/CPUID) command returns the vendor ID string in EBX, EDX and ECX registers. Writing these to memory in this order results in a 12-character string.
+  */
+  HOST_CPU_VENDOR_ID: 'host.cpu.vendor.id',
+
+  /**
+  * Unique host ID. For Cloud, this must be the instance_id assigned by the cloud provider. For non-containerized systems, this should be the `machine-id`. See the table below for the sources to use to determine the `machine-id` based on operating system.
+  */
+  HOST_ID: 'host.id',
+
+  /**
+  * VM image ID or host OS image ID. For Cloud, this value is from the provider.
+  */
+  HOST_IMAGE_ID: 'host.image.id',
+
+  /**
+  * Name of the VM image or OS install the host was instantiated from.
+  */
+  HOST_IMAGE_NAME: 'host.image.name',
+
+  /**
+  * The version string of the VM image or host OS as defined in [Version Attributes](/docs/resource/README.md#version-attributes).
+  */
+  HOST_IMAGE_VERSION: 'host.image.version',
+
+  /**
+  * Available IP addresses of the host, excluding loopback interfaces.
+  *
+  * Note: IPv4 Addresses MUST be specified in dotted-quad notation. IPv6 addresses MUST be specified in the [RFC 5952](https://www.rfc-editor.org/rfc/rfc5952.html) format.
+  */
+  HOST_IP: 'host.ip',
+
+  /**
+  * Available MAC addresses of the host, excluding loopback interfaces.
+  *
+  * Note: MAC Addresses MUST be represented in [IEEE RA hexadecimal form](https://standards.ieee.org/wp-content/uploads/import/documents/tutorials/eui.pdf): as hyphen-separated octets in uppercase hexadecimal form from most to least significant.
+  */
+  HOST_MAC: 'host.mac',
+
+  /**
+  * Name of the host. On Unix systems, it may contain what the hostname command returns, or the fully qualified hostname, or another name specified by the user.
+  */
+  HOST_NAME: 'host.name',
+
+  /**
+  * Type of host. For Cloud, this must be the machine type.
+  */
+  HOST_TYPE: 'host.type',
+
+  /**
+  * State of the HTTP connection in the HTTP connection pool.
+  */
+  HTTP_CONNECTION_STATE: 'http.connection.state',
+
+  /**
   * The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the [Content-Length](https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length) header. For requests using transport encoding, this should be the compressed size.
   */
   HTTP_REQUEST_BODY_SIZE: 'http.request.body.size',
@@ -2590,9 +4136,19 @@ Tracing instrumentations that do so, MUST also set `http.request.method_original
   HTTP_REQUEST_RESEND_COUNT: 'http.request.resend_count',
 
   /**
+  * The total size of the request in bytes. This should be the total number of bytes sent over the wire, including the request line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and request body if any.
+  */
+  HTTP_REQUEST_SIZE: 'http.request.size',
+
+  /**
   * The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the [Content-Length](https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length) header. For requests using transport encoding, this should be the compressed size.
   */
   HTTP_RESPONSE_BODY_SIZE: 'http.response.body.size',
+
+  /**
+  * The total size of the response in bytes. This should be the total number of bytes sent over the wire, including the status line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and response body and trailers if any.
+  */
+  HTTP_RESPONSE_SIZE: 'http.response.size',
 
   /**
   * [HTTP response status code](https://tools.ietf.org/html/rfc7231#section-6).
@@ -2606,6 +4162,134 @@ Tracing instrumentations that do so, MUST also set `http.request.method_original
 SHOULD include the [application root](/docs/http/http-spans.md#http-server-definitions) if there is one.
   */
   HTTP_ROUTE: 'http.route',
+
+  /**
+  * The name of the cluster.
+  */
+  K8S_CLUSTER_NAME: 'k8s.cluster.name',
+
+  /**
+  * A pseudo-ID for the cluster, set to the UID of the `kube-system` namespace.
+  *
+  * Note: K8s doesn&#39;t have support for obtaining a cluster ID. If this is ever
+added, we will recommend collecting the `k8s.cluster.uid` through the
+official APIs. In the meantime, we are able to use the `uid` of the
+`kube-system` namespace as a proxy for cluster ID. Read on for the
+rationale.
+
+Every object created in a K8s cluster is assigned a distinct UID. The
+`kube-system` namespace is used by Kubernetes itself and will exist
+for the lifetime of the cluster. Using the `uid` of the `kube-system`
+namespace is a reasonable proxy for the K8s ClusterID as it will only
+change if the cluster is rebuilt. Furthermore, Kubernetes UIDs are
+UUIDs as standardized by
+[ISO/IEC 9834-8 and ITU-T X.667](https://www.itu.int/ITU-T/studygroups/com17/oid.html).
+Which states:
+
+&gt; If generated according to one of the mechanisms defined in Rec.
+  ITU-T X.667 | ISO/IEC 9834-8, a UUID is either guaranteed to be
+  different from all other UUIDs generated before 3603 A.D., or is
+  extremely likely to be different (depending on the mechanism chosen).
+
+Therefore, UIDs between clusters should be extremely unlikely to
+conflict.
+  */
+  K8S_CLUSTER_UID: 'k8s.cluster.uid',
+
+  /**
+  * The name of the Container from Pod specification, must be unique within a Pod. Container runtime usually uses different globally unique name (`container.name`).
+  */
+  K8S_CONTAINER_NAME: 'k8s.container.name',
+
+  /**
+  * Number of times the container was restarted. This attribute can be used to identify a particular container (running or stopped) within a container spec.
+  */
+  K8S_CONTAINER_RESTART_COUNT: 'k8s.container.restart_count',
+
+  /**
+  * The name of the CronJob.
+  */
+  K8S_CRONJOB_NAME: 'k8s.cronjob.name',
+
+  /**
+  * The UID of the CronJob.
+  */
+  K8S_CRONJOB_UID: 'k8s.cronjob.uid',
+
+  /**
+  * The name of the DaemonSet.
+  */
+  K8S_DAEMONSET_NAME: 'k8s.daemonset.name',
+
+  /**
+  * The UID of the DaemonSet.
+  */
+  K8S_DAEMONSET_UID: 'k8s.daemonset.uid',
+
+  /**
+  * The name of the Deployment.
+  */
+  K8S_DEPLOYMENT_NAME: 'k8s.deployment.name',
+
+  /**
+  * The UID of the Deployment.
+  */
+  K8S_DEPLOYMENT_UID: 'k8s.deployment.uid',
+
+  /**
+  * The name of the Job.
+  */
+  K8S_JOB_NAME: 'k8s.job.name',
+
+  /**
+  * The UID of the Job.
+  */
+  K8S_JOB_UID: 'k8s.job.uid',
+
+  /**
+  * The name of the namespace that the pod is running in.
+  */
+  K8S_NAMESPACE_NAME: 'k8s.namespace.name',
+
+  /**
+  * The name of the Node.
+  */
+  K8S_NODE_NAME: 'k8s.node.name',
+
+  /**
+  * The UID of the Node.
+  */
+  K8S_NODE_UID: 'k8s.node.uid',
+
+  /**
+  * The name of the Pod.
+  */
+  K8S_POD_NAME: 'k8s.pod.name',
+
+  /**
+  * The UID of the Pod.
+  */
+  K8S_POD_UID: 'k8s.pod.uid',
+
+  /**
+  * The name of the ReplicaSet.
+  */
+  K8S_REPLICASET_NAME: 'k8s.replicaset.name',
+
+  /**
+  * The UID of the ReplicaSet.
+  */
+  K8S_REPLICASET_UID: 'k8s.replicaset.uid',
+
+  /**
+  * The name of the StatefulSet.
+  */
+  K8S_STATEFULSET_NAME: 'k8s.statefulset.name',
+
+  /**
+  * The UID of the StatefulSet.
+  */
+  K8S_STATEFULSET_UID: 'k8s.statefulset.uid',
 
   /**
   * The number of messages sent, received, or processed in the scope of the batching operation.
@@ -2633,6 +4317,11 @@ the broker doesn&#39;t have such notion, the destination name SHOULD uniquely id
   MESSAGING_DESTINATION_NAME: 'messaging.destination.name',
 
   /**
+  * The identifier of the partition messages are sent to or received from, unique within the `messaging.destination.name`.
+  */
+  MESSAGING_DESTINATION_PARTITION_ID: 'messaging.destination.partition.id',
+
+  /**
   * Low cardinality representation of the messaging destination name.
   *
   * Note: Destination names could be constructed from templates. An example would be a destination name involving a user name or product id. Although the destination name in this case is of high cardinality, the underlying template is of low cardinality and can be effectively used for grouping and aggregation.
@@ -2658,6 +4347,16 @@ the broker doesn&#39;t have such notion, the original destination name SHOULD un
   MESSAGING_DESTINATION_PUBLISH_NAME: 'messaging.destination_publish.name',
 
   /**
+  * The name of the consumer group the event consumer is associated with.
+  */
+  MESSAGING_EVENTHUBS_CONSUMER_GROUP: 'messaging.eventhubs.consumer.group',
+
+  /**
+  * The UTC epoch seconds at which the message has been accepted and stored in the entity.
+  */
+  MESSAGING_EVENTHUBS_MESSAGE_ENQUEUED_TIME: 'messaging.eventhubs.message.enqueued_time',
+
+  /**
   * The ordering key for a given message. If the attribute is not present, the message does not have an ordering key.
   */
   MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY: 'messaging.gcp_pubsub.message.ordering_key',
@@ -2666,11 +4365,6 @@ the broker doesn&#39;t have such notion, the original destination name SHOULD un
   * Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.
   */
   MESSAGING_KAFKA_CONSUMER_GROUP: 'messaging.kafka.consumer.group',
-
-  /**
-  * Partition the message is sent to.
-  */
-  MESSAGING_KAFKA_DESTINATION_PARTITION: 'messaging.kafka.destination.partition',
 
   /**
   * Message keys in Kafka are used for grouping alike messages to ensure they&#39;re processed on the same partition. They differ from `messaging.message.id` in that they&#39;re not unique. If the key is `null`, the attribute MUST NOT be set.
@@ -2728,6 +4422,11 @@ size should be used.
   MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY: 'messaging.rabbitmq.destination.routing_key',
 
   /**
+  * RabbitMQ message delivery tag.
+  */
+  MESSAGING_RABBITMQ_MESSAGE_DELIVERY_TAG: 'messaging.rabbitmq.message.delivery_tag',
+
+  /**
   * Name of the RocketMQ producer/consumer group that is handling the message. The client type is identified by the SpanKind.
   */
   MESSAGING_ROCKETMQ_CLIENT_GROUP: 'messaging.rocketmq.client_group',
@@ -2771,6 +4470,26 @@ size should be used.
   * Namespace of RocketMQ resources, resources in different namespaces are individual.
   */
   MESSAGING_ROCKETMQ_NAMESPACE: 'messaging.rocketmq.namespace',
+
+  /**
+  * The name of the subscription in the topic messages are received from.
+  */
+  MESSAGING_SERVICEBUS_DESTINATION_SUBSCRIPTION_NAME: 'messaging.servicebus.destination.subscription_name',
+
+  /**
+  * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+  */
+  MESSAGING_SERVICEBUS_DISPOSITION_STATUS: 'messaging.servicebus.disposition_status',
+
+  /**
+  * Number of deliveries that have been attempted for this message.
+  */
+  MESSAGING_SERVICEBUS_MESSAGE_DELIVERY_COUNT: 'messaging.servicebus.message.delivery_count',
+
+  /**
+  * The UTC epoch seconds at which the message has been accepted and stored in the entity.
+  */
+  MESSAGING_SERVICEBUS_MESSAGE_ENQUEUED_TIME: 'messaging.servicebus.message.enqueued_time',
 
   /**
   * An identifier for the messaging system being used. See below for a list of well-known identifiers.
@@ -2840,9 +4559,9 @@ size should be used.
   NETWORK_PROTOCOL_NAME: 'network.protocol.name',
 
   /**
-  * Version of the protocol specified in `network.protocol.name`.
+  * The actual version of the protocol used for network communication.
   *
-  * Note: `network.protocol.version` refers to the version of the protocol used and might be different from the protocol client&#39;s version. If the HTTP client has a version of `0.27.2`, but sends HTTP version `1.1`, this attribute should be set to `1.1`.
+  * Note: If protocol version is subject to negotiation (for example using [ALPN](https://www.rfc-editor.org/rfc/rfc7301.html)), this attribute SHOULD be set to the negotiated version. If the actual protocol version is not known, this attribute SHOULD NOT be set.
   */
   NETWORK_PROTOCOL_VERSION: 'network.protocol.version',
 
@@ -2863,6 +4582,99 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   * Note: The value SHOULD be normalized to lowercase.
   */
   NETWORK_TYPE: 'network.type',
+
+  /**
+  * The digest of the OCI image manifest. For container images specifically is the digest by which the container image is known.
+  *
+  * Note: Follows [OCI Image Manifest Specification](https://github.com/opencontainers/image-spec/blob/main/manifest.md), and specifically the [Digest property](https://github.com/opencontainers/image-spec/blob/main/descriptor.md#digests).
+An example can be found in [Example Image Manifest](https://docs.docker.com/registry/spec/manifest-v2-2/#example-image-manifest).
+  */
+  OCI_MANIFEST_DIGEST: 'oci.manifest.digest',
+
+  /**
+  * Unique identifier for a particular build or compilation of the operating system.
+  */
+  OS_BUILD_ID: 'os.build_id',
+
+  /**
+  * Human readable (not intended to be parsed) OS version information, like e.g. reported by `ver` or `lsb_release -a` commands.
+  */
+  OS_DESCRIPTION: 'os.description',
+
+  /**
+  * Human readable operating system name.
+  */
+  OS_NAME: 'os.name',
+
+  /**
+  * The operating system type.
+  */
+  OS_TYPE: 'os.type',
+
+  /**
+  * The version string of the operating system as defined in [Version Attributes](/docs/resource/README.md#version-attributes).
+  */
+  OS_VERSION: 'os.version',
+
+  /**
+  * The [`service.name`](/docs/resource/README.md#service) of the remote service. SHOULD be equal to the actual `service.name` resource attribute of the remote service if any.
+  */
+  PEER_SERVICE: 'peer.service',
+
+  /**
+  * The command used to launch the process (i.e. the command name). On Linux based systems, can be set to the zeroth string in `proc/[pid]/cmdline`. On Windows, can be set to the first parameter extracted from `GetCommandLineW`.
+  */
+  PROCESS_COMMAND: 'process.command',
+
+  /**
+  * All the command arguments (including the command/executable itself) as received by the process. On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according to the list of null-delimited strings extracted from `proc/[pid]/cmdline`. For libc-based executables, this would be the full argv vector passed to `main`.
+  */
+  PROCESS_COMMAND_ARGS: 'process.command_args',
+
+  /**
+  * The full command used to launch the process as a single string representing the full command. On Windows, can be set to the result of `GetCommandLineW`. Do not set this if you have to assemble it just for monitoring; use `process.command_args` instead.
+  */
+  PROCESS_COMMAND_LINE: 'process.command_line',
+
+  /**
+  * The name of the process executable. On Linux based systems, can be set to the `Name` in `proc/[pid]/status`. On Windows, can be set to the base name of `GetProcessImageFileNameW`.
+  */
+  PROCESS_EXECUTABLE_NAME: 'process.executable.name',
+
+  /**
+  * The full path to the process executable. On Linux based systems, can be set to the target of `proc/[pid]/exe`. On Windows, can be set to the result of `GetProcessImageFileNameW`.
+  */
+  PROCESS_EXECUTABLE_PATH: 'process.executable.path',
+
+  /**
+  * The username of the user that owns the process.
+  */
+  PROCESS_OWNER: 'process.owner',
+
+  /**
+  * Parent Process identifier (PPID).
+  */
+  PROCESS_PARENT_PID: 'process.parent_pid',
+
+  /**
+  * Process identifier (PID).
+  */
+  PROCESS_PID: 'process.pid',
+
+  /**
+  * An additional description about the runtime of the process, for example a specific vendor customization of the runtime environment.
+  */
+  PROCESS_RUNTIME_DESCRIPTION: 'process.runtime.description',
+
+  /**
+  * The name of the runtime of this process. For compiled native binaries, this SHOULD be the name of the compiler.
+  */
+  PROCESS_RUNTIME_NAME: 'process.runtime.name',
+
+  /**
+  * The version of the runtime of this process, as returned by the runtime without modification.
+  */
+  PROCESS_RUNTIME_VERSION: 'process.runtime.version',
 
   /**
   * The [error codes](https://connect.build/docs/protocol/#error-codes) of the Connect request. Error codes are always string values.
@@ -2928,6 +4740,67 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   SERVER_PORT: 'server.port',
 
   /**
+  * The string ID of the service instance.
+  *
+  * Note: MUST be unique for each instance of the same `service.namespace,service.name` pair (in other words
+`service.namespace,service.name,service.instance.id` triplet MUST be globally unique). The ID helps to
+distinguish instances of the same service that exist at the same time (e.g. instances of a horizontally scaled
+service).
+
+Implementations, such as SDKs, are recommended to generate a random Version 1 or Version 4 [RFC
+4122](https://www.ietf.org/rfc/rfc4122.txt) UUID, but are free to use an inherent unique ID as the source of
+this value if stability is desirable. In that case, the ID SHOULD be used as source of a UUID Version 5 and
+SHOULD use the following UUID as the namespace: `4d63009a-8d0f-11ee-aad7-4c796ed8e320`.
+
+UUIDs are typically recommended, as only an opaque value for the purposes of identifying a service instance is
+needed. Similar to what can be seen in the man page for the
+[`/etc/machine-id`](https://www.freedesktop.org/software/systemd/man/machine-id.html) file, the underlying
+data, such as pod name and namespace should be treated as confidential, being the user&#39;s choice to expose it
+or not via another resource attribute.
+
+For applications running behind an application server (like unicorn), we do not recommend using one identifier
+for all processes participating in the application. Instead, it&#39;s recommended each division (e.g. a worker
+thread in unicorn) to have its own instance.id.
+
+It&#39;s not recommended for a Collector to set `service.instance.id` if it can&#39;t unambiguously determine the
+service instance that is generating that telemetry. For instance, creating an UUID based on `pod.name` will
+likely be wrong, as the Collector might not know from which container within that pod the telemetry originated.
+However, Collectors can set the `service.instance.id` if they can unambiguously determine the service instance
+for that telemetry. This is typically the case for scraping receivers, as they know the target address and
+port.
+  */
+  SERVICE_INSTANCE_ID: 'service.instance.id',
+
+  /**
+  * Logical name of the service.
+  *
+  * Note: MUST be the same for all instances of horizontally scaled services. If the value was not specified, SDKs MUST fallback to `unknown_service:` concatenated with [`process.executable.name`](process.md#process), e.g. `unknown_service:bash`. If `process.executable.name` is not available, the value MUST be set to `unknown_service`.
+  */
+  SERVICE_NAME: 'service.name',
+
+  /**
+  * A namespace for `service.name`.
+  *
+  * Note: A string value having a meaning that helps to distinguish a group of services, for example the team name that owns a group of services. `service.name` is expected to be unique within the same namespace. If `service.namespace` is not specified in the Resource then `service.name` is expected to be unique for all services that have no explicit namespace defined (so the empty/unspecified namespace is simply one more valid namespace). Zero-length namespace string is assumed equal to unspecified namespace.
+  */
+  SERVICE_NAMESPACE: 'service.namespace',
+
+  /**
+  * The version string of the service API or implementation. The format is not defined by these conventions.
+  */
+  SERVICE_VERSION: 'service.version',
+
+  /**
+  * A unique id to identify a session.
+  */
+  SESSION_ID: 'session.id',
+
+  /**
+  * The previous `session.id` for this user, when known.
+  */
+  SESSION_PREVIOUS_ID: 'session.previous_id',
+
+  /**
   * Source address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
   *
   * Note: When observed from the destination side, and when communicating through an intermediary, `source.address` SHOULD represent the source address behind any intermediaries, for example proxies, if it&#39;s available.
@@ -2938,6 +4811,41 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   * Source port number.
   */
   SOURCE_PORT: 'source.port',
+
+  /**
+  * The language of the telemetry SDK.
+  */
+  TELEMETRY_SDK_LANGUAGE: 'telemetry.sdk.language',
+
+  /**
+  * The name of the telemetry SDK as defined above.
+  *
+  * Note: The OpenTelemetry SDK MUST set the `telemetry.sdk.name` attribute to `opentelemetry`.
+If another SDK, like a fork or a vendor-provided implementation, is used, this SDK MUST set the
+`telemetry.sdk.name` attribute to the fully-qualified class or module name of this SDK&#39;s main entry point
+or another suitable identifier depending on the language.
+The identifier `opentelemetry` is reserved and MUST NOT be used in this case.
+All custom identifiers SHOULD be stable across different versions of an implementation.
+  */
+  TELEMETRY_SDK_NAME: 'telemetry.sdk.name',
+
+  /**
+  * The version string of the telemetry SDK.
+  */
+  TELEMETRY_SDK_VERSION: 'telemetry.sdk.version',
+
+  /**
+  * The name of the auto instrumentation agent or distribution, if used.
+  *
+  * Note: Official auto instrumentation agents and distributions SHOULD set the `telemetry.distro.name` attribute to
+a string starting with `opentelemetry-`, e.g. `opentelemetry-java-instrumentation`.
+  */
+  TELEMETRY_DISTRO_NAME: 'telemetry.distro.name',
+
+  /**
+  * The version string of the auto instrumentation agent or distribution, if used.
+  */
+  TELEMETRY_DISTRO_VERSION: 'telemetry.distro.version',
 
   /**
   * Current &#34;managed&#34; thread ID (as opposed to OS thread ID).
@@ -3097,6 +5005,20 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   TLS_SERVER_SUBJECT: 'tls.server.subject',
 
   /**
+  * Domain extracted from the `url.full`, such as &#34;opentelemetry.io&#34;.
+  *
+  * Note: In some cases a URL may refer to an IP and/or port directly, without a domain name. In this case, the IP address would go to the domain field. If the URL contains a [literal IPv6 address](https://www.rfc-editor.org/rfc/rfc2732#section-2) enclosed by `[` and `]`, the `[` and `]` characters should also be captured in the domain field.
+  */
+  URL_DOMAIN: 'url.domain',
+
+  /**
+  * The file extension extracted from the `url.full`, excluding the leading dot.
+  *
+  * Note: The file extension is only set if it exists, as not every url has a file extension. When the file name has multiple extensions `example.tar.gz`, only the last one should be captured `gz`, not `tar.gz`.
+  */
+  URL_EXTENSION: 'url.extension',
+
+  /**
   * The [URI fragment](https://www.rfc-editor.org/rfc/rfc3986#section-3.5) component.
   */
   URL_FRAGMENT: 'url.fragment',
@@ -3106,21 +5028,43 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   *
   * Note: For network calls, URL usually has `scheme://host[:port][path][?query][#fragment]` format, where the fragment is not transmitted over HTTP, but if it is known, it SHOULD be included nevertheless.
 `url.full` MUST NOT contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case username and password SHOULD be redacted and attribute&#39;s value SHOULD be `https://REDACTED:REDACTED@www.example.com/`.
-`url.full` SHOULD capture the absolute URL when it is available (or can be reconstructed) and SHOULD NOT be validated or modified except for sanitizing purposes.
+`url.full` SHOULD capture the absolute URL when it is available (or can be reconstructed). Sensitive content provided in `url.full` SHOULD be scrubbed when instrumentations can identify it.
   */
   URL_FULL: 'url.full',
 
   /**
+  * Unmodified original URL as seen in the event source.
+  *
+  * Note: In network monitoring, the observed URL may be a full URL, whereas in access logs, the URL is often just represented as a path. This field is meant to represent the URL as it was observed, complete or not.
+`url.original` might contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case password and username SHOULD NOT be redacted and attribute&#39;s value SHOULD remain the same.
+  */
+  URL_ORIGINAL: 'url.original',
+
+  /**
   * The [URI path](https://www.rfc-editor.org/rfc/rfc3986#section-3.3) component.
+  *
+  * Note: Sensitive content provided in `url.path` SHOULD be scrubbed when instrumentations can identify it.
   */
   URL_PATH: 'url.path',
 
   /**
+  * Port extracted from the `url.full`.
+  */
+  URL_PORT: 'url.port',
+
+  /**
   * The [URI query](https://www.rfc-editor.org/rfc/rfc3986#section-3.4) component.
   *
-  * Note: Sensitive content provided in query string SHOULD be scrubbed when instrumentations can identify it.
+  * Note: Sensitive content provided in `url.query` SHOULD be scrubbed when instrumentations can identify it.
   */
   URL_QUERY: 'url.query',
+
+  /**
+  * The highest registered url domain, stripped of the subdomain.
+  *
+  * Note: This value can be determined precisely with the [public suffix list](http://publicsuffix.org). For example, the registered domain for `foo.example.com` is `example.com`. Trying to approximate this by simply taking the last two labels will not work well for TLDs such as `co.uk`.
+  */
+  URL_REGISTERED_DOMAIN: 'url.registered_domain',
 
   /**
   * The [URI scheme](https://www.rfc-editor.org/rfc/rfc3986#section-3.1) component identifying the used protocol.
@@ -3128,19 +5072,37 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   URL_SCHEME: 'url.scheme',
 
   /**
+  * The subdomain portion of a fully qualified domain name includes all of the names except the host name under the registered_domain. In a partially qualified domain, or if the qualification level of the full name cannot be determined, subdomain contains all of the names below the registered domain.
+  *
+  * Note: The subdomain portion of `www.east.mydomain.co.uk` is `east`. If the domain has multiple levels of subdomain, such as `sub2.sub1.example.com`, the subdomain field should contain `sub2.sub1`, with no trailing period.
+  */
+  URL_SUBDOMAIN: 'url.subdomain',
+
+  /**
+  * The effective top level domain (eTLD), also known as the domain suffix, is the last part of the domain name. For example, the top level domain for example.com is `com`.
+  *
+  * Note: This value can be determined precisely with the [public suffix list](http://publicsuffix.org).
+  */
+  URL_TOP_LEVEL_DOMAIN: 'url.top_level_domain',
+
+  /**
+  * Name of the user-agent extracted from original. Usually refers to the browser&#39;s name.
+  *
+  * Note: [Example](https://www.whatsmyua.info) of extracting browser&#39;s name from original string. In the case of using a user-agent for non-browser products, such as microservices with multiple names/versions inside the `user_agent.original`, the most significant name SHOULD be selected. In such a scenario it should align with `user_agent.version`.
+  */
+  USER_AGENT_NAME: 'user_agent.name',
+
+  /**
   * Value of the [HTTP User-Agent](https://www.rfc-editor.org/rfc/rfc9110.html#field.user-agent) header sent by the client.
   */
   USER_AGENT_ORIGINAL: 'user_agent.original',
 
   /**
-  * A unique id to identify a session.
+  * Version of the user-agent extracted from original. Usually refers to the browser&#39;s version.
+  *
+  * Note: [Example](https://www.whatsmyua.info) of extracting browser&#39;s version from original string. In the case of using a user-agent for non-browser products, such as microservices with multiple names/versions inside the `user_agent.original`, the most significant version SHOULD be selected. In such a scenario it should align with `user_agent.name`.
   */
-  SESSION_ID: 'session.id',
-
-  /**
-  * The previous `session.id` for this user, when known.
-  */
-  SESSION_PREVIOUS_ID: 'session.previous_id',
+  USER_AGENT_VERSION: 'user_agent.version',
 
   /**
   * The full invoked ARN as provided on the `Context` passed to the function (`Lambda-Runtime-Invoked-Function-Arn` header on the `/runtime/invocation/next` applicable).
@@ -3148,31 +5110,6 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   * Note: This may be different from `cloud.resource_id` if an alias is involved.
   */
   AWS_LAMBDA_INVOKED_ARN: 'aws.lambda.invoked_arn',
-
-  /**
-  * The [event_id](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id) uniquely identifies the event.
-  */
-  CLOUDEVENTS_EVENT_ID: 'cloudevents.event_id',
-
-  /**
-  * The [source](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1) identifies the context in which an event happened.
-  */
-  CLOUDEVENTS_EVENT_SOURCE: 'cloudevents.event_source',
-
-  /**
-  * The [version of the CloudEvents specification](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion) which the event uses.
-  */
-  CLOUDEVENTS_EVENT_SPEC_VERSION: 'cloudevents.event_spec_version',
-
-  /**
-  * The [subject](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject) of the event in the context of the event producer (identified by source).
-  */
-  CLOUDEVENTS_EVENT_SUBJECT: 'cloudevents.event_subject',
-
-  /**
-  * The [event_type](https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type) contains a value describing the type of event related to the originating occurrence.
-  */
-  CLOUDEVENTS_EVENT_TYPE: 'cloudevents.event_type',
 
   /**
   * Parent-child Reference type.
@@ -3192,183 +5129,9 @@ different processes could be listening on TCP port 12345 and UDP port 12345.
   OTEL_STATUS_DESCRIPTION: 'otel.status_description',
 
   /**
-  * The invocation ID of the current function invocation.
-  */
-  FAAS_INVOCATION_ID: 'faas.invocation_id',
-
-  /**
-  * The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
-  */
-  FAAS_DOCUMENT_COLLECTION: 'faas.document.collection',
-
-  /**
-  * The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.
-  */
-  FAAS_DOCUMENT_NAME: 'faas.document.name',
-
-  /**
-  * Describes the type of the operation that was performed on the data.
-  */
-  FAAS_DOCUMENT_OPERATION: 'faas.document.operation',
-
-  /**
-  * A string containing the time when the data was accessed in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
-  */
-  FAAS_DOCUMENT_TIME: 'faas.document.time',
-
-  /**
-  * A string containing the schedule period as [Cron Expression](https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm).
-  */
-  FAAS_CRON: 'faas.cron',
-
-  /**
-  * A string containing the function invocation time in the [ISO 8601](https://www.iso.org/iso-8601-date-and-time-format.html) format expressed in [UTC](https://www.w3.org/TR/NOTE-datetime).
-  */
-  FAAS_TIME: 'faas.time',
-
-  /**
-  * A boolean that is true if the serverless function is executed for the first time (aka cold-start).
-  */
-  FAAS_COLDSTART: 'faas.coldstart',
-
-  /**
-  * The unique identifier of the feature flag.
-  */
-  FEATURE_FLAG_KEY: 'feature_flag.key',
-
-  /**
-  * The name of the service provider that performs the flag evaluation.
-  */
-  FEATURE_FLAG_PROVIDER_NAME: 'feature_flag.provider_name',
-
-  /**
-  * SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.
-  *
-  * Note: A semantic identifier, commonly referred to as a variant, provides a means
-for referring to a value without including the value itself. This can
-provide additional context for understanding the meaning behind a value.
-For example, the variant `red` maybe be used for the value `#c05543`.
-
-A stringified version of the value can be used in situations where a
-semantic identifier is unavailable. String representation of the value
-should be determined by the implementer.
-  */
-  FEATURE_FLAG_VARIANT: 'feature_flag.variant',
-
-  /**
   * The AWS request ID as returned in the response headers `x-amz-request-id` or `x-amz-requestid`.
   */
   AWS_REQUEST_ID: 'aws.request_id',
-
-  /**
-  * The value of the `AttributesToGet` request parameter.
-  */
-  AWS_DYNAMODB_ATTRIBUTES_TO_GET: 'aws.dynamodb.attributes_to_get',
-
-  /**
-  * The value of the `ConsistentRead` request parameter.
-  */
-  AWS_DYNAMODB_CONSISTENT_READ: 'aws.dynamodb.consistent_read',
-
-  /**
-  * The JSON-serialized value of each item in the `ConsumedCapacity` response field.
-  */
-  AWS_DYNAMODB_CONSUMED_CAPACITY: 'aws.dynamodb.consumed_capacity',
-
-  /**
-  * The value of the `IndexName` request parameter.
-  */
-  AWS_DYNAMODB_INDEX_NAME: 'aws.dynamodb.index_name',
-
-  /**
-  * The JSON-serialized value of the `ItemCollectionMetrics` response field.
-  */
-  AWS_DYNAMODB_ITEM_COLLECTION_METRICS: 'aws.dynamodb.item_collection_metrics',
-
-  /**
-  * The value of the `Limit` request parameter.
-  */
-  AWS_DYNAMODB_LIMIT: 'aws.dynamodb.limit',
-
-  /**
-  * The value of the `ProjectionExpression` request parameter.
-  */
-  AWS_DYNAMODB_PROJECTION: 'aws.dynamodb.projection',
-
-  /**
-  * The value of the `ProvisionedThroughput.ReadCapacityUnits` request parameter.
-  */
-  AWS_DYNAMODB_PROVISIONED_READ_CAPACITY: 'aws.dynamodb.provisioned_read_capacity',
-
-  /**
-  * The value of the `ProvisionedThroughput.WriteCapacityUnits` request parameter.
-  */
-  AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY: 'aws.dynamodb.provisioned_write_capacity',
-
-  /**
-  * The value of the `Select` request parameter.
-  */
-  AWS_DYNAMODB_SELECT: 'aws.dynamodb.select',
-
-  /**
-  * The keys in the `RequestItems` object field.
-  */
-  AWS_DYNAMODB_TABLE_NAMES: 'aws.dynamodb.table_names',
-
-  /**
-  * The JSON-serialized value of each item of the `GlobalSecondaryIndexes` request field.
-  */
-  AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES: 'aws.dynamodb.global_secondary_indexes',
-
-  /**
-  * The JSON-serialized value of each item of the `LocalSecondaryIndexes` request field.
-  */
-  AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES: 'aws.dynamodb.local_secondary_indexes',
-
-  /**
-  * The value of the `ExclusiveStartTableName` request parameter.
-  */
-  AWS_DYNAMODB_EXCLUSIVE_START_TABLE: 'aws.dynamodb.exclusive_start_table',
-
-  /**
-  * The the number of items in the `TableNames` response parameter.
-  */
-  AWS_DYNAMODB_TABLE_COUNT: 'aws.dynamodb.table_count',
-
-  /**
-  * The value of the `ScanIndexForward` request parameter.
-  */
-  AWS_DYNAMODB_SCAN_FORWARD: 'aws.dynamodb.scan_forward',
-
-  /**
-  * The value of the `Count` response parameter.
-  */
-  AWS_DYNAMODB_COUNT: 'aws.dynamodb.count',
-
-  /**
-  * The value of the `ScannedCount` response parameter.
-  */
-  AWS_DYNAMODB_SCANNED_COUNT: 'aws.dynamodb.scanned_count',
-
-  /**
-  * The value of the `Segment` request parameter.
-  */
-  AWS_DYNAMODB_SEGMENT: 'aws.dynamodb.segment',
-
-  /**
-  * The value of the `TotalSegments` request parameter.
-  */
-  AWS_DYNAMODB_TOTAL_SEGMENTS: 'aws.dynamodb.total_segments',
-
-  /**
-  * The JSON-serialized value of each item in the `AttributeDefinitions` request field.
-  */
-  AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS: 'aws.dynamodb.attribute_definitions',
-
-  /**
-  * The JSON-serialized value of each item in the the `GlobalSecondaryIndexUpdates` request field.
-  */
-  AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES: 'aws.dynamodb.global_secondary_index_updates',
 
   /**
   * The S3 bucket name the request refers to. Corresponds to the `--bucket` parameter of the [S3 API](https://docs.aws.amazon.com/cli/latest/reference/s3api/index.html) operations.
@@ -3491,14 +5254,6 @@ This applies in particular to the following operations:
  * @deprecated Use the SEMATTRS_XXXXX constants rather than the SemanticAttributes.XXXXX for bundle minification
  */
 export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap<SemanticAttributes>([
-  TMP_FAAS_INVOKED_NAME,
-  TMP_FAAS_INVOKED_PROVIDER,
-  TMP_FAAS_INVOKED_REGION,
-  TMP_FAAS_TRIGGER,
-  TMP_PEER_SERVICE,
-  TMP_ENDUSER_ID,
-  TMP_ENDUSER_ROLE,
-  TMP_ENDUSER_SCOPE,
   TMP_EVENT_NAME,
   TMP_LOG_RECORD_UID,
   TMP_LOG_IOSTREAM,
@@ -3510,9 +5265,9 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_ANDROID_STATE,
   TMP_POOL_NAME,
   TMP_STATE,
+  TMP_ASPNETCORE_RATE_LIMITING_RESULT,
   TMP_ASPNETCORE_DIAGNOSTICS_HANDLER_TYPE,
   TMP_ASPNETCORE_RATE_LIMITING_POLICY,
-  TMP_ASPNETCORE_RATE_LIMITING_RESULT,
   TMP_ASPNETCORE_REQUEST_IS_UNHANDLED,
   TMP_ASPNETCORE_ROUTING_IS_FALLBACK,
   TMP_SIGNALR_CONNECTION_STATUS,
@@ -3520,6 +5275,7 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_JVM_BUFFER_POOL_NAME,
   TMP_JVM_MEMORY_POOL_NAME,
   TMP_JVM_MEMORY_TYPE,
+  TMP_PROCESS_CPU_STATE,
   TMP_SYSTEM_DEVICE,
   TMP_SYSTEM_CPU_LOGICAL_NUMBER,
   TMP_SYSTEM_CPU_STATE,
@@ -3532,15 +5288,64 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_SYSTEM_FILESYSTEM_STATE,
   TMP_SYSTEM_FILESYSTEM_TYPE,
   TMP_SYSTEM_NETWORK_STATE,
-  TMP_SYSTEM_PROCESSES_STATUS,
+  TMP_SYSTEM_PROCESS_STATUS,
+  TMP_ANDROID_OS_API_LEVEL,
+  TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS,
+  TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET,
+  TMP_AWS_DYNAMODB_CONSISTENT_READ,
+  TMP_AWS_DYNAMODB_CONSUMED_CAPACITY,
+  TMP_AWS_DYNAMODB_COUNT,
+  TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE,
+  TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES,
+  TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES,
+  TMP_AWS_DYNAMODB_INDEX_NAME,
+  TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS,
+  TMP_AWS_DYNAMODB_LIMIT,
+  TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES,
+  TMP_AWS_DYNAMODB_PROJECTION,
+  TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY,
+  TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY,
+  TMP_AWS_DYNAMODB_SCAN_FORWARD,
+  TMP_AWS_DYNAMODB_SCANNED_COUNT,
+  TMP_AWS_DYNAMODB_SEGMENT,
+  TMP_AWS_DYNAMODB_SELECT,
+  TMP_AWS_DYNAMODB_TABLE_COUNT,
+  TMP_AWS_DYNAMODB_TABLE_NAMES,
+  TMP_AWS_DYNAMODB_TOTAL_SEGMENTS,
+  TMP_BROWSER_BRANDS,
+  TMP_BROWSER_LANGUAGE,
+  TMP_BROWSER_MOBILE,
+  TMP_BROWSER_PLATFORM,
   TMP_CLIENT_ADDRESS,
   TMP_CLIENT_PORT,
+  TMP_CLOUD_ACCOUNT_ID,
+  TMP_CLOUD_AVAILABILITY_ZONE,
+  TMP_CLOUD_PLATFORM,
+  TMP_CLOUD_PROVIDER,
+  TMP_CLOUD_REGION,
+  TMP_CLOUD_RESOURCE_ID,
+  TMP_CLOUDEVENTS_EVENT_ID,
+  TMP_CLOUDEVENTS_EVENT_SOURCE,
+  TMP_CLOUDEVENTS_EVENT_SPEC_VERSION,
+  TMP_CLOUDEVENTS_EVENT_SUBJECT,
+  TMP_CLOUDEVENTS_EVENT_TYPE,
   TMP_CODE_COLUMN,
   TMP_CODE_FILEPATH,
   TMP_CODE_FUNCTION,
   TMP_CODE_LINENO,
   TMP_CODE_NAMESPACE,
   TMP_CODE_STACKTRACE,
+  TMP_CONTAINER_COMMAND,
+  TMP_CONTAINER_COMMAND_ARGS,
+  TMP_CONTAINER_COMMAND_LINE,
+  TMP_CONTAINER_CPU_STATE,
+  TMP_CONTAINER_ID,
+  TMP_CONTAINER_IMAGE_ID,
+  TMP_CONTAINER_IMAGE_NAME,
+  TMP_CONTAINER_IMAGE_REPO_DIGESTS,
+  TMP_CONTAINER_IMAGE_TAGS,
+  TMP_CONTAINER_NAME,
+  TMP_CONTAINER_RUNTIME,
   TMP_DB_CASSANDRA_CONSISTENCY_LEVEL,
   TMP_DB_CASSANDRA_COORDINATOR_DC,
   TMP_DB_CASSANDRA_COORDINATOR_ID,
@@ -3548,7 +5353,6 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_DB_CASSANDRA_PAGE_SIZE,
   TMP_DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT,
   TMP_DB_CASSANDRA_TABLE,
-  TMP_DB_CONNECTION_STRING,
   TMP_DB_COSMOSDB_CLIENT_ID,
   TMP_DB_COSMOSDB_CONNECTION_MODE,
   TMP_DB_COSMOSDB_CONTAINER,
@@ -3558,9 +5362,7 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_DB_COSMOSDB_STATUS_CODE,
   TMP_DB_COSMOSDB_SUB_STATUS_CODE,
   TMP_DB_ELASTICSEARCH_CLUSTER_NAME,
-  TMP_DB_ELASTICSEARCH_NODE_NAME,
   TMP_DB_INSTANCE_ID,
-  TMP_DB_JDBC_DRIVER_CLASSNAME,
   TMP_DB_MONGODB_COLLECTION,
   TMP_DB_MSSQL_INSTANCE_NAME,
   TMP_DB_NAME,
@@ -3570,6 +5372,10 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_DB_STATEMENT,
   TMP_DB_SYSTEM,
   TMP_DB_USER,
+  TMP_DEPLOYMENT_ENVIRONMENT,
+  TMP_DB_CONNECTION_STRING,
+  TMP_DB_ELASTICSEARCH_NODE_NAME,
+  TMP_DB_JDBC_DRIVER_CLASSNAME,
   TMP_HTTP_FLAVOR,
   TMP_HTTP_METHOD,
   TMP_HTTP_REQUEST_CONTENT_LENGTH,
@@ -3579,6 +5385,7 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_HTTP_TARGET,
   TMP_HTTP_URL,
   TMP_HTTP_USER_AGENT,
+  TMP_MESSAGING_KAFKA_DESTINATION_PARTITION,
   TMP_NET_HOST_NAME,
   TMP_NET_HOST_PORT,
   TMP_NET_PEER_NAME,
@@ -3592,32 +5399,110 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_NET_SOCK_PEER_NAME,
   TMP_NET_SOCK_PEER_PORT,
   TMP_NET_TRANSPORT,
+  TMP_SYSTEM_PROCESSES_STATUS,
   TMP_DESTINATION_ADDRESS,
   TMP_DESTINATION_PORT,
+  TMP_DEVICE_ID,
+  TMP_DEVICE_MANUFACTURER,
+  TMP_DEVICE_MODEL_IDENTIFIER,
+  TMP_DEVICE_MODEL_NAME,
   TMP_DISK_IO_DIRECTION,
+  TMP_DNS_QUESTION_NAME,
+  TMP_ENDUSER_ID,
+  TMP_ENDUSER_ROLE,
+  TMP_ENDUSER_SCOPE,
   TMP_ERROR_TYPE,
   TMP_EXCEPTION_ESCAPED,
   TMP_EXCEPTION_MESSAGE,
   TMP_EXCEPTION_STACKTRACE,
   TMP_EXCEPTION_TYPE,
+  TMP_FAAS_COLDSTART,
+  TMP_FAAS_CRON,
+  TMP_FAAS_DOCUMENT_COLLECTION,
+  TMP_FAAS_DOCUMENT_NAME,
+  TMP_FAAS_DOCUMENT_OPERATION,
+  TMP_FAAS_DOCUMENT_TIME,
+  TMP_FAAS_INSTANCE,
+  TMP_FAAS_INVOCATION_ID,
+  TMP_FAAS_INVOKED_NAME,
+  TMP_FAAS_INVOKED_PROVIDER,
+  TMP_FAAS_INVOKED_REGION,
+  TMP_FAAS_MAX_MEMORY,
+  TMP_FAAS_NAME,
+  TMP_FAAS_TIME,
+  TMP_FAAS_TRIGGER,
+  TMP_FAAS_VERSION,
+  TMP_FEATURE_FLAG_KEY,
+  TMP_FEATURE_FLAG_PROVIDER_NAME,
+  TMP_FEATURE_FLAG_VARIANT,
+  TMP_FILE_DIRECTORY,
+  TMP_FILE_EXTENSION,
+  TMP_FILE_NAME,
+  TMP_FILE_PATH,
+  TMP_FILE_SIZE,
+  TMP_GCP_CLOUD_RUN_JOB_EXECUTION,
+  TMP_GCP_CLOUD_RUN_JOB_TASK_INDEX,
+  TMP_GCP_GCE_INSTANCE_HOSTNAME,
+  TMP_GCP_GCE_INSTANCE_NAME,
+  TMP_HOST_ARCH,
+  TMP_HOST_CPU_CACHE_L2_SIZE,
+  TMP_HOST_CPU_FAMILY,
+  TMP_HOST_CPU_MODEL_ID,
+  TMP_HOST_CPU_MODEL_NAME,
+  TMP_HOST_CPU_STEPPING,
+  TMP_HOST_CPU_VENDOR_ID,
+  TMP_HOST_ID,
+  TMP_HOST_IMAGE_ID,
+  TMP_HOST_IMAGE_NAME,
+  TMP_HOST_IMAGE_VERSION,
+  TMP_HOST_IP,
+  TMP_HOST_MAC,
+  TMP_HOST_NAME,
+  TMP_HOST_TYPE,
+  TMP_HTTP_CONNECTION_STATE,
   TMP_HTTP_REQUEST_BODY_SIZE,
   TMP_HTTP_REQUEST_METHOD,
   TMP_HTTP_REQUEST_METHOD_ORIGINAL,
   TMP_HTTP_REQUEST_RESEND_COUNT,
+  TMP_HTTP_REQUEST_SIZE,
   TMP_HTTP_RESPONSE_BODY_SIZE,
+  TMP_HTTP_RESPONSE_SIZE,
   TMP_HTTP_RESPONSE_STATUS_CODE,
   TMP_HTTP_ROUTE,
+  TMP_K8S_CLUSTER_NAME,
+  TMP_K8S_CLUSTER_UID,
+  TMP_K8S_CONTAINER_NAME,
+  TMP_K8S_CONTAINER_RESTART_COUNT,
+  TMP_K8S_CRONJOB_NAME,
+  TMP_K8S_CRONJOB_UID,
+  TMP_K8S_DAEMONSET_NAME,
+  TMP_K8S_DAEMONSET_UID,
+  TMP_K8S_DEPLOYMENT_NAME,
+  TMP_K8S_DEPLOYMENT_UID,
+  TMP_K8S_JOB_NAME,
+  TMP_K8S_JOB_UID,
+  TMP_K8S_NAMESPACE_NAME,
+  TMP_K8S_NODE_NAME,
+  TMP_K8S_NODE_UID,
+  TMP_K8S_POD_NAME,
+  TMP_K8S_POD_UID,
+  TMP_K8S_REPLICASET_NAME,
+  TMP_K8S_REPLICASET_UID,
+  TMP_K8S_STATEFULSET_NAME,
+  TMP_K8S_STATEFULSET_UID,
   TMP_MESSAGING_BATCH_MESSAGE_COUNT,
   TMP_MESSAGING_CLIENT_ID,
   TMP_MESSAGING_DESTINATION_ANONYMOUS,
   TMP_MESSAGING_DESTINATION_NAME,
+  TMP_MESSAGING_DESTINATION_PARTITION_ID,
   TMP_MESSAGING_DESTINATION_TEMPLATE,
   TMP_MESSAGING_DESTINATION_TEMPORARY,
   TMP_MESSAGING_DESTINATION_PUBLISH_ANONYMOUS,
   TMP_MESSAGING_DESTINATION_PUBLISH_NAME,
+  TMP_MESSAGING_EVENTHUBS_CONSUMER_GROUP,
+  TMP_MESSAGING_EVENTHUBS_MESSAGE_ENQUEUED_TIME,
   TMP_MESSAGING_GCP_PUBSUB_MESSAGE_ORDERING_KEY,
   TMP_MESSAGING_KAFKA_CONSUMER_GROUP,
-  TMP_MESSAGING_KAFKA_DESTINATION_PARTITION,
   TMP_MESSAGING_KAFKA_MESSAGE_KEY,
   TMP_MESSAGING_KAFKA_MESSAGE_OFFSET,
   TMP_MESSAGING_KAFKA_MESSAGE_TOMBSTONE,
@@ -3627,6 +5512,7 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_MESSAGING_MESSAGE_ID,
   TMP_MESSAGING_OPERATION,
   TMP_MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY,
+  TMP_MESSAGING_RABBITMQ_MESSAGE_DELIVERY_TAG,
   TMP_MESSAGING_ROCKETMQ_CLIENT_GROUP,
   TMP_MESSAGING_ROCKETMQ_CONSUMPTION_MODEL,
   TMP_MESSAGING_ROCKETMQ_MESSAGE_DELAY_TIME_LEVEL,
@@ -3636,6 +5522,10 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_MESSAGING_ROCKETMQ_MESSAGE_TAG,
   TMP_MESSAGING_ROCKETMQ_MESSAGE_TYPE,
   TMP_MESSAGING_ROCKETMQ_NAMESPACE,
+  TMP_MESSAGING_SERVICEBUS_DESTINATION_SUBSCRIPTION_NAME,
+  TMP_MESSAGING_SERVICEBUS_DISPOSITION_STATUS,
+  TMP_MESSAGING_SERVICEBUS_MESSAGE_DELIVERY_COUNT,
+  TMP_MESSAGING_SERVICEBUS_MESSAGE_ENQUEUED_TIME,
   TMP_MESSAGING_SYSTEM,
   TMP_NETWORK_CARRIER_ICC,
   TMP_NETWORK_CARRIER_MCC,
@@ -3652,6 +5542,24 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_NETWORK_PROTOCOL_VERSION,
   TMP_NETWORK_TRANSPORT,
   TMP_NETWORK_TYPE,
+  TMP_OCI_MANIFEST_DIGEST,
+  TMP_OS_BUILD_ID,
+  TMP_OS_DESCRIPTION,
+  TMP_OS_NAME,
+  TMP_OS_TYPE,
+  TMP_OS_VERSION,
+  TMP_PEER_SERVICE,
+  TMP_PROCESS_COMMAND,
+  TMP_PROCESS_COMMAND_ARGS,
+  TMP_PROCESS_COMMAND_LINE,
+  TMP_PROCESS_EXECUTABLE_NAME,
+  TMP_PROCESS_EXECUTABLE_PATH,
+  TMP_PROCESS_OWNER,
+  TMP_PROCESS_PARENT_PID,
+  TMP_PROCESS_PID,
+  TMP_PROCESS_RUNTIME_DESCRIPTION,
+  TMP_PROCESS_RUNTIME_NAME,
+  TMP_PROCESS_RUNTIME_VERSION,
   TMP_RPC_CONNECT_RPC_ERROR_CODE,
   TMP_RPC_GRPC_STATUS_CODE,
   TMP_RPC_JSONRPC_ERROR_CODE,
@@ -3663,8 +5571,19 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_RPC_SYSTEM,
   TMP_SERVER_ADDRESS,
   TMP_SERVER_PORT,
+  TMP_SERVICE_INSTANCE_ID,
+  TMP_SERVICE_NAME,
+  TMP_SERVICE_NAMESPACE,
+  TMP_SERVICE_VERSION,
+  TMP_SESSION_ID,
+  TMP_SESSION_PREVIOUS_ID,
   TMP_SOURCE_ADDRESS,
   TMP_SOURCE_PORT,
+  TMP_TELEMETRY_SDK_LANGUAGE,
+  TMP_TELEMETRY_SDK_NAME,
+  TMP_TELEMETRY_SDK_VERSION,
+  TMP_TELEMETRY_DISTRO_NAME,
+  TMP_TELEMETRY_DISTRO_VERSION,
   TMP_THREAD_ID,
   TMP_THREAD_NAME,
   TMP_TLS_CIPHER,
@@ -3696,57 +5615,26 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_TLS_SERVER_NOT_AFTER,
   TMP_TLS_SERVER_NOT_BEFORE,
   TMP_TLS_SERVER_SUBJECT,
+  TMP_URL_DOMAIN,
+  TMP_URL_EXTENSION,
   TMP_URL_FRAGMENT,
   TMP_URL_FULL,
+  TMP_URL_ORIGINAL,
   TMP_URL_PATH,
+  TMP_URL_PORT,
   TMP_URL_QUERY,
+  TMP_URL_REGISTERED_DOMAIN,
   TMP_URL_SCHEME,
+  TMP_URL_SUBDOMAIN,
+  TMP_URL_TOP_LEVEL_DOMAIN,
+  TMP_USER_AGENT_NAME,
   TMP_USER_AGENT_ORIGINAL,
-  TMP_SESSION_ID,
-  TMP_SESSION_PREVIOUS_ID,
+  TMP_USER_AGENT_VERSION,
   TMP_AWS_LAMBDA_INVOKED_ARN,
-  TMP_CLOUDEVENTS_EVENT_ID,
-  TMP_CLOUDEVENTS_EVENT_SOURCE,
-  TMP_CLOUDEVENTS_EVENT_SPEC_VERSION,
-  TMP_CLOUDEVENTS_EVENT_SUBJECT,
-  TMP_CLOUDEVENTS_EVENT_TYPE,
   TMP_OPENTRACING_REF_TYPE,
   TMP_OTEL_STATUS_CODE,
   TMP_OTEL_STATUS_DESCRIPTION,
-  TMP_FAAS_INVOCATION_ID,
-  TMP_FAAS_DOCUMENT_COLLECTION,
-  TMP_FAAS_DOCUMENT_NAME,
-  TMP_FAAS_DOCUMENT_OPERATION,
-  TMP_FAAS_DOCUMENT_TIME,
-  TMP_FAAS_CRON,
-  TMP_FAAS_TIME,
-  TMP_FAAS_COLDSTART,
-  TMP_FEATURE_FLAG_KEY,
-  TMP_FEATURE_FLAG_PROVIDER_NAME,
-  TMP_FEATURE_FLAG_VARIANT,
   TMP_AWS_REQUEST_ID,
-  TMP_AWS_DYNAMODB_ATTRIBUTES_TO_GET,
-  TMP_AWS_DYNAMODB_CONSISTENT_READ,
-  TMP_AWS_DYNAMODB_CONSUMED_CAPACITY,
-  TMP_AWS_DYNAMODB_INDEX_NAME,
-  TMP_AWS_DYNAMODB_ITEM_COLLECTION_METRICS,
-  TMP_AWS_DYNAMODB_LIMIT,
-  TMP_AWS_DYNAMODB_PROJECTION,
-  TMP_AWS_DYNAMODB_PROVISIONED_READ_CAPACITY,
-  TMP_AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY,
-  TMP_AWS_DYNAMODB_SELECT,
-  TMP_AWS_DYNAMODB_TABLE_NAMES,
-  TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES,
-  TMP_AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES,
-  TMP_AWS_DYNAMODB_EXCLUSIVE_START_TABLE,
-  TMP_AWS_DYNAMODB_TABLE_COUNT,
-  TMP_AWS_DYNAMODB_SCAN_FORWARD,
-  TMP_AWS_DYNAMODB_COUNT,
-  TMP_AWS_DYNAMODB_SCANNED_COUNT,
-  TMP_AWS_DYNAMODB_SEGMENT,
-  TMP_AWS_DYNAMODB_TOTAL_SEGMENTS,
-  TMP_AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS,
-  TMP_AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES,
   TMP_AWS_S3_BUCKET,
   TMP_AWS_S3_COPY_SOURCE,
   TMP_AWS_S3_DELETE,
@@ -3761,174 +5649,6 @@ export const SemanticAttributes:SemanticAttributes = /*#__PURE__*/createConstMap
   TMP_MESSAGE_TYPE,
   TMP_MESSAGE_UNCOMPRESSED_SIZE,
 ]);
-
-/* ----------------------------------------------------------------------------------------------------------
- * Constant values for FaasInvokedProviderValues enum definition
- *
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- * ---------------------------------------------------------------------------------------------------------- */
-
-// Temporary local constants to assign to the individual exports and the namespaced version
-// Required to avoid the namespace exports using the unminifiable export names for some package types
-const TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD = 'alibaba_cloud';
-const TMP_FAASINVOKEDPROVIDERVALUES_AWS = 'aws';
-const TMP_FAASINVOKEDPROVIDERVALUES_AZURE = 'azure';
-const TMP_FAASINVOKEDPROVIDERVALUES_GCP = 'gcp';
-const TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD = 'tencent_cloud';
-
-/**
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- */
-export const FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD = TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD;
-
-/**
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- */
-export const FAASINVOKEDPROVIDERVALUES_AWS = TMP_FAASINVOKEDPROVIDERVALUES_AWS;
-
-/**
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- */
-export const FAASINVOKEDPROVIDERVALUES_AZURE = TMP_FAASINVOKEDPROVIDERVALUES_AZURE;
-
-/**
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- */
-export const FAASINVOKEDPROVIDERVALUES_GCP = TMP_FAASINVOKEDPROVIDERVALUES_GCP;
-
-/**
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- */
-export const FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD = TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD;
-
-/**
- * Identifies the Values for FaasInvokedProviderValues enum definition
- *
- * The cloud provider of the invoked function.
- *
- * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
- * @deprecated Use the FAASINVOKEDPROVIDERVALUES_XXXXX constants rather than the FaasInvokedProviderValues.XXXXX for bundle minification.
- */
-export type FaasInvokedProviderValues = {
-
-  /** Alibaba Cloud. */
-  ALIBABA_CLOUD: 'alibaba_cloud',
-
-  /** Amazon Web Services. */
-  AWS: 'aws',
-
-  /** Microsoft Azure. */
-  AZURE: 'azure',
-
-  /** Google Cloud Platform. */
-  GCP: 'gcp',
-
-  /** Tencent Cloud. */
-  TENCENT_CLOUD: 'tencent_cloud',
-}
-
-/**
- * The constant map of values for FaasInvokedProviderValues.
- * @deprecated Use the FAASINVOKEDPROVIDERVALUES_XXXXX constants rather than the FaasInvokedProviderValues.XXXXX for bundle minification.
- */
-export const FaasInvokedProviderValues:FaasInvokedProviderValues = /*#__PURE__*/createConstMap<FaasInvokedProviderValues>([
-  TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD,
-  TMP_FAASINVOKEDPROVIDERVALUES_AWS,
-  TMP_FAASINVOKEDPROVIDERVALUES_AZURE,
-  TMP_FAASINVOKEDPROVIDERVALUES_GCP,
-  TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD,
-]);
-
-
-
-/* ----------------------------------------------------------------------------------------------------------
- * Constant values for FaasTriggerValues enum definition
- *
- * Type of the trigger which caused this function invocation.
- * ---------------------------------------------------------------------------------------------------------- */
-
-// Temporary local constants to assign to the individual exports and the namespaced version
-// Required to avoid the namespace exports using the unminifiable export names for some package types
-const TMP_FAASTRIGGERVALUES_DATASOURCE = 'datasource';
-const TMP_FAASTRIGGERVALUES_HTTP = 'http';
-const TMP_FAASTRIGGERVALUES_PUBSUB = 'pubsub';
-const TMP_FAASTRIGGERVALUES_TIMER = 'timer';
-const TMP_FAASTRIGGERVALUES_OTHER = 'other';
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-export const FAASTRIGGERVALUES_DATASOURCE = TMP_FAASTRIGGERVALUES_DATASOURCE;
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-export const FAASTRIGGERVALUES_HTTP = TMP_FAASTRIGGERVALUES_HTTP;
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-export const FAASTRIGGERVALUES_PUBSUB = TMP_FAASTRIGGERVALUES_PUBSUB;
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-export const FAASTRIGGERVALUES_TIMER = TMP_FAASTRIGGERVALUES_TIMER;
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-export const FAASTRIGGERVALUES_OTHER = TMP_FAASTRIGGERVALUES_OTHER;
-
-/**
- * Identifies the Values for FaasTriggerValues enum definition
- *
- * Type of the trigger which caused this function invocation.
- * @deprecated Use the FAASTRIGGERVALUES_XXXXX constants rather than the FaasTriggerValues.XXXXX for bundle minification.
- */
-export type FaasTriggerValues = {
-
-  /** A response to some data source operation such as a database or filesystem read/write. */
-  DATASOURCE: 'datasource',
-
-  /** To provide an answer to an inbound HTTP request. */
-  HTTP: 'http',
-
-  /** A function is set to be executed when messages are sent to a messaging system. */
-  PUBSUB: 'pubsub',
-
-  /** A function is scheduled to be executed regularly. */
-  TIMER: 'timer',
-
-  /** If none of the others apply. */
-  OTHER: 'other',
-}
-
-/**
- * The constant map of values for FaasTriggerValues.
- * @deprecated Use the FAASTRIGGERVALUES_XXXXX constants rather than the FaasTriggerValues.XXXXX for bundle minification.
- */
-export const FaasTriggerValues:FaasTriggerValues = /*#__PURE__*/createConstMap<FaasTriggerValues>([
-  TMP_FAASTRIGGERVALUES_DATASOURCE,
-  TMP_FAASTRIGGERVALUES_HTTP,
-  TMP_FAASTRIGGERVALUES_PUBSUB,
-  TMP_FAASTRIGGERVALUES_TIMER,
-  TMP_FAASTRIGGERVALUES_OTHER,
-]);
-
-
 
 /* ----------------------------------------------------------------------------------------------------------
  * Constant values for LogIostreamValues enum definition
@@ -4411,9 +6131,66 @@ export const JvmMemoryTypeValues:JvmMemoryTypeValues = /*#__PURE__*/createConstM
 
 
 /* ----------------------------------------------------------------------------------------------------------
+ * Constant values for ProcessCpuStateValues enum definition
+ *
+ * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_PROCESSCPUSTATEVALUES_SYSTEM = 'system';
+const TMP_PROCESSCPUSTATEVALUES_USER = 'user';
+const TMP_PROCESSCPUSTATEVALUES_WAIT = 'wait';
+
+/**
+ * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+ */
+export const PROCESSCPUSTATEVALUES_SYSTEM = TMP_PROCESSCPUSTATEVALUES_SYSTEM;
+
+/**
+ * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+ */
+export const PROCESSCPUSTATEVALUES_USER = TMP_PROCESSCPUSTATEVALUES_USER;
+
+/**
+ * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+ */
+export const PROCESSCPUSTATEVALUES_WAIT = TMP_PROCESSCPUSTATEVALUES_WAIT;
+
+/**
+ * Identifies the Values for ProcessCpuStateValues enum definition
+ *
+ * The CPU state for this data point. A process SHOULD be characterized _either_ by data points with no `state` labels, _or only_ data points with `state` labels.
+ * @deprecated Use the PROCESSCPUSTATEVALUES_XXXXX constants rather than the ProcessCpuStateValues.XXXXX for bundle minification.
+ */
+export type ProcessCpuStateValues = {
+
+  /** system. */
+  SYSTEM: 'system',
+
+  /** user. */
+  USER: 'user',
+
+  /** wait. */
+  WAIT: 'wait',
+}
+
+/**
+ * The constant map of values for ProcessCpuStateValues.
+ * @deprecated Use the PROCESSCPUSTATEVALUES_XXXXX constants rather than the ProcessCpuStateValues.XXXXX for bundle minification.
+ */
+export const ProcessCpuStateValues:ProcessCpuStateValues = /*#__PURE__*/createConstMap<ProcessCpuStateValues>([
+  TMP_PROCESSCPUSTATEVALUES_SYSTEM,
+  TMP_PROCESSCPUSTATEVALUES_USER,
+  TMP_PROCESSCPUSTATEVALUES_WAIT,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
  * Constant values for SystemCpuStateValues enum definition
  *
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  * ---------------------------------------------------------------------------------------------------------- */
 
 // Temporary local constants to assign to the individual exports and the namespaced version
@@ -4427,44 +6204,44 @@ const TMP_SYSTEMCPUSTATEVALUES_INTERRUPT = 'interrupt';
 const TMP_SYSTEMCPUSTATEVALUES_STEAL = 'steal';
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_USER = TMP_SYSTEMCPUSTATEVALUES_USER;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_SYSTEM = TMP_SYSTEMCPUSTATEVALUES_SYSTEM;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_NICE = TMP_SYSTEMCPUSTATEVALUES_NICE;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_IDLE = TMP_SYSTEMCPUSTATEVALUES_IDLE;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_IOWAIT = TMP_SYSTEMCPUSTATEVALUES_IOWAIT;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_INTERRUPT = TMP_SYSTEMCPUSTATEVALUES_INTERRUPT;
 
 /**
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  */
 export const SYSTEMCPUSTATEVALUES_STEAL = TMP_SYSTEMCPUSTATEVALUES_STEAL;
 
 /**
  * Identifies the Values for SystemCpuStateValues enum definition
  *
- * The state of the CPU.
+ * The CPU state for this data point. A system&#39;s CPU SHOULD be characterized *either* by data points with no `state` labels, *or only* data points with `state` labels.
  * @deprecated Use the SYSTEMCPUSTATEVALUES_XXXXX constants rather than the SystemCpuStateValues.XXXXX for bundle minification.
  */
 export type SystemCpuStateValues = {
@@ -5017,45 +6794,45 @@ export const SystemNetworkStateValues:SystemNetworkStateValues = /*#__PURE__*/cr
 
 
 /* ----------------------------------------------------------------------------------------------------------
- * Constant values for SystemProcessesStatusValues enum definition
+ * Constant values for SystemProcessStatusValues enum definition
  *
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
  * ---------------------------------------------------------------------------------------------------------- */
 
 // Temporary local constants to assign to the individual exports and the namespaced version
 // Required to avoid the namespace exports using the unminifiable export names for some package types
-const TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING = 'running';
-const TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING = 'sleeping';
-const TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED = 'stopped';
-const TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT = 'defunct';
+const TMP_SYSTEMPROCESSSTATUSVALUES_RUNNING = 'running';
+const TMP_SYSTEMPROCESSSTATUSVALUES_SLEEPING = 'sleeping';
+const TMP_SYSTEMPROCESSSTATUSVALUES_STOPPED = 'stopped';
+const TMP_SYSTEMPROCESSSTATUSVALUES_DEFUNCT = 'defunct';
 
 /**
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
  */
-export const SYSTEMPROCESSESSTATUSVALUES_RUNNING = TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING;
+export const SYSTEMPROCESSSTATUSVALUES_RUNNING = TMP_SYSTEMPROCESSSTATUSVALUES_RUNNING;
 
 /**
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
  */
-export const SYSTEMPROCESSESSTATUSVALUES_SLEEPING = TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING;
+export const SYSTEMPROCESSSTATUSVALUES_SLEEPING = TMP_SYSTEMPROCESSSTATUSVALUES_SLEEPING;
 
 /**
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
  */
-export const SYSTEMPROCESSESSTATUSVALUES_STOPPED = TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED;
+export const SYSTEMPROCESSSTATUSVALUES_STOPPED = TMP_SYSTEMPROCESSSTATUSVALUES_STOPPED;
 
 /**
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
  */
-export const SYSTEMPROCESSESSTATUSVALUES_DEFUNCT = TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT;
+export const SYSTEMPROCESSSTATUSVALUES_DEFUNCT = TMP_SYSTEMPROCESSSTATUSVALUES_DEFUNCT;
 
 /**
- * Identifies the Values for SystemProcessesStatusValues enum definition
+ * Identifies the Values for SystemProcessStatusValues enum definition
  *
  * The process state, e.g., [Linux Process State Codes](https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES).
- * @deprecated Use the SYSTEMPROCESSESSTATUSVALUES_XXXXX constants rather than the SystemProcessesStatusValues.XXXXX for bundle minification.
+ * @deprecated Use the SYSTEMPROCESSSTATUSVALUES_XXXXX constants rather than the SystemProcessStatusValues.XXXXX for bundle minification.
  */
-export type SystemProcessesStatusValues = {
+export type SystemProcessStatusValues = {
 
   /** running. */
   RUNNING: 'running',
@@ -5071,14 +6848,535 @@ export type SystemProcessesStatusValues = {
 }
 
 /**
- * The constant map of values for SystemProcessesStatusValues.
- * @deprecated Use the SYSTEMPROCESSESSTATUSVALUES_XXXXX constants rather than the SystemProcessesStatusValues.XXXXX for bundle minification.
+ * The constant map of values for SystemProcessStatusValues.
+ * @deprecated Use the SYSTEMPROCESSSTATUSVALUES_XXXXX constants rather than the SystemProcessStatusValues.XXXXX for bundle minification.
  */
-export const SystemProcessesStatusValues:SystemProcessesStatusValues = /*#__PURE__*/createConstMap<SystemProcessesStatusValues>([
-  TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING,
-  TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING,
-  TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED,
-  TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT,
+export const SystemProcessStatusValues:SystemProcessStatusValues = /*#__PURE__*/createConstMap<SystemProcessStatusValues>([
+  TMP_SYSTEMPROCESSSTATUSVALUES_RUNNING,
+  TMP_SYSTEMPROCESSSTATUSVALUES_SLEEPING,
+  TMP_SYSTEMPROCESSSTATUSVALUES_STOPPED,
+  TMP_SYSTEMPROCESSSTATUSVALUES_DEFUNCT,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for CloudPlatformValues enum definition
+ *
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_ECS = 'alibaba_cloud_ecs';
+const TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_FC = 'alibaba_cloud_fc';
+const TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_OPENSHIFT = 'alibaba_cloud_openshift';
+const TMP_CLOUDPLATFORMVALUES_AWS_EC2 = 'aws_ec2';
+const TMP_CLOUDPLATFORMVALUES_AWS_ECS = 'aws_ecs';
+const TMP_CLOUDPLATFORMVALUES_AWS_EKS = 'aws_eks';
+const TMP_CLOUDPLATFORMVALUES_AWS_LAMBDA = 'aws_lambda';
+const TMP_CLOUDPLATFORMVALUES_AWS_ELASTIC_BEANSTALK = 'aws_elastic_beanstalk';
+const TMP_CLOUDPLATFORMVALUES_AWS_APP_RUNNER = 'aws_app_runner';
+const TMP_CLOUDPLATFORMVALUES_AWS_OPENSHIFT = 'aws_openshift';
+const TMP_CLOUDPLATFORMVALUES_AZURE_VM = 'azure_vm';
+const TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_APPS = 'azure_container_apps';
+const TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_INSTANCES = 'azure_container_instances';
+const TMP_CLOUDPLATFORMVALUES_AZURE_AKS = 'azure_aks';
+const TMP_CLOUDPLATFORMVALUES_AZURE_FUNCTIONS = 'azure_functions';
+const TMP_CLOUDPLATFORMVALUES_AZURE_APP_SERVICE = 'azure_app_service';
+const TMP_CLOUDPLATFORMVALUES_AZURE_OPENSHIFT = 'azure_openshift';
+const TMP_CLOUDPLATFORMVALUES_GCP_BARE_METAL_SOLUTION = 'gcp_bare_metal_solution';
+const TMP_CLOUDPLATFORMVALUES_GCP_COMPUTE_ENGINE = 'gcp_compute_engine';
+const TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_RUN = 'gcp_cloud_run';
+const TMP_CLOUDPLATFORMVALUES_GCP_KUBERNETES_ENGINE = 'gcp_kubernetes_engine';
+const TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_FUNCTIONS = 'gcp_cloud_functions';
+const TMP_CLOUDPLATFORMVALUES_GCP_APP_ENGINE = 'gcp_app_engine';
+const TMP_CLOUDPLATFORMVALUES_GCP_OPENSHIFT = 'gcp_openshift';
+const TMP_CLOUDPLATFORMVALUES_IBM_CLOUD_OPENSHIFT = 'ibm_cloud_openshift';
+const TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_CVM = 'tencent_cloud_cvm';
+const TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_EKS = 'tencent_cloud_eks';
+const TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_SCF = 'tencent_cloud_scf';
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_ALIBABA_CLOUD_ECS = TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_ECS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_ALIBABA_CLOUD_FC = TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_FC;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_ALIBABA_CLOUD_OPENSHIFT = TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_OPENSHIFT;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_EC2 = TMP_CLOUDPLATFORMVALUES_AWS_EC2;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_ECS = TMP_CLOUDPLATFORMVALUES_AWS_ECS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_EKS = TMP_CLOUDPLATFORMVALUES_AWS_EKS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_LAMBDA = TMP_CLOUDPLATFORMVALUES_AWS_LAMBDA;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_ELASTIC_BEANSTALK = TMP_CLOUDPLATFORMVALUES_AWS_ELASTIC_BEANSTALK;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_APP_RUNNER = TMP_CLOUDPLATFORMVALUES_AWS_APP_RUNNER;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AWS_OPENSHIFT = TMP_CLOUDPLATFORMVALUES_AWS_OPENSHIFT;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_VM = TMP_CLOUDPLATFORMVALUES_AZURE_VM;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_CONTAINER_APPS = TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_APPS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_CONTAINER_INSTANCES = TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_INSTANCES;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_AKS = TMP_CLOUDPLATFORMVALUES_AZURE_AKS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_FUNCTIONS = TMP_CLOUDPLATFORMVALUES_AZURE_FUNCTIONS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_APP_SERVICE = TMP_CLOUDPLATFORMVALUES_AZURE_APP_SERVICE;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_AZURE_OPENSHIFT = TMP_CLOUDPLATFORMVALUES_AZURE_OPENSHIFT;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_BARE_METAL_SOLUTION = TMP_CLOUDPLATFORMVALUES_GCP_BARE_METAL_SOLUTION;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_COMPUTE_ENGINE = TMP_CLOUDPLATFORMVALUES_GCP_COMPUTE_ENGINE;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_CLOUD_RUN = TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_RUN;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_KUBERNETES_ENGINE = TMP_CLOUDPLATFORMVALUES_GCP_KUBERNETES_ENGINE;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_CLOUD_FUNCTIONS = TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_FUNCTIONS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_APP_ENGINE = TMP_CLOUDPLATFORMVALUES_GCP_APP_ENGINE;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_GCP_OPENSHIFT = TMP_CLOUDPLATFORMVALUES_GCP_OPENSHIFT;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_IBM_CLOUD_OPENSHIFT = TMP_CLOUDPLATFORMVALUES_IBM_CLOUD_OPENSHIFT;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_TENCENT_CLOUD_CVM = TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_CVM;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_TENCENT_CLOUD_EKS = TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_EKS;
+
+/**
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ */
+export const CLOUDPLATFORMVALUES_TENCENT_CLOUD_SCF = TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_SCF;
+
+/**
+ * Identifies the Values for CloudPlatformValues enum definition
+ *
+ * The cloud platform in use.
+ *
+ * Note: The prefix of the service SHOULD match the one specified in `cloud.provider`.
+ * @deprecated Use the CLOUDPLATFORMVALUES_XXXXX constants rather than the CloudPlatformValues.XXXXX for bundle minification.
+ */
+export type CloudPlatformValues = {
+
+  /** Alibaba Cloud Elastic Compute Service. */
+  ALIBABA_CLOUD_ECS: 'alibaba_cloud_ecs',
+
+  /** Alibaba Cloud Function Compute. */
+  ALIBABA_CLOUD_FC: 'alibaba_cloud_fc',
+
+  /** Red Hat OpenShift on Alibaba Cloud. */
+  ALIBABA_CLOUD_OPENSHIFT: 'alibaba_cloud_openshift',
+
+  /** AWS Elastic Compute Cloud. */
+  AWS_EC2: 'aws_ec2',
+
+  /** AWS Elastic Container Service. */
+  AWS_ECS: 'aws_ecs',
+
+  /** AWS Elastic Kubernetes Service. */
+  AWS_EKS: 'aws_eks',
+
+  /** AWS Lambda. */
+  AWS_LAMBDA: 'aws_lambda',
+
+  /** AWS Elastic Beanstalk. */
+  AWS_ELASTIC_BEANSTALK: 'aws_elastic_beanstalk',
+
+  /** AWS App Runner. */
+  AWS_APP_RUNNER: 'aws_app_runner',
+
+  /** Red Hat OpenShift on AWS (ROSA). */
+  AWS_OPENSHIFT: 'aws_openshift',
+
+  /** Azure Virtual Machines. */
+  AZURE_VM: 'azure_vm',
+
+  /** Azure Container Apps. */
+  AZURE_CONTAINER_APPS: 'azure_container_apps',
+
+  /** Azure Container Instances. */
+  AZURE_CONTAINER_INSTANCES: 'azure_container_instances',
+
+  /** Azure Kubernetes Service. */
+  AZURE_AKS: 'azure_aks',
+
+  /** Azure Functions. */
+  AZURE_FUNCTIONS: 'azure_functions',
+
+  /** Azure App Service. */
+  AZURE_APP_SERVICE: 'azure_app_service',
+
+  /** Azure Red Hat OpenShift. */
+  AZURE_OPENSHIFT: 'azure_openshift',
+
+  /** Google Bare Metal Solution (BMS). */
+  GCP_BARE_METAL_SOLUTION: 'gcp_bare_metal_solution',
+
+  /** Google Cloud Compute Engine (GCE). */
+  GCP_COMPUTE_ENGINE: 'gcp_compute_engine',
+
+  /** Google Cloud Run. */
+  GCP_CLOUD_RUN: 'gcp_cloud_run',
+
+  /** Google Cloud Kubernetes Engine (GKE). */
+  GCP_KUBERNETES_ENGINE: 'gcp_kubernetes_engine',
+
+  /** Google Cloud Functions (GCF). */
+  GCP_CLOUD_FUNCTIONS: 'gcp_cloud_functions',
+
+  /** Google Cloud App Engine (GAE). */
+  GCP_APP_ENGINE: 'gcp_app_engine',
+
+  /** Red Hat OpenShift on Google Cloud. */
+  GCP_OPENSHIFT: 'gcp_openshift',
+
+  /** Red Hat OpenShift on IBM Cloud. */
+  IBM_CLOUD_OPENSHIFT: 'ibm_cloud_openshift',
+
+  /** Tencent Cloud Cloud Virtual Machine (CVM). */
+  TENCENT_CLOUD_CVM: 'tencent_cloud_cvm',
+
+  /** Tencent Cloud Elastic Kubernetes Service (EKS). */
+  TENCENT_CLOUD_EKS: 'tencent_cloud_eks',
+
+  /** Tencent Cloud Serverless Cloud Function (SCF). */
+  TENCENT_CLOUD_SCF: 'tencent_cloud_scf',
+}
+
+/**
+ * The constant map of values for CloudPlatformValues.
+ * @deprecated Use the CLOUDPLATFORMVALUES_XXXXX constants rather than the CloudPlatformValues.XXXXX for bundle minification.
+ */
+export const CloudPlatformValues:CloudPlatformValues = /*#__PURE__*/createConstMap<CloudPlatformValues>([
+  TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_ECS,
+  TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_FC,
+  TMP_CLOUDPLATFORMVALUES_ALIBABA_CLOUD_OPENSHIFT,
+  TMP_CLOUDPLATFORMVALUES_AWS_EC2,
+  TMP_CLOUDPLATFORMVALUES_AWS_ECS,
+  TMP_CLOUDPLATFORMVALUES_AWS_EKS,
+  TMP_CLOUDPLATFORMVALUES_AWS_LAMBDA,
+  TMP_CLOUDPLATFORMVALUES_AWS_ELASTIC_BEANSTALK,
+  TMP_CLOUDPLATFORMVALUES_AWS_APP_RUNNER,
+  TMP_CLOUDPLATFORMVALUES_AWS_OPENSHIFT,
+  TMP_CLOUDPLATFORMVALUES_AZURE_VM,
+  TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_APPS,
+  TMP_CLOUDPLATFORMVALUES_AZURE_CONTAINER_INSTANCES,
+  TMP_CLOUDPLATFORMVALUES_AZURE_AKS,
+  TMP_CLOUDPLATFORMVALUES_AZURE_FUNCTIONS,
+  TMP_CLOUDPLATFORMVALUES_AZURE_APP_SERVICE,
+  TMP_CLOUDPLATFORMVALUES_AZURE_OPENSHIFT,
+  TMP_CLOUDPLATFORMVALUES_GCP_BARE_METAL_SOLUTION,
+  TMP_CLOUDPLATFORMVALUES_GCP_COMPUTE_ENGINE,
+  TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_RUN,
+  TMP_CLOUDPLATFORMVALUES_GCP_KUBERNETES_ENGINE,
+  TMP_CLOUDPLATFORMVALUES_GCP_CLOUD_FUNCTIONS,
+  TMP_CLOUDPLATFORMVALUES_GCP_APP_ENGINE,
+  TMP_CLOUDPLATFORMVALUES_GCP_OPENSHIFT,
+  TMP_CLOUDPLATFORMVALUES_IBM_CLOUD_OPENSHIFT,
+  TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_CVM,
+  TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_EKS,
+  TMP_CLOUDPLATFORMVALUES_TENCENT_CLOUD_SCF,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for CloudProviderValues enum definition
+ *
+ * Name of the cloud provider.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_CLOUDPROVIDERVALUES_ALIBABA_CLOUD = 'alibaba_cloud';
+const TMP_CLOUDPROVIDERVALUES_AWS = 'aws';
+const TMP_CLOUDPROVIDERVALUES_AZURE = 'azure';
+const TMP_CLOUDPROVIDERVALUES_GCP = 'gcp';
+const TMP_CLOUDPROVIDERVALUES_HEROKU = 'heroku';
+const TMP_CLOUDPROVIDERVALUES_IBM_CLOUD = 'ibm_cloud';
+const TMP_CLOUDPROVIDERVALUES_TENCENT_CLOUD = 'tencent_cloud';
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_ALIBABA_CLOUD = TMP_CLOUDPROVIDERVALUES_ALIBABA_CLOUD;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_AWS = TMP_CLOUDPROVIDERVALUES_AWS;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_AZURE = TMP_CLOUDPROVIDERVALUES_AZURE;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_GCP = TMP_CLOUDPROVIDERVALUES_GCP;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_HEROKU = TMP_CLOUDPROVIDERVALUES_HEROKU;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_IBM_CLOUD = TMP_CLOUDPROVIDERVALUES_IBM_CLOUD;
+
+/**
+ * Name of the cloud provider.
+ */
+export const CLOUDPROVIDERVALUES_TENCENT_CLOUD = TMP_CLOUDPROVIDERVALUES_TENCENT_CLOUD;
+
+/**
+ * Identifies the Values for CloudProviderValues enum definition
+ *
+ * Name of the cloud provider.
+ * @deprecated Use the CLOUDPROVIDERVALUES_XXXXX constants rather than the CloudProviderValues.XXXXX for bundle minification.
+ */
+export type CloudProviderValues = {
+
+  /** Alibaba Cloud. */
+  ALIBABA_CLOUD: 'alibaba_cloud',
+
+  /** Amazon Web Services. */
+  AWS: 'aws',
+
+  /** Microsoft Azure. */
+  AZURE: 'azure',
+
+  /** Google Cloud Platform. */
+  GCP: 'gcp',
+
+  /** Heroku Platform as a Service. */
+  HEROKU: 'heroku',
+
+  /** IBM Cloud. */
+  IBM_CLOUD: 'ibm_cloud',
+
+  /** Tencent Cloud. */
+  TENCENT_CLOUD: 'tencent_cloud',
+}
+
+/**
+ * The constant map of values for CloudProviderValues.
+ * @deprecated Use the CLOUDPROVIDERVALUES_XXXXX constants rather than the CloudProviderValues.XXXXX for bundle minification.
+ */
+export const CloudProviderValues:CloudProviderValues = /*#__PURE__*/createConstMap<CloudProviderValues>([
+  TMP_CLOUDPROVIDERVALUES_ALIBABA_CLOUD,
+  TMP_CLOUDPROVIDERVALUES_AWS,
+  TMP_CLOUDPROVIDERVALUES_AZURE,
+  TMP_CLOUDPROVIDERVALUES_GCP,
+  TMP_CLOUDPROVIDERVALUES_HEROKU,
+  TMP_CLOUDPROVIDERVALUES_IBM_CLOUD,
+  TMP_CLOUDPROVIDERVALUES_TENCENT_CLOUD,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for ContainerCpuStateValues enum definition
+ *
+ * The CPU state for this data point.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_CONTAINERCPUSTATEVALUES_USER = 'user';
+const TMP_CONTAINERCPUSTATEVALUES_SYSTEM = 'system';
+const TMP_CONTAINERCPUSTATEVALUES_KERNEL = 'kernel';
+
+/**
+ * The CPU state for this data point.
+ */
+export const CONTAINERCPUSTATEVALUES_USER = TMP_CONTAINERCPUSTATEVALUES_USER;
+
+/**
+ * The CPU state for this data point.
+ */
+export const CONTAINERCPUSTATEVALUES_SYSTEM = TMP_CONTAINERCPUSTATEVALUES_SYSTEM;
+
+/**
+ * The CPU state for this data point.
+ */
+export const CONTAINERCPUSTATEVALUES_KERNEL = TMP_CONTAINERCPUSTATEVALUES_KERNEL;
+
+/**
+ * Identifies the Values for ContainerCpuStateValues enum definition
+ *
+ * The CPU state for this data point.
+ * @deprecated Use the CONTAINERCPUSTATEVALUES_XXXXX constants rather than the ContainerCpuStateValues.XXXXX for bundle minification.
+ */
+export type ContainerCpuStateValues = {
+
+  /** When tasks of the cgroup are in user mode (Linux). When all container processes are in user mode (Windows). */
+  USER: 'user',
+
+  /** When CPU is used by the system (host OS). */
+  SYSTEM: 'system',
+
+  /** When tasks of the cgroup are in kernel mode (Linux). When all container processes are in kernel mode (Windows). */
+  KERNEL: 'kernel',
+}
+
+/**
+ * The constant map of values for ContainerCpuStateValues.
+ * @deprecated Use the CONTAINERCPUSTATEVALUES_XXXXX constants rather than the ContainerCpuStateValues.XXXXX for bundle minification.
+ */
+export const ContainerCpuStateValues:ContainerCpuStateValues = /*#__PURE__*/createConstMap<ContainerCpuStateValues>([
+  TMP_CONTAINERCPUSTATEVALUES_USER,
+  TMP_CONTAINERCPUSTATEVALUES_SYSTEM,
+  TMP_CONTAINERCPUSTATEVALUES_KERNEL,
 ]);
 
 
@@ -6250,6 +8548,84 @@ export const NetTransportValues:NetTransportValues = /*#__PURE__*/createConstMap
 
 
 /* ----------------------------------------------------------------------------------------------------------
+ * Constant values for SystemProcessesStatusValues enum definition
+ *
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING = 'running';
+const TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING = 'sleeping';
+const TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED = 'stopped';
+const TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT = 'defunct';
+
+/**
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.
+ */
+export const SYSTEMPROCESSESSTATUSVALUES_RUNNING = TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING;
+
+/**
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.
+ */
+export const SYSTEMPROCESSESSTATUSVALUES_SLEEPING = TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING;
+
+/**
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.
+ */
+export const SYSTEMPROCESSESSTATUSVALUES_STOPPED = TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED;
+
+/**
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.
+ */
+export const SYSTEMPROCESSESSTATUSVALUES_DEFUNCT = TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT;
+
+/**
+ * Identifies the Values for SystemProcessesStatusValues enum definition
+ *
+ * Deprecated, use `system.process.status` instead.
+ *
+ * @deprecated Replaced by `system.process.status`.  Use the SYSTEMPROCESSESSTATUSVALUES_XXXXX constants rather than the SystemProcessesStatusValues.XXXXX for bundle minification.
+ */
+export type SystemProcessesStatusValues = {
+
+  /** running. */
+  RUNNING: 'running',
+
+  /** sleeping. */
+  SLEEPING: 'sleeping',
+
+  /** stopped. */
+  STOPPED: 'stopped',
+
+  /** defunct. */
+  DEFUNCT: 'defunct',
+}
+
+/**
+ * The constant map of values for SystemProcessesStatusValues.
+ * @deprecated Use the SYSTEMPROCESSESSTATUSVALUES_XXXXX constants rather than the SystemProcessesStatusValues.XXXXX for bundle minification.
+ */
+export const SystemProcessesStatusValues:SystemProcessesStatusValues = /*#__PURE__*/createConstMap<SystemProcessesStatusValues>([
+  TMP_SYSTEMPROCESSESSTATUSVALUES_RUNNING,
+  TMP_SYSTEMPROCESSESSTATUSVALUES_SLEEPING,
+  TMP_SYSTEMPROCESSESSTATUSVALUES_STOPPED,
+  TMP_SYSTEMPROCESSESSTATUSVALUES_DEFUNCT,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
  * Constant values for DiskIoDirectionValues enum definition
  *
  * The disk IO operation direction.
@@ -6378,6 +8754,385 @@ export type ErrorTypeValues = {
 export const ErrorTypeValues:ErrorTypeValues = {
   OTHER: TMP_ERRORTYPEVALUES_OTHER,
 };
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for FaasDocumentOperationValues enum definition
+ *
+ * Describes the type of the operation that was performed on the data.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_FAASDOCUMENTOPERATIONVALUES_INSERT = 'insert';
+const TMP_FAASDOCUMENTOPERATIONVALUES_EDIT = 'edit';
+const TMP_FAASDOCUMENTOPERATIONVALUES_DELETE = 'delete';
+
+/**
+ * Describes the type of the operation that was performed on the data.
+ */
+export const FAASDOCUMENTOPERATIONVALUES_INSERT = TMP_FAASDOCUMENTOPERATIONVALUES_INSERT;
+
+/**
+ * Describes the type of the operation that was performed on the data.
+ */
+export const FAASDOCUMENTOPERATIONVALUES_EDIT = TMP_FAASDOCUMENTOPERATIONVALUES_EDIT;
+
+/**
+ * Describes the type of the operation that was performed on the data.
+ */
+export const FAASDOCUMENTOPERATIONVALUES_DELETE = TMP_FAASDOCUMENTOPERATIONVALUES_DELETE;
+
+/**
+ * Identifies the Values for FaasDocumentOperationValues enum definition
+ *
+ * Describes the type of the operation that was performed on the data.
+ * @deprecated Use the FAASDOCUMENTOPERATIONVALUES_XXXXX constants rather than the FaasDocumentOperationValues.XXXXX for bundle minification.
+ */
+export type FaasDocumentOperationValues = {
+
+  /** When a new object is created. */
+  INSERT: 'insert',
+
+  /** When an object is modified. */
+  EDIT: 'edit',
+
+  /** When an object is deleted. */
+  DELETE: 'delete',
+}
+
+/**
+ * The constant map of values for FaasDocumentOperationValues.
+ * @deprecated Use the FAASDOCUMENTOPERATIONVALUES_XXXXX constants rather than the FaasDocumentOperationValues.XXXXX for bundle minification.
+ */
+export const FaasDocumentOperationValues:FaasDocumentOperationValues = /*#__PURE__*/createConstMap<FaasDocumentOperationValues>([
+  TMP_FAASDOCUMENTOPERATIONVALUES_INSERT,
+  TMP_FAASDOCUMENTOPERATIONVALUES_EDIT,
+  TMP_FAASDOCUMENTOPERATIONVALUES_DELETE,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for FaasInvokedProviderValues enum definition
+ *
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD = 'alibaba_cloud';
+const TMP_FAASINVOKEDPROVIDERVALUES_AWS = 'aws';
+const TMP_FAASINVOKEDPROVIDERVALUES_AZURE = 'azure';
+const TMP_FAASINVOKEDPROVIDERVALUES_GCP = 'gcp';
+const TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD = 'tencent_cloud';
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ */
+export const FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD = TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD;
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ */
+export const FAASINVOKEDPROVIDERVALUES_AWS = TMP_FAASINVOKEDPROVIDERVALUES_AWS;
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ */
+export const FAASINVOKEDPROVIDERVALUES_AZURE = TMP_FAASINVOKEDPROVIDERVALUES_AZURE;
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ */
+export const FAASINVOKEDPROVIDERVALUES_GCP = TMP_FAASINVOKEDPROVIDERVALUES_GCP;
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ */
+export const FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD = TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD;
+
+/**
+ * Identifies the Values for FaasInvokedProviderValues enum definition
+ *
+ * The cloud provider of the invoked function.
+ *
+ * Note: SHOULD be equal to the `cloud.provider` resource attribute of the invoked function.
+ * @deprecated Use the FAASINVOKEDPROVIDERVALUES_XXXXX constants rather than the FaasInvokedProviderValues.XXXXX for bundle minification.
+ */
+export type FaasInvokedProviderValues = {
+
+  /** Alibaba Cloud. */
+  ALIBABA_CLOUD: 'alibaba_cloud',
+
+  /** Amazon Web Services. */
+  AWS: 'aws',
+
+  /** Microsoft Azure. */
+  AZURE: 'azure',
+
+  /** Google Cloud Platform. */
+  GCP: 'gcp',
+
+  /** Tencent Cloud. */
+  TENCENT_CLOUD: 'tencent_cloud',
+}
+
+/**
+ * The constant map of values for FaasInvokedProviderValues.
+ * @deprecated Use the FAASINVOKEDPROVIDERVALUES_XXXXX constants rather than the FaasInvokedProviderValues.XXXXX for bundle minification.
+ */
+export const FaasInvokedProviderValues:FaasInvokedProviderValues = /*#__PURE__*/createConstMap<FaasInvokedProviderValues>([
+  TMP_FAASINVOKEDPROVIDERVALUES_ALIBABA_CLOUD,
+  TMP_FAASINVOKEDPROVIDERVALUES_AWS,
+  TMP_FAASINVOKEDPROVIDERVALUES_AZURE,
+  TMP_FAASINVOKEDPROVIDERVALUES_GCP,
+  TMP_FAASINVOKEDPROVIDERVALUES_TENCENT_CLOUD,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for FaasTriggerValues enum definition
+ *
+ * Type of the trigger which caused this function invocation.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_FAASTRIGGERVALUES_DATASOURCE = 'datasource';
+const TMP_FAASTRIGGERVALUES_HTTP = 'http';
+const TMP_FAASTRIGGERVALUES_PUBSUB = 'pubsub';
+const TMP_FAASTRIGGERVALUES_TIMER = 'timer';
+const TMP_FAASTRIGGERVALUES_OTHER = 'other';
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+export const FAASTRIGGERVALUES_DATASOURCE = TMP_FAASTRIGGERVALUES_DATASOURCE;
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+export const FAASTRIGGERVALUES_HTTP = TMP_FAASTRIGGERVALUES_HTTP;
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+export const FAASTRIGGERVALUES_PUBSUB = TMP_FAASTRIGGERVALUES_PUBSUB;
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+export const FAASTRIGGERVALUES_TIMER = TMP_FAASTRIGGERVALUES_TIMER;
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+export const FAASTRIGGERVALUES_OTHER = TMP_FAASTRIGGERVALUES_OTHER;
+
+/**
+ * Identifies the Values for FaasTriggerValues enum definition
+ *
+ * Type of the trigger which caused this function invocation.
+ * @deprecated Use the FAASTRIGGERVALUES_XXXXX constants rather than the FaasTriggerValues.XXXXX for bundle minification.
+ */
+export type FaasTriggerValues = {
+
+  /** A response to some data source operation such as a database or filesystem read/write. */
+  DATASOURCE: 'datasource',
+
+  /** To provide an answer to an inbound HTTP request. */
+  HTTP: 'http',
+
+  /** A function is set to be executed when messages are sent to a messaging system. */
+  PUBSUB: 'pubsub',
+
+  /** A function is scheduled to be executed regularly. */
+  TIMER: 'timer',
+
+  /** If none of the others apply. */
+  OTHER: 'other',
+}
+
+/**
+ * The constant map of values for FaasTriggerValues.
+ * @deprecated Use the FAASTRIGGERVALUES_XXXXX constants rather than the FaasTriggerValues.XXXXX for bundle minification.
+ */
+export const FaasTriggerValues:FaasTriggerValues = /*#__PURE__*/createConstMap<FaasTriggerValues>([
+  TMP_FAASTRIGGERVALUES_DATASOURCE,
+  TMP_FAASTRIGGERVALUES_HTTP,
+  TMP_FAASTRIGGERVALUES_PUBSUB,
+  TMP_FAASTRIGGERVALUES_TIMER,
+  TMP_FAASTRIGGERVALUES_OTHER,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for HostArchValues enum definition
+ *
+ * The CPU architecture the host system is running on.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_HOSTARCHVALUES_AMD64 = 'amd64';
+const TMP_HOSTARCHVALUES_ARM32 = 'arm32';
+const TMP_HOSTARCHVALUES_ARM64 = 'arm64';
+const TMP_HOSTARCHVALUES_IA64 = 'ia64';
+const TMP_HOSTARCHVALUES_PPC32 = 'ppc32';
+const TMP_HOSTARCHVALUES_PPC64 = 'ppc64';
+const TMP_HOSTARCHVALUES_S390X = 's390x';
+const TMP_HOSTARCHVALUES_X86 = 'x86';
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_AMD64 = TMP_HOSTARCHVALUES_AMD64;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_ARM32 = TMP_HOSTARCHVALUES_ARM32;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_ARM64 = TMP_HOSTARCHVALUES_ARM64;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_IA64 = TMP_HOSTARCHVALUES_IA64;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_PPC32 = TMP_HOSTARCHVALUES_PPC32;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_PPC64 = TMP_HOSTARCHVALUES_PPC64;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_S390X = TMP_HOSTARCHVALUES_S390X;
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+export const HOSTARCHVALUES_X86 = TMP_HOSTARCHVALUES_X86;
+
+/**
+ * Identifies the Values for HostArchValues enum definition
+ *
+ * The CPU architecture the host system is running on.
+ * @deprecated Use the HOSTARCHVALUES_XXXXX constants rather than the HostArchValues.XXXXX for bundle minification.
+ */
+export type HostArchValues = {
+
+  /** AMD64. */
+  AMD64: 'amd64',
+
+  /** ARM32. */
+  ARM32: 'arm32',
+
+  /** ARM64. */
+  ARM64: 'arm64',
+
+  /** Itanium. */
+  IA64: 'ia64',
+
+  /** 32-bit PowerPC. */
+  PPC32: 'ppc32',
+
+  /** 64-bit PowerPC. */
+  PPC64: 'ppc64',
+
+  /** IBM z/Architecture. */
+  S390X: 's390x',
+
+  /** 32-bit x86. */
+  X86: 'x86',
+}
+
+/**
+ * The constant map of values for HostArchValues.
+ * @deprecated Use the HOSTARCHVALUES_XXXXX constants rather than the HostArchValues.XXXXX for bundle minification.
+ */
+export const HostArchValues:HostArchValues = /*#__PURE__*/createConstMap<HostArchValues>([
+  TMP_HOSTARCHVALUES_AMD64,
+  TMP_HOSTARCHVALUES_ARM32,
+  TMP_HOSTARCHVALUES_ARM64,
+  TMP_HOSTARCHVALUES_IA64,
+  TMP_HOSTARCHVALUES_PPC32,
+  TMP_HOSTARCHVALUES_PPC64,
+  TMP_HOSTARCHVALUES_S390X,
+  TMP_HOSTARCHVALUES_X86,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
+ * Constant values for HttpConnectionStateValues enum definition
+ *
+ * State of the HTTP connection in the HTTP connection pool.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_HTTPCONNECTIONSTATEVALUES_ACTIVE = 'active';
+const TMP_HTTPCONNECTIONSTATEVALUES_IDLE = 'idle';
+
+/**
+ * State of the HTTP connection in the HTTP connection pool.
+ */
+export const HTTPCONNECTIONSTATEVALUES_ACTIVE = TMP_HTTPCONNECTIONSTATEVALUES_ACTIVE;
+
+/**
+ * State of the HTTP connection in the HTTP connection pool.
+ */
+export const HTTPCONNECTIONSTATEVALUES_IDLE = TMP_HTTPCONNECTIONSTATEVALUES_IDLE;
+
+/**
+ * Identifies the Values for HttpConnectionStateValues enum definition
+ *
+ * State of the HTTP connection in the HTTP connection pool.
+ * @deprecated Use the HTTPCONNECTIONSTATEVALUES_XXXXX constants rather than the HttpConnectionStateValues.XXXXX for bundle minification.
+ */
+export type HttpConnectionStateValues = {
+
+  /** active state. */
+  ACTIVE: 'active',
+
+  /** idle state. */
+  IDLE: 'idle',
+}
+
+/**
+ * The constant map of values for HttpConnectionStateValues.
+ * @deprecated Use the HTTPCONNECTIONSTATEVALUES_XXXXX constants rather than the HttpConnectionStateValues.XXXXX for bundle minification.
+ */
+export const HttpConnectionStateValues:HttpConnectionStateValues = /*#__PURE__*/createConstMap<HttpConnectionStateValues>([
+  TMP_HTTPCONNECTIONSTATEVALUES_ACTIVE,
+  TMP_HTTPCONNECTIONSTATEVALUES_IDLE,
+]);
 
 
 
@@ -6701,7 +9456,8 @@ export const HttpRequestMethodValues:HttpRequestMethodValues = {
 const TMP_MESSAGINGOPERATIONVALUES_PUBLISH = 'publish';
 const TMP_MESSAGINGOPERATIONVALUES_CREATE = 'create';
 const TMP_MESSAGINGOPERATIONVALUES_RECEIVE = 'receive';
-const TMP_MESSAGINGOPERATIONVALUES_DELIVER = 'deliver';
+const TMP_MESSAGINGOPERATIONVALUES_DELIVER = 'process';
+const TMP_MESSAGINGOPERATIONVALUES_SETTLE = 'settle';
 
 /**
  * A string identifying the kind of messaging operation.
@@ -6732,6 +9488,13 @@ export const MESSAGINGOPERATIONVALUES_RECEIVE = TMP_MESSAGINGOPERATIONVALUES_REC
 export const MESSAGINGOPERATIONVALUES_DELIVER = TMP_MESSAGINGOPERATIONVALUES_DELIVER;
 
 /**
+ * A string identifying the kind of messaging operation.
+ *
+ * Note: If a custom value is used, it MUST be of low cardinality.
+ */
+export const MESSAGINGOPERATIONVALUES_SETTLE = TMP_MESSAGINGOPERATIONVALUES_SETTLE;
+
+/**
  * Identifies the Values for MessagingOperationValues enum definition
  *
  * A string identifying the kind of messaging operation.
@@ -6750,20 +9513,24 @@ export type MessagingOperationValues = {
   /** One or more messages are requested by a consumer. This operation refers to pull-based scenarios, where consumers explicitly call methods of messaging SDKs to receive messages. */
   RECEIVE: 'receive',
 
-  /** One or more messages are passed to a consumer. This operation refers to push-based scenarios, where consumer register callbacks which get called by messaging SDKs. */
-  DELIVER: 'deliver',
+  /** One or more messages are delivered to or processed by a consumer. */
+  DELIVER: 'process',
+
+  /** One or more messages are settled. */
+  SETTLE: 'settle',
 }
 
 /**
  * The constant map of values for MessagingOperationValues.
  * @deprecated Use the MESSAGINGOPERATIONVALUES_XXXXX constants rather than the MessagingOperationValues.XXXXX for bundle minification.
  */
-export const MessagingOperationValues:MessagingOperationValues = /*#__PURE__*/createConstMap<MessagingOperationValues>([
-  TMP_MESSAGINGOPERATIONVALUES_PUBLISH,
-  TMP_MESSAGINGOPERATIONVALUES_CREATE,
-  TMP_MESSAGINGOPERATIONVALUES_RECEIVE,
-  TMP_MESSAGINGOPERATIONVALUES_DELIVER,
-]);
+export const MessagingOperationValues:MessagingOperationValues = {
+  PUBLISH: TMP_MESSAGINGOPERATIONVALUES_PUBLISH,
+  CREATE: TMP_MESSAGINGOPERATIONVALUES_CREATE,
+  RECEIVE: TMP_MESSAGINGOPERATIONVALUES_RECEIVE,
+  DELIVER: TMP_MESSAGINGOPERATIONVALUES_DELIVER,
+  SETTLE: TMP_MESSAGINGOPERATIONVALUES_SETTLE,
+};
 
 
 
@@ -6882,6 +9649,73 @@ export const MessagingRocketmqMessageTypeValues:MessagingRocketmqMessageTypeValu
 
 
 /* ----------------------------------------------------------------------------------------------------------
+ * Constant values for MessagingServicebusDispositionStatusValues enum definition
+ *
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_COMPLETE = 'complete';
+const TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_ABANDON = 'abandon';
+const TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEAD_LETTER = 'dead_letter';
+const TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEFER = 'defer';
+
+/**
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ */
+export const MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_COMPLETE = TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_COMPLETE;
+
+/**
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ */
+export const MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_ABANDON = TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_ABANDON;
+
+/**
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ */
+export const MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEAD_LETTER = TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEAD_LETTER;
+
+/**
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ */
+export const MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEFER = TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEFER;
+
+/**
+ * Identifies the Values for MessagingServicebusDispositionStatusValues enum definition
+ *
+ * Describes the [settlement type](https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock).
+ * @deprecated Use the MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_XXXXX constants rather than the MessagingServicebusDispositionStatusValues.XXXXX for bundle minification.
+ */
+export type MessagingServicebusDispositionStatusValues = {
+
+  /** Message is completed. */
+  COMPLETE: 'complete',
+
+  /** Message is abandoned. */
+  ABANDON: 'abandon',
+
+  /** Message is sent to dead letter queue. */
+  DEAD_LETTER: 'dead_letter',
+
+  /** Message is deferred. */
+  DEFER: 'defer',
+}
+
+/**
+ * The constant map of values for MessagingServicebusDispositionStatusValues.
+ * @deprecated Use the MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_XXXXX constants rather than the MessagingServicebusDispositionStatusValues.XXXXX for bundle minification.
+ */
+export const MessagingServicebusDispositionStatusValues:MessagingServicebusDispositionStatusValues = /*#__PURE__*/createConstMap<MessagingServicebusDispositionStatusValues>([
+  TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_COMPLETE,
+  TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_ABANDON,
+  TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEAD_LETTER,
+  TMP_MESSAGINGSERVICEBUSDISPOSITIONSTATUSVALUES_DEFER,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
  * Constant values for MessagingSystemValues enum definition
  *
  * An identifier for the messaging system being used. See below for a list of well-known identifiers.
@@ -6891,9 +9725,9 @@ export const MessagingRocketmqMessageTypeValues:MessagingRocketmqMessageTypeValu
 // Required to avoid the namespace exports using the unminifiable export names for some package types
 const TMP_MESSAGINGSYSTEMVALUES_ACTIVEMQ = 'activemq';
 const TMP_MESSAGINGSYSTEMVALUES_AWS_SQS = 'aws_sqs';
-const TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTGRID = 'azure_eventgrid';
-const TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTHUBS = 'azure_eventhubs';
-const TMP_MESSAGINGSYSTEMVALUES_AZURE_SERVICEBUS = 'azure_servicebus';
+const TMP_MESSAGINGSYSTEMVALUES_EVENTGRID = 'eventgrid';
+const TMP_MESSAGINGSYSTEMVALUES_EVENTHUBS = 'eventhubs';
+const TMP_MESSAGINGSYSTEMVALUES_SERVICEBUS = 'servicebus';
 const TMP_MESSAGINGSYSTEMVALUES_GCP_PUBSUB = 'gcp_pubsub';
 const TMP_MESSAGINGSYSTEMVALUES_JMS = 'jms';
 const TMP_MESSAGINGSYSTEMVALUES_KAFKA = 'kafka';
@@ -6913,17 +9747,17 @@ export const MESSAGINGSYSTEMVALUES_AWS_SQS = TMP_MESSAGINGSYSTEMVALUES_AWS_SQS;
 /**
  * An identifier for the messaging system being used. See below for a list of well-known identifiers.
  */
-export const MESSAGINGSYSTEMVALUES_AZURE_EVENTGRID = TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTGRID;
+export const MESSAGINGSYSTEMVALUES_EVENTGRID = TMP_MESSAGINGSYSTEMVALUES_EVENTGRID;
 
 /**
  * An identifier for the messaging system being used. See below for a list of well-known identifiers.
  */
-export const MESSAGINGSYSTEMVALUES_AZURE_EVENTHUBS = TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTHUBS;
+export const MESSAGINGSYSTEMVALUES_EVENTHUBS = TMP_MESSAGINGSYSTEMVALUES_EVENTHUBS;
 
 /**
  * An identifier for the messaging system being used. See below for a list of well-known identifiers.
  */
-export const MESSAGINGSYSTEMVALUES_AZURE_SERVICEBUS = TMP_MESSAGINGSYSTEMVALUES_AZURE_SERVICEBUS;
+export const MESSAGINGSYSTEMVALUES_SERVICEBUS = TMP_MESSAGINGSYSTEMVALUES_SERVICEBUS;
 
 /**
  * An identifier for the messaging system being used. See below for a list of well-known identifiers.
@@ -6965,13 +9799,13 @@ export type MessagingSystemValues = {
   AWS_SQS: 'aws_sqs',
 
   /** Azure Event Grid. */
-  AZURE_EVENTGRID: 'azure_eventgrid',
+  EVENTGRID: 'eventgrid',
 
   /** Azure Event Hubs. */
-  AZURE_EVENTHUBS: 'azure_eventhubs',
+  EVENTHUBS: 'eventhubs',
 
   /** Azure Service Bus. */
-  AZURE_SERVICEBUS: 'azure_servicebus',
+  SERVICEBUS: 'servicebus',
 
   /** Google Cloud Pub/Sub. */
   GCP_PUBSUB: 'gcp_pubsub',
@@ -6996,9 +9830,9 @@ export type MessagingSystemValues = {
 export const MessagingSystemValues:MessagingSystemValues = /*#__PURE__*/createConstMap<MessagingSystemValues>([
   TMP_MESSAGINGSYSTEMVALUES_ACTIVEMQ,
   TMP_MESSAGINGSYSTEMVALUES_AWS_SQS,
-  TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTGRID,
-  TMP_MESSAGINGSYSTEMVALUES_AZURE_EVENTHUBS,
-  TMP_MESSAGINGSYSTEMVALUES_AZURE_SERVICEBUS,
+  TMP_MESSAGINGSYSTEMVALUES_EVENTGRID,
+  TMP_MESSAGINGSYSTEMVALUES_EVENTHUBS,
+  TMP_MESSAGINGSYSTEMVALUES_SERVICEBUS,
   TMP_MESSAGINGSYSTEMVALUES_GCP_PUBSUB,
   TMP_MESSAGINGSYSTEMVALUES_JMS,
   TMP_MESSAGINGSYSTEMVALUES_KAFKA,
@@ -7528,6 +10362,143 @@ export const NetworkTypeValues:NetworkTypeValues = /*#__PURE__*/createConstMap<N
 
 
 /* ----------------------------------------------------------------------------------------------------------
+ * Constant values for OsTypeValues enum definition
+ *
+ * The operating system type.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_OSTYPEVALUES_WINDOWS = 'windows';
+const TMP_OSTYPEVALUES_LINUX = 'linux';
+const TMP_OSTYPEVALUES_DARWIN = 'darwin';
+const TMP_OSTYPEVALUES_FREEBSD = 'freebsd';
+const TMP_OSTYPEVALUES_NETBSD = 'netbsd';
+const TMP_OSTYPEVALUES_OPENBSD = 'openbsd';
+const TMP_OSTYPEVALUES_DRAGONFLYBSD = 'dragonflybsd';
+const TMP_OSTYPEVALUES_HPUX = 'hpux';
+const TMP_OSTYPEVALUES_AIX = 'aix';
+const TMP_OSTYPEVALUES_SOLARIS = 'solaris';
+const TMP_OSTYPEVALUES_Z_OS = 'z_os';
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_WINDOWS = TMP_OSTYPEVALUES_WINDOWS;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_LINUX = TMP_OSTYPEVALUES_LINUX;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_DARWIN = TMP_OSTYPEVALUES_DARWIN;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_FREEBSD = TMP_OSTYPEVALUES_FREEBSD;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_NETBSD = TMP_OSTYPEVALUES_NETBSD;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_OPENBSD = TMP_OSTYPEVALUES_OPENBSD;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_DRAGONFLYBSD = TMP_OSTYPEVALUES_DRAGONFLYBSD;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_HPUX = TMP_OSTYPEVALUES_HPUX;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_AIX = TMP_OSTYPEVALUES_AIX;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_SOLARIS = TMP_OSTYPEVALUES_SOLARIS;
+
+/**
+ * The operating system type.
+ */
+export const OSTYPEVALUES_Z_OS = TMP_OSTYPEVALUES_Z_OS;
+
+/**
+ * Identifies the Values for OsTypeValues enum definition
+ *
+ * The operating system type.
+ * @deprecated Use the OSTYPEVALUES_XXXXX constants rather than the OsTypeValues.XXXXX for bundle minification.
+ */
+export type OsTypeValues = {
+
+  /** Microsoft Windows. */
+  WINDOWS: 'windows',
+
+  /** Linux. */
+  LINUX: 'linux',
+
+  /** Apple Darwin. */
+  DARWIN: 'darwin',
+
+  /** FreeBSD. */
+  FREEBSD: 'freebsd',
+
+  /** NetBSD. */
+  NETBSD: 'netbsd',
+
+  /** OpenBSD. */
+  OPENBSD: 'openbsd',
+
+  /** DragonFly BSD. */
+  DRAGONFLYBSD: 'dragonflybsd',
+
+  /** HP-UX (Hewlett Packard Unix). */
+  HPUX: 'hpux',
+
+  /** AIX (Advanced Interactive eXecutive). */
+  AIX: 'aix',
+
+  /** SunOS, Oracle Solaris. */
+  SOLARIS: 'solaris',
+
+  /** IBM z/OS. */
+  Z_OS: 'z_os',
+}
+
+/**
+ * The constant map of values for OsTypeValues.
+ * @deprecated Use the OSTYPEVALUES_XXXXX constants rather than the OsTypeValues.XXXXX for bundle minification.
+ */
+export const OsTypeValues:OsTypeValues = /*#__PURE__*/createConstMap<OsTypeValues>([
+  TMP_OSTYPEVALUES_WINDOWS,
+  TMP_OSTYPEVALUES_LINUX,
+  TMP_OSTYPEVALUES_DARWIN,
+  TMP_OSTYPEVALUES_FREEBSD,
+  TMP_OSTYPEVALUES_NETBSD,
+  TMP_OSTYPEVALUES_OPENBSD,
+  TMP_OSTYPEVALUES_DRAGONFLYBSD,
+  TMP_OSTYPEVALUES_HPUX,
+  TMP_OSTYPEVALUES_AIX,
+  TMP_OSTYPEVALUES_SOLARIS,
+  TMP_OSTYPEVALUES_Z_OS,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
  * Constant values for RpcConnectRpcErrorCodeValues enum definition
  *
  * The [error codes](https://connect.build/docs/protocol/#error-codes) of the Connect request. Error codes are always string values.
@@ -7989,6 +10960,153 @@ export const RpcSystemValues:RpcSystemValues = /*#__PURE__*/createConstMap<RpcSy
 
 
 /* ----------------------------------------------------------------------------------------------------------
+ * Constant values for TelemetrySdkLanguageValues enum definition
+ *
+ * The language of the telemetry SDK.
+ * ---------------------------------------------------------------------------------------------------------- */
+
+// Temporary local constants to assign to the individual exports and the namespaced version
+// Required to avoid the namespace exports using the unminifiable export names for some package types
+const TMP_TELEMETRYSDKLANGUAGEVALUES_CPP = 'cpp';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_DOTNET = 'dotnet';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_ERLANG = 'erlang';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_GO = 'go';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_JAVA = 'java';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_NODEJS = 'nodejs';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_PHP = 'php';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_PYTHON = 'python';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_RUBY = 'ruby';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_RUST = 'rust';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_SWIFT = 'swift';
+const TMP_TELEMETRYSDKLANGUAGEVALUES_WEBJS = 'webjs';
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_CPP = TMP_TELEMETRYSDKLANGUAGEVALUES_CPP;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_DOTNET = TMP_TELEMETRYSDKLANGUAGEVALUES_DOTNET;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_ERLANG = TMP_TELEMETRYSDKLANGUAGEVALUES_ERLANG;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_GO = TMP_TELEMETRYSDKLANGUAGEVALUES_GO;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_JAVA = TMP_TELEMETRYSDKLANGUAGEVALUES_JAVA;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_NODEJS = TMP_TELEMETRYSDKLANGUAGEVALUES_NODEJS;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_PHP = TMP_TELEMETRYSDKLANGUAGEVALUES_PHP;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_PYTHON = TMP_TELEMETRYSDKLANGUAGEVALUES_PYTHON;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_RUBY = TMP_TELEMETRYSDKLANGUAGEVALUES_RUBY;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_RUST = TMP_TELEMETRYSDKLANGUAGEVALUES_RUST;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_SWIFT = TMP_TELEMETRYSDKLANGUAGEVALUES_SWIFT;
+
+/**
+ * The language of the telemetry SDK.
+ */
+export const TELEMETRYSDKLANGUAGEVALUES_WEBJS = TMP_TELEMETRYSDKLANGUAGEVALUES_WEBJS;
+
+/**
+ * Identifies the Values for TelemetrySdkLanguageValues enum definition
+ *
+ * The language of the telemetry SDK.
+ * @deprecated Use the TELEMETRYSDKLANGUAGEVALUES_XXXXX constants rather than the TelemetrySdkLanguageValues.XXXXX for bundle minification.
+ */
+export type TelemetrySdkLanguageValues = {
+
+  /** cpp. */
+  CPP: 'cpp',
+
+  /** dotnet. */
+  DOTNET: 'dotnet',
+
+  /** erlang. */
+  ERLANG: 'erlang',
+
+  /** go. */
+  GO: 'go',
+
+  /** java. */
+  JAVA: 'java',
+
+  /** nodejs. */
+  NODEJS: 'nodejs',
+
+  /** php. */
+  PHP: 'php',
+
+  /** python. */
+  PYTHON: 'python',
+
+  /** ruby. */
+  RUBY: 'ruby',
+
+  /** rust. */
+  RUST: 'rust',
+
+  /** swift. */
+  SWIFT: 'swift',
+
+  /** webjs. */
+  WEBJS: 'webjs',
+}
+
+/**
+ * The constant map of values for TelemetrySdkLanguageValues.
+ * @deprecated Use the TELEMETRYSDKLANGUAGEVALUES_XXXXX constants rather than the TelemetrySdkLanguageValues.XXXXX for bundle minification.
+ */
+export const TelemetrySdkLanguageValues:TelemetrySdkLanguageValues = /*#__PURE__*/createConstMap<TelemetrySdkLanguageValues>([
+  TMP_TELEMETRYSDKLANGUAGEVALUES_CPP,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_DOTNET,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_ERLANG,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_GO,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_JAVA,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_NODEJS,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_PHP,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_PYTHON,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_RUBY,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_RUST,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_SWIFT,
+  TMP_TELEMETRYSDKLANGUAGEVALUES_WEBJS,
+]);
+
+
+
+/* ----------------------------------------------------------------------------------------------------------
  * Constant values for TlsProtocolNameValues enum definition
  *
  * Normalized lowercase protocol name parsed from original string of the negotiated [SSL/TLS protocol version](https://www.openssl.org/docs/man1.1.1/man3/SSL_get_version.html#RETURN-VALUES).
@@ -8133,63 +11251,6 @@ export type OtelStatusCodeValues = {
 export const OtelStatusCodeValues:OtelStatusCodeValues = /*#__PURE__*/createConstMap<OtelStatusCodeValues>([
   TMP_OTELSTATUSCODEVALUES_OK,
   TMP_OTELSTATUSCODEVALUES_ERROR,
-]);
-
-
-
-/* ----------------------------------------------------------------------------------------------------------
- * Constant values for FaasDocumentOperationValues enum definition
- *
- * Describes the type of the operation that was performed on the data.
- * ---------------------------------------------------------------------------------------------------------- */
-
-// Temporary local constants to assign to the individual exports and the namespaced version
-// Required to avoid the namespace exports using the unminifiable export names for some package types
-const TMP_FAASDOCUMENTOPERATIONVALUES_INSERT = 'insert';
-const TMP_FAASDOCUMENTOPERATIONVALUES_EDIT = 'edit';
-const TMP_FAASDOCUMENTOPERATIONVALUES_DELETE = 'delete';
-
-/**
- * Describes the type of the operation that was performed on the data.
- */
-export const FAASDOCUMENTOPERATIONVALUES_INSERT = TMP_FAASDOCUMENTOPERATIONVALUES_INSERT;
-
-/**
- * Describes the type of the operation that was performed on the data.
- */
-export const FAASDOCUMENTOPERATIONVALUES_EDIT = TMP_FAASDOCUMENTOPERATIONVALUES_EDIT;
-
-/**
- * Describes the type of the operation that was performed on the data.
- */
-export const FAASDOCUMENTOPERATIONVALUES_DELETE = TMP_FAASDOCUMENTOPERATIONVALUES_DELETE;
-
-/**
- * Identifies the Values for FaasDocumentOperationValues enum definition
- *
- * Describes the type of the operation that was performed on the data.
- * @deprecated Use the FAASDOCUMENTOPERATIONVALUES_XXXXX constants rather than the FaasDocumentOperationValues.XXXXX for bundle minification.
- */
-export type FaasDocumentOperationValues = {
-
-  /** When a new object is created. */
-  INSERT: 'insert',
-
-  /** When an object is modified. */
-  EDIT: 'edit',
-
-  /** When an object is deleted. */
-  DELETE: 'delete',
-}
-
-/**
- * The constant map of values for FaasDocumentOperationValues.
- * @deprecated Use the FAASDOCUMENTOPERATIONVALUES_XXXXX constants rather than the FaasDocumentOperationValues.XXXXX for bundle minification.
- */
-export const FaasDocumentOperationValues:FaasDocumentOperationValues = /*#__PURE__*/createConstMap<FaasDocumentOperationValues>([
-  TMP_FAASDOCUMENTOPERATIONVALUES_INSERT,
-  TMP_FAASDOCUMENTOPERATIONVALUES_EDIT,
-  TMP_FAASDOCUMENTOPERATIONVALUES_DELETE,
 ]);
